EMC2: Bayesian Hierarchical Analysis of Cognitive Models of Choice

Fit Bayesian (hierarchical) cognitive models using a linear modeling language interface using particle Metropolis Markov chain Monte Carlo sampling with Gibbs steps. The diffusion decision model (DDM), linear ballistic accumulator model (LBA), racing diffusion model (RDM), and the lognormal race model (LNR) are supported. Additionally, users can specify their own likelihood function and/or choose for non-hierarchical estimation, as well as for a diagonal, blocked or full multivariate normal group-level distribution to test individual differences. Prior specification is facilitated through methods that visualize the (implied) prior. A wide range of plotting functions assist in assessing model convergence and posterior inference. Models can be easily evaluated using functions that plot posterior predictions or using relative model comparison metrics such as information criteria or Bayes factors. References: Stevenson et al. (2024) doi:10.31234/osf.io/2e4dq .

See also

Author

Maintainer: Niek Stevenson niek.stevenson@gmail.com (ORCID)

Authors:

Michelle Donzallaz

Andrew Heathcote

Other contributors:

Steven Miletić [contributor]

Raphael Hartmann [contributor]

Karl C. Klauer [contributor]

Steven G. Johnson [contributor]

Jean M. Linhart [contributor]

Brian Gough [contributor]

Gerard Jungman [contributor]

Rudolf Schuerer [contributor]

Przemyslaw Sliwa [contributor]

Jason H. Stover [contributor]