Calculate information criteria (DIC, BPIC), effective number of parameters and constituent posterior deviance (D) summaries (meanD = mean of D, Dmean = D for mean of posterior parameters and minD = minimum of D).

Calculate information criteria (DIC, BPIC), effective number of parameters and constituent posterior deviance (D) summaries (meanD = mean of D, Dmean = D for mean of posterior parameters and minD = minimum of D).

Usage

IC(
  emc,
  stage = "sample",
  filter = 0,
  use_best_fit = TRUE,
  print_summary = TRUE,
  digits = 0,
  subject = NULL,
  group_only = FALSE
)

Arguments

emc

emc object or list of these

stage

A string. Specifies which stage you want to plot.

filter

An integer or vector. If it's an integer, iterations up until the value set by filter will be excluded. If a vector is supplied, only the iterations in the vector will be considered.

use_best_fit

Boolean, default TRUE use best of minD and Dmean in calculation otherwise always use Dmean

print_summary

Boolean (default TRUE) print table of results

digits

Integer, significant digits in printed table

subject

Integer or string selecting a single subject, default NULL returns sums over all subjects

group_only

Boolean. If TRUE will calculate the IC for the group-level only

Value

Table of DIC, BPIC, EffectiveN, meanD, Dmean, and minD