Get Data

Extracts data from an emc object

Usage

# S3 method for class 'emc'
get_data(emc)

get_data(emc)

Arguments

emc

an emc object

Value

A dataframe of the original data

Details

emc adds columns and rows to a dataframe in order to facilitate efficient likelihood calculations. This function will return the data as provided originally.

Examples

get_data(samples_LNR)
#>     subjects        E     S     R   rt
#> 1       as1t accuracy right right 0.44
#> 3       as1t    speed right right 0.50
#> 5       as1t    speed  left  left 0.32
#> 7       as1t accuracy right right 0.48
#> 9       as1t accuracy  left  left 0.34
#> 11      as1t accuracy  left  left 0.34
#> 13      as1t  neutral  left  left 0.36
#> 15      as1t  neutral right right 0.40
#> 17      as1t  neutral  left  left 0.38
#> 19      as1t accuracy right right 0.36
#> 21      as1t accuracy right right 0.38
#> 23      as1t  neutral  left  left 0.36
#> 25      as1t accuracy  left right 0.32
#> 27      as1t accuracy  left  left 0.42
#> 29      as1t  neutral right right 0.52
#> 31      as1t    speed right  left 0.42
#> 33      as1t accuracy right right 0.48
#> 35      as1t    speed right  left 0.44
#> 37      as1t accuracy right right 0.38
#> 39      as1t    speed  left  left 0.52
#> 41      as1t accuracy right right 0.48
#> 43      as1t  neutral  left  left 0.60
#> 45      as1t  neutral  left  left 0.38
#> 47      as1t    speed  left  left 0.50
#> 49      as1t accuracy right right 0.46
#> 51      as1t  neutral right  left 0.32
#> 53      as1t accuracy  left right 0.48
#> 55      as1t  neutral  left  left 0.30
#> 57      as1t  neutral right right 0.52
#> 59      as1t    speed  left  left 0.28
#> 61      as1t    speed  left  left 0.44
#> 63      as1t  neutral right right 0.42
#> 65      as1t    speed  left  left 0.32
#> 67      as1t accuracy  left  left 0.42
#> 69      as1t accuracy  left  left 0.42
#> 71      as1t    speed right right 0.34
#> 73      as1t  neutral right  left 0.32
#> 75      as1t accuracy right right 0.46
#> 77      as1t    speed  left  left 0.52
#> 79      as1t  neutral right right 0.42
#> 81      as1t    speed right  left 0.42
#> 83      as1t    speed  left  left 0.50
#> 85      as1t  neutral right right 0.42
#> 87      as1t  neutral right right 0.40
#> 89      as1t  neutral right right 0.58
#> 91      as1t accuracy  left  left 0.42
#> 93      as1t    speed  left  left 0.44
#> 95      as1t accuracy right right 0.48
#> 97      as1t  neutral right right 0.40
#> 99      as1t    speed right right 0.50
#> 101     as1t    speed  left  left 0.32
#> 103     as1t  neutral right right 0.78
#> 105     as1t accuracy right right 0.46
#> 107     as1t  neutral  left  left 0.46
#> 109     as1t accuracy right right 0.54
#> 111     as1t    speed right right 0.50
#> 113     as1t    speed  left  left 0.32
#> 115     as1t    speed right right 0.56
#> 117     as1t accuracy right  left 0.60
#> 119     as1t accuracy right right 0.46
#> 121     as1t    speed right  left 0.36
#> 123     as1t accuracy right  left 0.60
#> 125     as1t    speed  left  left 0.52
#> 127     as1t  neutral  left  left 0.40
#> 129     as1t  neutral  left  left 0.46
#> 131     as1t accuracy  left  left 0.48
#> 133     as1t    speed  left  left 0.56
#> 135     as1t  neutral right  left 0.48
#> 137     as1t  neutral  left  left 0.46
#> 139     as1t  neutral  left  left 0.40
#> 141     as1t    speed right right 0.50
#> 143     as1t    speed  left  left 0.44
#> 145     as1t  neutral right right 0.42
#> 147     as1t  neutral right  left 0.72
#> 149     as1t    speed  left right 0.50
#> 151     as1t    speed  left  left 0.44
#> 153     as1t    speed right right 0.50
#> 155     as1t  neutral  left  left 0.38
#> 157     as1t accuracy right right 0.46
#> 159     as1t    speed right  left 0.42
#> 161     as1t    speed  left  left 0.44
#> 163     as1t accuracy right right 0.48
#> 165     as1t accuracy  left  left 0.70
#> 167     as1t  neutral right right 0.52
#> 169     as1t    speed right right 0.60
#> 171     as1t    speed right  left 0.34
#> 173     as1t accuracy  left right 0.56
#> 175     as1t    speed  left  left 0.44
#> 177     as1t accuracy right right 0.46
#> 179     as1t accuracy  left  left 0.42
#> 181     as1t accuracy right right 0.46
#> 183     as1t    speed  left right 0.26
#> 185     as1t    speed  left right 0.30
#> 187     as1t accuracy right right 0.48
#> 189     as1t    speed  left  left 0.32
#> 191     as1t  neutral right  left 0.40
#> 193     as1t  neutral  left  left 0.40
#> 195     as1t    speed  left  left 0.50
#> 197     as1t accuracy  left  left 0.48
#> 199     as1t  neutral right right 0.52
#> 201     as1t accuracy right right 0.46
#> 203     as1t accuracy right right 0.48
#> 205     as1t  neutral right right 0.42
#> 207     as1t accuracy right right 0.44
#> 209     as1t    speed right right 0.56
#> 211     as1t accuracy right right 0.46
#> 213     as1t  neutral  left  left 0.46
#> 215     as1t  neutral right right 0.78
#> 217     as1t  neutral right right 0.42
#> 219     as1t accuracy right right 0.38
#> 221     as1t  neutral right right 0.40
#> 223     as1t    speed  left  left 0.56
#> 225     as1t  neutral  left  left 0.36
#> 227     as1t  neutral right right 0.42
#> 229     as1t accuracy  left  left 0.48
#> 231     as1t  neutral right right 0.42
#> 233     as1t    speed right  left 0.42
#> 235     as1t accuracy  left right 0.32
#> 237     as1t  neutral right right 0.40
#> 239     as1t  neutral  left  left 0.54
#> 241     as1t accuracy right right 0.48
#> 243     as1t  neutral  left  left 0.60
#> 245     as1t accuracy  left right 0.34
#> 247     as1t accuracy right right 0.46
#> 249     as1t accuracy right right 0.48
#> 251     as1t  neutral  left  left 0.40
#> 253     as1t accuracy right right 0.36
#> 255     as1t  neutral  left  left 0.38
#> 257     as1t    speed  left  left 0.58
#> 259     as1t accuracy right right 0.38
#> 261     as1t  neutral right  left 0.46
#> 263     as1t  neutral right right 0.42
#> 265     as1t  neutral  left  left 0.46
#> 267     as1t accuracy right right 0.36
#> 269     as1t    speed right right 0.34
#> 271     as1t accuracy right right 0.48
#> 273     as1t  neutral right right 0.40
#> 275     as1t accuracy right right 0.38
#> 277     as1t  neutral  left  left 0.60
#> 279     as1t  neutral  left  left 0.46
#> 281     as1t accuracy right right 0.44
#> 283     as1t accuracy right right 0.46
#> 285     as1t  neutral  left  left 0.46
#> 287     as1t  neutral  left  left 0.40
#> 289     as1t  neutral right right 0.40
#> 291     as1t accuracy  left  left 0.64
#> 293     as1t accuracy right right 0.38
#> 295     as1t  neutral right  left 0.58
#> 297     as1t    speed  left right 0.42
#> 299     as1t  neutral right right 0.40
#> 301     as1t    speed  left right 0.42
#> 303     as1t accuracy  left right 0.78
#> 305     as1t    speed  left right 0.42
#> 307     as1t  neutral right  left 0.48
#> 309     as1t    speed  left  left 0.44
#> 311     as1t    speed  left  left 0.52
#> 313     as1t    speed right  left 0.26
#> 315     as1t    speed right  left 0.36
#> 317     as1t    speed  left right 0.28
#> 319     as1t accuracy  left right 0.46
#> 321     as1t accuracy right right 0.46
#> 323     as1t  neutral  left  left 0.40
#> 325     as1t    speed right right 0.30
#> 327     as1t  neutral  left  left 0.46
#> 329     as1t  neutral  left  left 0.68
#> 331     as1t    speed right right 0.56
#> 333     as1t    speed  left  left 0.44
#> 335     as1t  neutral  left right 0.58
#> 337     as1t    speed right right 0.56
#> 339     as1t  neutral right right 0.42
#> 341     as1t accuracy  left  left 0.48
#> 343     as1t accuracy  left  left 0.66
#> 345     as1t accuracy right right 0.48
#> 347     as1t  neutral right right 0.58
#> 349     as1t  neutral right right 0.40
#> 351     as1t  neutral right right 0.42
#> 353     as1t accuracy  left  left 0.70
#> 355     as1t accuracy  left right 0.46
#> 357     as1t accuracy right right 0.54
#> 359     as1t accuracy  left  left 0.48
#> 361     as1t  neutral right right 0.40
#> 363     as1t    speed  left  left 0.44
#> 365     as1t  neutral  left right 0.58
#> 367     as1t    speed  left  left 0.44
#> 369     as1t  neutral  left  left 0.60
#> 371     as1t    speed right  left 0.44
#> 373     as1t accuracy right right 0.48
#> 375     as1t accuracy  left right 0.36
#> 377     as1t accuracy  left right 0.52
#> 379     as1t  neutral right right 0.40
#> 381     as1t  neutral  left  left 0.68
#> 383     as1t    speed  left  left 0.80
#> 385     as1t accuracy right right 0.44
#> 387     as1t  neutral  left right 0.66
#> 389     as1t  neutral  left right 0.60
#> 391     as1t    speed right right 0.56
#> 393     as1t  neutral right  left 0.58
#> 395     as1t  neutral right right 0.52
#> 397     as1t    speed right right 0.28
#> 399     as1t    speed  left  left 0.26
#>  [ reached 'max' / getOption("max.print") -- omitted 2302 rows ]