Filter/manipulate parameters from emc object

Underlying function used in most plotting and object handling functions in EMC2. Can for example be used to filter/thin a parameter type (i.e, group-level means mu) and convert to an mcmc.list.

Usage

get_pars(
  emc,
  selection = "mu",
  stage = "sample",
  thin = 1,
  filter = 0,
  map = FALSE,
  add_recalculated = FALSE,
  length.out = NULL,
  by_subject = FALSE,
  return_mcmc = TRUE,
  merge_chains = FALSE,
  subject = NULL,
  flatten = FALSE,
  remove_dup = FALSE,
  remove_constants = TRUE,
  use_par = NULL,
  type = NULL,
  true_pars = NULL,
  chain = NULL,
  covariates = NULL
)

Arguments

emc

an emc object.

selection

A Character string. Indicates which parameter type to select (e.g., alpha, mu, sigma2, correlation).

stage

A character string. Indicates from which sampling stage(s) to take the samples from (i.e. preburn, burn, adapt, sample)

thin

An integer. By how much to thin the chains

filter

Integer or numeric vector. If an integer is supplied, iterations up until that integer are removed. If a vector is supplied, the iterations within the range are kept.

map

Boolean. If TRUE parameters will be mapped back to the cells of the experimental design using the design matrices. Otherwise the sampled parameters are returned. Only works for selection = mu or selection = alpha.

add_recalculated

Boolean. If TRUE will also add recalculated parameters, such as b in the LBA (b = B + A; see ?LBA), or z in the DDM z = Z*A (see ?DDM) only works when map = TRUE

length.out

Integer. Alternatively to thinning, you can also select a desired length of the MCMC chains, which will be thinned appropriately.

by_subject

Boolean. If TRUE for selections that include subject parameters (e.g. alpha), plot/stats are organized by subject, otherwise by parameter.

return_mcmc

Boolean. If TRUE returns an mcmc.list object, otherwise a matrix/array with the parameter type.

merge_chains

Boolean. If TRUE returns parameter type merged across chains.

subject

Integer (vector) or character (vector). If an integer will select the 'x'th subject(s), if a character it should match subject names in the data which will be selected.

flatten

Boolean. If FALSE for 3-dimensional samples (e.g., correlations: n-pars x n-pars x iterations). organizes by the dimension containing parameter names, otherwise collapses names across the first and second dimension. Does not apply for selection = "alpha"

remove_dup

Boolean. If TRUE removes duplicate values from the samples. Automatically set to TRUE if flatten = TRUE

remove_constants

Boolean. If TRUE removes constant values from the samples (e.g. 0s in the covariance matrix).

use_par

Character (vector). If specified, only these parameters are returned. Should match the parameter names (i.e. these are collapsed when flatten = TRUE and use_par should also be collapsed names).

type

Character indicating the group-level model selected. Only necessary if sampler isn't specified.

true_pars

Set of true_parameters can be specified to apply flatten or use_par on a set of true parameters

chain

Integer. Which of the chain(s) to return

covariates

Only needed with plot_prior and covariates in the design

Value

An mcmc.list object of the selected parameter types with the specified manipulations

Examples

# E.g. get the group-level mean parameters mapped back to the design
get_pars(samples_LNR, stage = "sample", map = TRUE, selection = "mu")
#> Error in make_missing(data[, names(data) != "winner"], LT, UT, LC, UC,     LCresponse, UCresponse, LCdirection, UCdirection): could not find function "make_missing"

# Or return the flattened correlation, with 10 iterations per chain
get_pars(samples_LNR, stage = "sample", selection = "correlation", flatten = TRUE, length.out = 10)
#> $correlation
#> [[1]]
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 1 
#> End = 10 
#> Thinning interval = 1 
#>          m_lMd.m        s.m       t0.m     s.m_lMd   t0.m_lMd      t0.s
#>  [1,]  0.4161710 -0.6259101 -0.7718499 -0.35553290 -0.6256956 0.6667611
#>  [2,]  0.2501842 -0.9145107 -0.7955403 -0.44346196 -0.6067699 0.8940067
#>  [3,]  0.9380206 -0.7432630 -0.6485934 -0.80325567 -0.7607187 0.8958959
#>  [4,]  0.9012393 -0.8598360 -0.9070840 -0.81564785 -0.7857671 0.8023191
#>  [5,] -0.3318392 -0.6172538 -0.3607328 -0.01709717 -0.4680233 0.7224332
#>  [6,]  0.3632546 -0.5779674 -0.6217344 -0.33415065 -0.6140923 0.5190850
#>  [7,]  0.9089677 -0.4972707 -0.3765058 -0.67796702 -0.5270201 0.8069139
#>  [8,]  0.7921812 -0.6232306 -0.7189267 -0.26017151 -0.4338810 0.5790486
#>  [9,] -0.5027539  0.3108699  0.2131788 -0.57969274 -0.2809811 0.6623240
#> [10,]  0.9050753 -0.9158997 -0.9183176 -0.96503892 -0.9310012 0.9170239
#> 
#> [[2]]
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 1 
#> End = 10 
#> Thinning interval = 1 
#>           m_lMd.m        s.m       t0.m     s.m_lMd   t0.m_lMd       t0.s
#>  [1,]  0.72380800  0.3632221  0.3824542  0.01009507  0.1217113 0.10199593
#>  [2,]  0.63967736 -0.0872879 -0.3767488  0.26465814 -0.5279094 0.08677739
#>  [3,]  0.66764646 -0.1298830  0.2244160 -0.72193329 -0.2378108 0.29716339
#>  [4,]  0.50860601  0.3215838  0.1365122 -0.28725407 -0.2814896 0.72706995
#>  [5,]  0.97202375 -0.8540369 -0.9202158 -0.90616284 -0.9437163 0.86288112
#>  [6,]  0.71747092 -0.7931026 -0.8820200 -0.61464123 -0.8418412 0.75485969
#>  [7,]  0.32358089 -0.8265526 -0.8194899 -0.11420503 -0.1080113 0.98079341
#>  [8,]  0.09362899 -0.4534869 -0.1398391  0.05341269 -0.5076566 0.28797526
#>  [9,] -0.67982519  0.3466001 -0.1773134 -0.45229919  0.2681004 0.49333839
#> [10,]  0.50179387 -0.5953184 -0.4790218 -0.90900744 -0.9234997 0.83370013
#> 
#> [[3]]
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 1 
#> End = 10 
#> Thinning interval = 1 
#>          m_lMd.m         s.m        t0.m     s.m_lMd    t0.m_lMd      t0.s
#>  [1,]  0.8671442 -0.41315331 -0.67095310 -0.52884700 -0.72228925 0.5592671
#>  [2,]  0.7313510 -0.60859329 -0.51751044 -0.59550307 -0.85142037 0.5012968
#>  [3,]  0.7947831 -0.73853832 -0.17845371 -0.78742059 -0.27699367 0.5972722
#>  [4,]  0.6870603 -0.48150737 -0.54228968 -0.59407139 -0.82447241 0.8811392
#>  [5,]  0.6160009 -0.27463906 -0.41870063 -0.50370139 -0.73793340 0.6593811
#>  [6,] -0.1148367 -0.02781127  0.04183441 -0.88253651 -0.66663829 0.6022866
#>  [7,]  0.7717988 -0.81631346 -0.80039250 -0.45855096 -0.69259725 0.5922351
#>  [8,]  0.8803450 -0.05208476 -0.88156786 -0.19539243 -0.96658182 0.1106099
#>  [9,]  0.1273639  0.40332163  0.17323633 -0.55190424 -0.78948670 0.7809628
#> [10,]  0.1032909 -0.92818684 -0.92789494  0.03446882 -0.01983142 0.9086741
#> 
#> attr(,"class")
#> [1] "mcmc.list"
#>