Prior specification or prior sampling for blocked estimation

Works analogous to get_prior_standard. Blocks of the covariance matrix to estimate are only considered in sampling. To get the default prior for a created design: get_prior_diag(design = design, sample = FALSE)

Usage

get_prior_blocked(
  prior = NULL,
  n_pars = NULL,
  sample = TRUE,
  N = 1e+05,
  selection = "mu",
  design = NULL,
  par_groups = NULL
)

Arguments

prior

A named list that can contain the prior mean (theta_mu_mean) and variance (theta_mu_var) on the group-level mean, or the scale (A), or degrees of freedom (v) for the group-level variance-covariance matrix. For NULL entries, a default prior gets created.

n_pars

Often inferred from the design, but if design = NULL, n_pars will be used to determine the size of prior.

sample

Boolean, defaults to TRUE, sample from the prior or simply return the prior specifications?

N

How many samples to draw from the prior, the default is 1e5

selection

Character. If sample = TRUE, what prior to sample from. Options: "mu", "sigma2", "covariance" "Sigma", "alpha", "correlation".

design

The design obtained from design(), required when map = TRUE

par_groups

Integer vector indicating which parts of the covariance matrix should be blocked together

Value

A list with a single entry of type of samples from the prior (if sample = TRUE) or else a prior object

Details

For details see Huang, A., & Wand, M. P. (2013). Simple marginally noninformative prior distributions for covariance matrices. Bayesian Analysis, 8, 439-452. https://doi.org/10.1214/13-BA815.

Note that if sample = FALSE, prior$theta_mu_invar (the inverse of the prior covariance matrix on the group-level mean) is returned, which is only used for computational efficiency

Examples

# First define a design for the model
design_DDMaE <- design(data = forstmann,model=DDM,
                           formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                           constants=c(s=log(1)))
#> Parameter(s) DP, st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $DP
#>  DP
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Now get the default prior
prior <- get_prior_blocked(design = design_DDMaE, sample = FALSE)
# We can change values in the default prior or use `prior`
# Then we can get samples from this prior e.g.
samples <- get_prior_blocked(prior = prior, design = design_DDMaE,
  sample = TRUE, selection = "mu")