Prior specification or prior sampling for single subject estimation

With this type of estimation, one or multiple subjects are estimated independently, without any hierarchical constraint.

get_prior_single(
  prior = NULL,
  n_pars = NULL,
  sample = TRUE,
  N = 1e+05,
  selection = "alpha",
  design = NULL,
  map = FALSE
)

Arguments

prior

A named list containing the prior mean (theta_mu_mean) and variance (theta_mu_var). If NULL, the default prior is used.

n_pars

Often inferred from the design, but if design = NULL, n_pars will be used to determine the size of prior.

sample

Boolean, defaults to TRUE, sample from the prior or simply return the prior specifications?

N

How many samples to draw from the prior, the default is 1e5

selection

Character. If sample = TRUE, what prior to sample from. Options: "alpha".

design

The design obtained from design(), required when map = TRUE

map

Boolean, defaults to TRUE. If sample = TRUE, the implied prior is sampled. This includes back-transformations for naturally bounded parameters such as the non-decision time and an inverse mapping from the design matrix back to the cells of the design. If FALSE, the transformed, unmapped, parameters are used. Note that map does not affect the prior used in the sampling process.

Value

A list with a single entry named "alpha" and samples from the prior (if sample = TRUE) or else a prior object

Details

To specify a (multivariate normal) prior, prior$theta_mu_mean and prior$theta_mu_var an entry is needed for each parameter.

Examples

# First define a design for the model
design_DDMaE <- design(data = forstmann,model=DDM,
                           formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                           constants=c(s=log(1)))
#> Parameter(s) DP, st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $DP
#>  DP
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Now get the default prior
prior <- get_prior_single(design = design_DDMaE, sample = FALSE)
# We can change values in the default prior or use `prior`
# Then we can get samples from this prior e.g.
samples <- get_prior_single(prior = prior, design = design_DDMaE,
  sample = TRUE, selection = "alpha")