make_missing

Truncate or censor data. is.na(rt) not truncated or censored.

Usage

make_missing(
  data,
  LT = 0,
  UT = Inf,
  LC = 0,
  UC = Inf,
  LCresponse = TRUE,
  UCresponse = TRUE,
  LCdirection = TRUE,
  UCdirection = TRUE
)

Arguments

data

Data frame with rt and R columns

LT

lower truncation bound below which data are removed (scalar or subject named vector)

UT

upper truncation bound above which data are removed (scalar or subject named vector)

LC

lower censoring bound (scalar or subject named vector)

UC

upper censoring bound (scalar or subject named vector)

LCresponse

Boolean, default TRUE, if false set LC response to NA

UCresponse

Boolean, default TRUE, if false set UC response to NA

LCdirection

Boolean, default TRUE, set LC rt to 0, else to NA

UCdirection

Boolean, default TRUE, set LC rt to Inf, else to NA

Value

Truncated and censored data frame