Make random effects

Simulates subject-level parameters in the format required by make_data().

Usage

make_random_effects(
  design,
  group_means,
  n_subj = NULL,
  variance_proportion = 0.2,
  covariances = NULL
)

Arguments

design

A design list. The design as specified by design()

group_means

A numeric vector. The group level means for each parameter, in the same order as sampled_p_vector(design)

n_subj

An integer. The number of subjects to generate parameters for. If NULL will be inferred from design

variance_proportion

A double. Optional. If covariances are not specified, the variances will be created by multiplying the means by this number. The covariances will be 0.

covariances

A covariance matrix. Optional. Specify the intended covariance matrix.

Value

A matrix of subject-level parameters.

Examples

# First create a design
design_DDMaE <- design(data = forstmann,model=DDM,
                            formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                            constants=c(s=log(1)))
#> Parameter(s) DP, st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $DP
#>  DP
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then create a group-level means vector:
group_means =c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
               t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
# Now we can create subject-level parameters
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 5)

# We can also define a covariance matrix to simulate from
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 5,
             covariances = diag(.1, length(group_means)))

# The subject level parameters can be used to generate data
make_data(subj_pars, design_DDMaE, n_trials = 10)
#>     subjects        E     S trials     R        rt
#> 1          1    speed  left      1  left 0.2995414
#> 6          1  neutral  left      1  left 0.3583484
#> 11         1 accuracy  left      1  left 0.4415862
#> 16         1    speed right      1  left 0.4019611
#> 21         1  neutral right      1  left 0.3107904
#> 26         1 accuracy right      1 right 0.4193218
#> 31         1    speed  left      2  left 0.3393786
#> 36         1  neutral  left      2  left 0.2956302
#> 41         1 accuracy  left      2  left 0.6497355
#> 46         1    speed right      2 right 0.3237513
#> 51         1  neutral right      2 right 0.4884702
#> 56         1 accuracy right      2 right 1.7338065
#> 61         1    speed  left      3 right 0.5415515
#> 66         1  neutral  left      3  left 0.3358628
#> 71         1 accuracy  left      3  left 0.3873956
#> 76         1    speed right      3 right 0.4463453
#> 81         1  neutral right      3 right 0.9604016
#> 86         1 accuracy right      3 right 0.4157079
#> 91         1    speed  left      4  left 0.3769213
#> 96         1  neutral  left      4  left 0.3276819
#> 101        1 accuracy  left      4  left 0.3762759
#> 106        1    speed right      4  left 0.3113859
#> 111        1  neutral right      4 right 0.6306029
#> 116        1 accuracy right      4 right 0.5715446
#> 121        1    speed  left      5  left 0.5678780
#> 126        1  neutral  left      5  left 1.9781526
#> 131        1 accuracy  left      5  left 0.3153465
#> 136        1    speed right      5 right 0.4415034
#> 141        1  neutral right      5 right 0.6045911
#> 146        1 accuracy right      5  left 0.4614368
#> 151        1    speed  left      6  left 0.3009741
#> 156        1  neutral  left      6  left 0.3889159
#> 161        1 accuracy  left      6  left 0.4733615
#> 166        1    speed right      6 right 0.3416467
#> 171        1  neutral right      6 right 0.9924565
#> 176        1 accuracy right      6 right 0.9002699
#> 181        1    speed  left      7  left 0.3652762
#> 186        1  neutral  left      7  left 0.3300306
#> 191        1 accuracy  left      7  left 0.5758831
#> 196        1    speed right      7 right 0.3380202
#> 201        1  neutral right      7  left 0.3672024
#> 206        1 accuracy right      7 right 0.9204381
#> 211        1    speed  left      8  left 0.4399299
#> 216        1  neutral  left      8  left 0.3677265
#> 221        1 accuracy  left      8  left 0.3783819
#> 226        1    speed right      8 right 0.3622825
#> 231        1  neutral right      8 right 1.0131804
#> 236        1 accuracy right      8 right 1.0222337
#> 241        1    speed  left      9  left 0.2947718
#> 246        1  neutral  left      9  left 0.3546473
#> 251        1 accuracy  left      9  left 0.4542118
#> 256        1    speed right      9 right 0.6848879
#> 261        1  neutral right      9 right 0.5259962
#> 266        1 accuracy right      9  left 0.3109069
#> 271        1    speed  left     10  left 0.4941061
#> 276        1  neutral  left     10  left 0.3728231
#> 281        1 accuracy  left     10  left 0.4657056
#> 286        1    speed right     10  left 0.3137829
#> 291        1  neutral right     10 right 0.8876180
#> 296        1 accuracy right     10 right 0.4545681
#> 2          2    speed  left      1  left 0.3065769
#> 7          2  neutral  left      1  left 0.5576033
#> 12         2 accuracy  left      1  left 0.6827103
#> 17         2    speed right      1 right 0.3468671
#> 22         2  neutral right      1 right 0.8907775
#> 27         2 accuracy right      1 right 0.5494491
#> 32         2    speed  left      2  left 0.1660164
#> 37         2  neutral  left      2  left 1.6469618
#> 42         2 accuracy  left      2  left 0.7901413
#> 47         2    speed right      2 right 0.5455096
#> 52         2  neutral right      2 right 0.5669529
#> 57         2 accuracy right      2 right 0.7301292
#> 62         2    speed  left      3  left 0.1919806
#> 67         2  neutral  left      3  left 0.3780501
#> 72         2 accuracy  left      3  left 1.1251296
#> 77         2    speed right      3 right 0.7487350
#> 82         2  neutral right      3 right 0.7170905
#> 87         2 accuracy right      3 right 0.4629095
#> 92         2    speed  left      4  left 0.4317188
#> 97         2  neutral  left      4  left 0.6419852
#> 102        2 accuracy  left      4  left 0.5806264
#> 107        2    speed right      4 right 0.7576242
#> 112        2  neutral right      4 right 1.5597910
#> 117        2 accuracy right      4 right 0.7842194
#> 122        2    speed  left      5  left 0.5800167
#> 127        2  neutral  left      5  left 0.4969508
#> 132        2 accuracy  left      5  left 2.2796608
#> 137        2    speed right      5 right 0.3993220
#> 142        2  neutral right      5 right 0.8156119
#> 147        2 accuracy right      5 right 0.6370908
#> 152        2    speed  left      6  left 0.5476387
#> 157        2  neutral  left      6  left 0.6201715
#> 162        2 accuracy  left      6  left 0.9236033
#> 167        2    speed right      6 right 0.3221148
#> 172        2  neutral right      6 right 0.8300677
#> 177        2 accuracy right      6 right 1.2975116
#> 182        2    speed  left      7 right 0.3916724
#> 187        2  neutral  left      7  left 0.8953424
#> 192        2 accuracy  left      7  left 0.9535978
#> 197        2    speed right      7  left 0.2757758
#> 202        2  neutral right      7 right 0.3635952
#> 207        2 accuracy right      7 right 1.3285584
#> 212        2    speed  left      8  left 0.2295545
#> 217        2  neutral  left      8  left 0.6389105
#> 222        2 accuracy  left      8  left 0.5044983
#> 227        2    speed right      8 right 0.5975954
#> 232        2  neutral right      8 right 0.9169343
#> 237        2 accuracy right      8 right 2.0039226
#> 242        2    speed  left      9 right 0.1710650
#> 247        2  neutral  left      9  left 0.4396664
#> 252        2 accuracy  left      9  left 0.7632470
#> 257        2    speed right      9 right 0.3230873
#> 262        2  neutral right      9 right 0.4531456
#> 267        2 accuracy right      9 right 1.2446239
#> 272        2    speed  left     10  left 0.2972930
#> 277        2  neutral  left     10  left 0.5362876
#> 282        2 accuracy  left     10  left 0.4768326
#> 287        2    speed right     10 right 0.7188874
#> 292        2  neutral right     10 right 0.3910318
#> 297        2 accuracy right     10 right 0.5914659
#> 3          3    speed  left      1  left 0.3118934
#> 8          3  neutral  left      1  left 1.1413158
#> 13         3 accuracy  left      1 right 0.4701726
#> 18         3    speed right      1 right 0.3639354
#> 23         3  neutral right      1 right 0.5406580
#> 28         3 accuracy right      1  left 0.1953277
#> 33         3    speed  left      2  left 0.4502992
#> 38         3  neutral  left      2  left 1.0406997
#> 43         3 accuracy  left      2  left 0.2122231
#> 48         3    speed right      2 right 0.3944810
#> 53         3  neutral right      2 right 0.7858485
#> 58         3 accuracy right      2 right 0.4501216
#> 63         3    speed  left      3  left 0.4380281
#> 68         3  neutral  left      3  left 0.4665894
#> 73         3 accuracy  left      3  left 0.4619752
#> 78         3    speed right      3 right 0.3353085
#> 83         3  neutral right      3 right 0.4052883
#> 88         3 accuracy right      3  left 0.1930964
#> 93         3    speed  left      4  left 0.2544328
#> 98         3  neutral  left      4  left 0.3908202
#> 103        3 accuracy  left      4  left 0.5720017
#> 108        3    speed right      4 right 0.3199325
#> 113        3  neutral right      4 right 0.8328068
#> 118        3 accuracy right      4 right 0.4425924
#> 123        3    speed  left      5  left 0.2601813
#> 128        3  neutral  left      5  left 0.4240205
#> 133        3 accuracy  left      5  left 0.3504820
#> 138        3    speed right      5 right 0.3989284
#> 143        3  neutral right      5 right 0.9235932
#> 148        3 accuracy right      5 right 0.8814884
#> 153        3    speed  left      6  left 0.2295825
#> 158        3  neutral  left      6  left 1.1421131
#> 163        3 accuracy  left      6  left 0.3461711
#> 168        3    speed right      6 right 0.3451842
#> 173        3  neutral right      6 right 1.1824155
#> 178        3 accuracy right      6 right 0.5010314
#> 183        3    speed  left      7  left 0.8759249
#> 188        3  neutral  left      7  left 0.6676297
#> 193        3 accuracy  left      7  left 0.3966067
#> 198        3    speed right      7 right 0.2744290
#> 203        3  neutral right      7 right 0.6154196
#> 208        3 accuracy right      7 right 0.2357148
#> 213        3    speed  left      8  left 0.4767609
#> 218        3  neutral  left      8  left 0.8212048
#> 223        3 accuracy  left      8  left 0.3167934
#> 228        3    speed right      8 right 0.4667461
#> 233        3  neutral right      8 right 0.4319948
#> 238        3 accuracy right      8 right 0.7610118
#> 243        3    speed  left      9  left 0.3934128
#> 248        3  neutral  left      9  left 0.6856269
#> 253        3 accuracy  left      9  left 0.2310148
#> 258        3    speed right      9 right 0.3851705
#> 263        3  neutral right      9 right 0.5059137
#> 268        3 accuracy right      9 right 0.3828133
#> 273        3    speed  left     10  left 0.3626898
#> 278        3  neutral  left     10  left 0.4132706
#> 283        3 accuracy  left     10  left 0.4567003
#> 288        3    speed right     10 right 0.3093432
#> 293        3  neutral right     10 right 0.3447157
#> 298        3 accuracy right     10 right 0.3117571
#> 4          4    speed  left      1 right 0.2371078
#> 9          4  neutral  left      1  left 0.5161052
#> 14         4 accuracy  left      1  left 2.0222802
#> 19         4    speed right      1 right 0.4101068
#> 24         4  neutral right      1 right 0.5631388
#> 29         4 accuracy right      1 right 0.6954881
#> 34         4    speed  left      2 right 0.3139694
#> 39         4  neutral  left      2  left 0.9047485
#> 44         4 accuracy  left      2  left 0.6088628
#> 49         4    speed right      2 right 0.4797702
#> 54         4  neutral right      2 right 0.4459461
#> 59         4 accuracy right      2 right 0.3172765
#> 64         4    speed  left      3  left 0.3292415
#> 69         4  neutral  left      3  left 1.5523302
#> 74         4 accuracy  left      3  left 0.5403207
#> 79         4    speed right      3 right 0.3816624
#> 84         4  neutral right      3 right 0.4170256
#> 89         4 accuracy right      3 right 0.5024358
#> 94         4    speed  left      4  left 0.3199466
#> 99         4  neutral  left      4  left 0.4637900
#> 104        4 accuracy  left      4  left 0.4556296
#> 109        4    speed right      4 right 0.2599012
#> 114        4  neutral right      4 right 0.4735717
#> 119        4 accuracy right      4 right 0.4153411
#> 124        4    speed  left      5  left 0.2866277
#> 129        4  neutral  left      5  left 0.7779011
#> 134        4 accuracy  left      5  left 0.2813750
#> 139        4    speed right      5 right 0.5667261
#> 144        4  neutral right      5 right 0.2918544
#> 149        4 accuracy right      5 right 0.4245728
#> 154        4    speed  left      6  left 0.3092142
#> 159        4  neutral  left      6  left 0.5836727
#> 164        4 accuracy  left      6  left 0.3004753
#> 169        4    speed right      6 right 0.3449028
#> 174        4  neutral right      6 right 0.3355733
#> 179        4 accuracy right      6 right 0.3387262
#> 184        4    speed  left      7 right 0.4334856
#> 189        4  neutral  left      7  left 0.4566293
#> 194        4 accuracy  left      7  left 0.6120978
#> 199        4    speed right      7 right 0.2606721
#> 204        4  neutral right      7 right 0.3193047
#> 209        4 accuracy right      7 right 0.2964968
#> 214        4    speed  left      8  left 0.2692844
#> 219        4  neutral  left      8  left 0.2537158
#> 224        4 accuracy  left      8  left 0.3034771
#> 229        4    speed right      8 right 0.2568036
#> 234        4  neutral right      8 right 0.3121873
#> 239        4 accuracy right      8 right 0.2654319
#> 244        4    speed  left      9  left 0.2783192
#> 249        4  neutral  left      9  left 0.9479375
#> 254        4 accuracy  left      9  left 0.9015227
#> 259        4    speed right      9 right 0.3451739
#> 264        4  neutral right      9 right 0.2366224
#> 269        4 accuracy right      9 right 0.3653259
#> 274        4    speed  left     10  left 0.2525908
#> 279        4  neutral  left     10  left 0.8691395
#> 284        4 accuracy  left     10  left 1.4099690
#> 289        4    speed right     10 right 0.6803406
#> 294        4  neutral right     10 right 0.8730807
#> 299        4 accuracy right     10 right 0.2788521
#> 5          5    speed  left      1  left 0.6853570
#> 10         5  neutral  left      1  left 0.8158273
#> 15         5 accuracy  left      1  left 0.8458805
#> 20         5    speed right      1 right 0.4433538
#> 25         5  neutral right      1 right 0.7327169
#> 30         5 accuracy right      1 right 0.5802943
#> 35         5    speed  left      2  left 0.6094162
#> 40         5  neutral  left      2  left 0.7806771
#> 45         5 accuracy  left      2 right 1.9424628
#> 50         5    speed right      2 right 0.2953443
#> 55         5  neutral right      2 right 0.6992032
#> 60         5 accuracy right      2 right 0.4576715
#> 65         5    speed  left      3  left 1.4974853
#> 70         5  neutral  left      3  left 1.2196954
#> 75         5 accuracy  left      3  left 2.0396196
#> 80         5    speed right      3 right 0.3819205
#> 85         5  neutral right      3 right 0.7176688
#> 90         5 accuracy right      3 right 0.5219850
#> 95         5    speed  left      4  left 0.4511835
#> 100        5  neutral  left      4 right 0.6746380
#> 105        5 accuracy  left      4  left 1.8300425
#> 110        5    speed right      4 right 0.2721119
#> 115        5  neutral right      4 right 0.6533670
#> 120        5 accuracy right      4 right 0.4158001
#> 125        5    speed  left      5  left 0.8263121
#> 130        5  neutral  left      5  left 0.8925286
#> 135        5 accuracy  left      5  left 1.5660587
#> 140        5    speed right      5 right 0.2960153
#> 145        5  neutral right      5 right 0.4048868
#> 150        5 accuracy right      5 right 0.8388554
#> 155        5    speed  left      6  left 0.6176719
#> 160        5  neutral  left      6  left 0.8187872
#> 165        5 accuracy  left      6  left 1.4221717
#> 170        5    speed right      6 right 0.3966479
#> 175        5  neutral right      6 right 0.4863436
#> 180        5 accuracy right      6 right 0.8366788
#> 185        5    speed  left      7  left 0.9842520
#> 190        5  neutral  left      7  left 0.5293706
#> 195        5 accuracy  left      7  left 1.2587381
#> 200        5    speed right      7 right 0.4005901
#> 205        5  neutral right      7 right 0.3534678
#> 210        5 accuracy right      7 right 1.4318552
#> 215        5    speed  left      8  left 0.7645413
#> 220        5  neutral  left      8  left 0.7719454
#> 225        5 accuracy  left      8  left 0.8921597
#> 230        5    speed right      8 right 0.3485048
#> 235        5  neutral right      8 right 0.4753860
#> 240        5 accuracy right      8 right 1.2764059
#> 245        5    speed  left      9  left 0.5422707
#> 250        5  neutral  left      9  left 1.4753627
#> 255        5 accuracy  left      9  left 1.0093063
#> 260        5    speed right      9 right 0.3685794
#> 265        5  neutral right      9 right 0.6130678
#> 270        5 accuracy right      9 right 0.7595593
#> 275        5    speed  left     10  left 0.4852571
#> 280        5  neutral  left     10  left 1.4156648
#> 285        5 accuracy  left     10  left 1.9476000
#> 290        5    speed right     10 right 0.4369146
#> 295        5  neutral right     10 right 0.5551187
#> 300        5 accuracy right     10 right 0.5993224