Return Data Frame of Parameters

Return Data Frame of Parameters

Usage

# S3 method for class 'emc.prior'
parameters(x, selection = "mu", N = 1000, covariates = NULL, ...)

# S3 method for class 'emc'
parameters(x, selection = "mu", N = NULL, resample = FALSE, ...)

parameters(x, ...)

Arguments

x

An emc or emc.prior object

selection

String designating parameter type (e.g. mu, sigma2, correlation, alpha)

N

Integer. How many samples to take from the posterior/prior. If NULL will return the full posterior

covariates

For priors, possible covariates in the design

...

Optional arguments that can be passed to get_pars

resample

Boolean. If TRUE will sample N samples from the posterior with replacement

Value

A data frame with one row for each sample (with a subjects column if selection = "alpha" and using draws from the posterior)

Examples

# For prior inference:
# First set up a prior
design_DDMaE <- design(data = forstmann,model=DDM,
                       formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                       constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then set up a prior using make_prior
p_vector=c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
           t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
psd <- c(v_Sleft=1,v_Sright=1,a=.3,a_Eneutral=.3,a_Eaccuracy=.3,
         t0=.4,Z=1,sv=.4,SZ=1)
# Here we left the variance prior at default
prior_DDMaE <- prior(design_DDMaE,mu_mean=p_vector,mu_sd=psd)
# Get our prior samples
parameters(prior_DDMaE, N = 100)
#>          v_Sleft   v_Sright            a  a_Eneutral a_Eaccuracy         t0
#> 1   -2.814549941  2.3918204 -0.098133219  0.14356633   1.2122731 -1.8539313
#> 2   -1.565209647  2.5329291 -0.019495517  0.41070198   0.3416735 -1.4646093
#> 3   -0.881121277  1.6101171 -0.268860389  0.54759147   1.0362209 -1.8003244
#> 4   -1.988297686  1.7767420 -0.678565717  0.25813378   1.0181041 -2.2137528
#> 5   -0.056668487  2.3068490  0.097209701  0.61530338   0.9048190 -1.9112027
#> 6   -1.311394658  0.9178694  0.076787524  0.17549190   0.8335220 -2.5824321
#> 7   -1.277106553  2.3448182 -0.315687372  0.01693237   1.0968605 -2.6748143
#> 8   -3.606829955 -0.1007887  0.166464462  0.51771003   0.5656771 -2.1973013
#> 9   -2.618611730  2.7188282  0.650860216  0.49750336   0.7834693 -1.3133034
#> 10  -1.435552138  1.3204836  0.154784278  0.08673077   0.7368435 -1.6234234
#> 11  -1.336465122  3.0124143 -0.364773168  0.55359355   0.9556535 -1.0459253
#> 12  -1.525595739  3.4074964 -0.163923740  0.70728723   0.9397537 -1.6224007
#> 13  -1.783832607  3.3781271 -0.409280240  0.60205198   0.1687738 -2.2139183
#> 14  -3.301381380  2.8728885  0.217538789  0.69901326   0.9445871 -1.5525011
#> 15  -0.807823529  1.5511886 -0.164173451  0.99115445   0.6780801 -1.6926156
#> 16  -2.662445030  2.3982250 -0.009585893  0.81048195   0.8536689 -2.0826690
#> 17  -1.211811345  3.5405140  0.105724595  0.75569719   0.7400946 -1.6060801
#> 18  -2.689481310  1.9003080  0.016884428  0.13006392   0.7392806 -1.5973743
#> 19  -2.082780580  1.6139714 -0.269299951  0.70488605   0.8628881 -1.1781310
#> 20  -2.930476226  1.5786804  0.283468608 -0.31288817   0.1987859 -1.9624656
#> 21  -2.545291073  2.0326977 -0.065050040  0.86719030   1.0566132 -1.5358110
#> 22  -0.960641964  4.2685252  0.493596456  0.52067278   0.7814901 -1.6511034
#> 23  -4.197026153  1.8861021  0.618281991  0.05167972   0.2477658 -0.9569139
#> 24  -2.949688702  3.0599908 -0.278330484  0.03238722   0.8557874 -1.9987225
#> 25  -3.575990195  2.1112452  0.447202622  0.83132000   1.0563595 -2.0292590
#> 26  -0.183685606  2.0773491 -0.877359296  0.46000016   0.5856510 -1.1869025
#> 27  -2.999558597  0.4074061 -0.238937193  0.04109826   0.5088250 -1.2679940
#> 28  -2.850682317  2.0781772  0.009967773  0.75105806   1.3382610 -1.7927713
#> 29  -2.061014599  3.5083591 -0.123212521  0.55021990   1.0212684 -1.7292297
#> 30  -3.023414817  1.6367809 -0.254759805  0.81323842   0.4878363 -1.5607135
#> 31  -2.911805615  2.9554261  0.174742701  0.13686858   0.2217457 -1.8064497
#> 32  -2.484597170  1.1421021 -0.258391905  0.93593212   0.7477033 -1.1984451
#> 33  -0.832957374  2.3029749  0.116821904  0.55933108   0.6279827 -1.8761185
#> 34  -0.484690496  0.5953949  0.240075733  1.02649385   0.6422598 -1.8366849
#> 35  -1.873971143  0.5524264  0.515167885  0.49430415   0.2955884 -1.8792660
#> 36  -0.983617400  0.2678443  0.857460071  0.21327658   0.5433042 -1.5286148
#> 37  -3.086723146  1.0840262  0.045287997  0.31381122   1.4025551 -1.0998604
#> 38  -0.408010267  2.8074140  0.202296834  0.73439476   0.6071344 -1.5515493
#> 39  -2.221901185  0.8860931 -0.653073450  0.22182270   1.0212405 -1.2733317
#> 40  -2.262486754  2.0232335 -0.353220696  0.60498314   0.3281667 -1.1734683
#> 41  -2.166984169  2.1585089  0.192719027  0.51607528   1.2476510 -1.0706519
#> 42  -2.568063222  1.9755176 -0.231390907  0.44646050   0.5758382 -1.5170828
#> 43  -1.620815131  1.8782506 -0.135455581  0.63818751   0.6511956 -1.6127999
#> 44   0.618612232  1.7093632 -0.162420843  0.60291014   1.2145348 -1.8362682
#> 45  -1.257363905  2.7566561  0.384066784  0.77848632   0.3014970 -1.3080217
#> 46  -0.856609316  1.8830236  0.052559044  0.73613583   0.2438390 -1.4366882
#> 47  -1.498190415  1.9138314 -0.572056145  0.31185070   1.0383656 -2.6953264
#> 48  -3.238924382  3.0022945  0.481812188  0.21054846   0.6118798 -1.7345087
#> 49  -2.733250957  1.2545377  0.443911098  0.13503121   0.4569265 -1.2186488
#> 50  -0.862607346  0.8351157 -0.144457834  0.33657875   0.2005326 -1.2980866
#> 51  -0.732463507  2.8031870  0.207930332  0.51376834   0.8061434 -1.9519439
#> 52  -3.537271692  3.1299744  0.462434339  0.42953183   0.9218605 -1.8160000
#> 53  -2.784475552  2.0564545 -0.346331562  0.31196846   1.1484909 -2.0218671
#> 54  -2.119644755  0.5069330 -0.347334887  0.11664658   0.1605079 -1.4176634
#> 55  -1.500199177  1.1830829 -0.138878796  0.60017372   0.6615458 -2.0084343
#> 56  -0.685937057  1.6003121  0.437346938 -0.03766361   0.3584559 -1.4286940
#> 57  -3.119144126  1.6123594 -0.305507244  0.55911559   0.8423353 -1.4855337
#> 58  -1.432335296  0.2650108  0.004549814  1.00427367   0.7970699 -2.2285589
#> 59  -1.171167104  0.6483499 -0.113171292  0.47417846   0.6788746 -1.1693329
#> 60  -1.803718731  1.7076257  0.066264661  0.52139987   0.5316659 -2.3737049
#> 61  -2.904750773  1.0196192 -0.262614160  0.41846867   0.9923393 -1.3828367
#> 62  -0.009203507  2.8326688 -0.181493675  0.71383882   0.8020636 -1.8960517
#> 63  -1.941468223  3.0201948  0.068686777  0.31861926   0.6547390 -1.8958774
#> 64  -0.914719195  2.5616430 -0.318975303  0.34507601   0.6727680 -1.5884749
#> 65  -3.081734865  1.9356506  0.330961695  0.56921418   0.4761404 -1.6915063
#> 66  -3.138548972  2.7279727  0.117723856  0.47343946   0.5546000 -1.1022443
#> 67  -2.163112657  3.8072352 -0.229790958  0.80743733   1.0380860 -1.2721748
#> 68  -1.111879338  2.3688864  0.090850924  0.28344582   0.9811998 -0.9698999
#> 69  -3.624446189  2.2579605  0.220974247  0.68927067   0.5939515 -1.8783626
#> 70  -2.711713532  2.5954373  0.057590898  0.01520308   0.5069551 -1.6372326
#> 71  -3.059803595  0.8313578  0.212252146  0.83754086   0.4628690 -1.9382226
#> 72  -1.888797020  4.6968949  0.008146264  0.65323828   0.2761398 -2.0152743
#> 73  -0.647644737  0.3720250 -0.115594301  0.38381708   0.9160088 -1.7229397
#> 74  -0.424040002  0.5769106 -0.123156285  0.38914807   1.4194797 -0.9247248
#> 75  -0.794553423  2.4556712  0.094190944  0.68090139   0.6758108 -1.8574584
#> 76  -3.055334814  2.8887469 -0.176694120  0.80312839   0.4496008 -1.4462533
#> 77  -1.702551164  0.8862599 -0.348455377  0.30689656   1.1017558 -2.2075376
#> 78  -0.746228738  0.9042695 -0.212261733  0.50353370   1.0398946 -1.4786942
#> 79  -1.533525949  4.3129649 -0.068001820  0.42424104   0.8293143 -0.7022800
#> 80  -2.901353848  2.9922936 -0.718070872  0.26157948   0.1247128 -1.5955301
#> 81  -0.807029319  1.1113336  0.180244935  0.15514198   0.6907993 -1.7018482
#> 82  -3.069513029  2.4806586  0.310790234  0.26883970   0.1863030 -1.5526871
#> 83  -2.295906950  2.7475166  0.147628931  0.46288249   0.3972400 -0.8713768
#> 84  -2.176783739  2.7111540 -0.205641488  0.14721820   0.9932105 -1.1891003
#> 85  -2.651833017  0.3133915  0.333920938  0.06579742   0.1451616 -1.7120610
#> 86  -1.791257866  2.6612534 -0.053796832  0.99450115   0.5814859 -2.3489167
#> 87  -2.765109338  0.8112293 -0.235400329  0.38882631   0.9111311 -1.3565399
#> 88  -2.939721676  1.9842166 -0.051795629  0.63055966   0.6796783 -1.5782977
#> 89  -1.350365269  0.6644340 -0.056408428  0.44911559   0.4772597 -1.6040522
#> 90  -1.706097636  0.3418382 -0.254787102  0.05750970   0.6212237 -1.0840291
#> 91  -1.111057195  2.7115347 -0.075670239  0.93517783   1.1094576 -1.4532241
#> 92  -2.083778322  2.0590178 -0.570314847  0.09889521   0.6765126 -1.3261488
#> 93  -1.414992558 -0.2126605 -0.058890401  0.58893049   0.7438881 -1.3847475
#> 94  -3.307122712  3.2791191  0.453934126  0.28616378   0.8897834 -1.6053046
#> 95  -0.650865807  1.9883954  0.783644675  0.87247292   0.6588486 -1.7700121
#> 96  -2.670438889  2.0045269  0.122344062  0.61475086   0.2645301 -1.7163227
#> 97  -3.158883829  1.6061851 -0.493418416  1.21118338   0.4693742 -2.0636565
#> 98  -1.985829183  2.2392415  0.074194354  0.20740341   1.2468011 -2.5508240
#> 99  -2.162914526  1.8099766  0.495128831  0.44551999   0.3208185 -1.5191560
#> 100 -1.493720111  0.3743481  0.032482028  0.40316785   1.0236353 -1.6208386
#>                Z          sv          SZ
#> 1    0.712977979 -0.60786137  0.21671852
#> 2    0.869155635 -1.27786137  0.55190698
#> 3    0.428202410 -0.80922522 -0.22618978
#> 4    0.617046115 -0.46307105  0.21513524
#> 5    0.450637048 -0.77418499  1.52638406
#> 6    0.055917270 -0.78224197  0.62195012
#> 7    0.387100456 -0.93626077 -0.09580908
#> 8   -1.265487659 -0.90726480 -0.66427054
#> 9   -3.396145209 -1.23520615  1.12743985
#> 10  -0.300093270 -0.28479307 -0.23641886
#> 11   0.467629693 -0.12205094 -1.14196305
#> 12   0.060128680 -0.65327435  0.66761629
#> 13   0.496141983 -0.16739141 -0.18990922
#> 14  -0.055286983 -0.60815927 -0.31804230
#> 15  -0.480029035 -1.03257905 -0.08341233
#> 16   1.756637444 -0.23105936  0.47786992
#> 17  -0.678081290 -0.29807299 -0.01977655
#> 18   1.323181332 -0.54024038  2.38116649
#> 19   0.655984975 -0.66219073 -0.62931655
#> 20  -2.696043870 -0.92032489  0.81532261
#> 21   0.058111125 -0.86305826 -1.23490284
#> 22   0.212872758 -0.18128783  0.52600018
#> 23   0.733198953 -1.02000557  0.20561017
#> 24  -1.108583572 -0.70906966 -1.03481017
#> 25  -1.599614531 -0.56292567 -1.24180250
#> 26   0.803292378 -0.19282428  1.46215216
#> 27   1.192657301 -0.81679109  0.77595789
#> 28  -0.258799869 -0.61251383  0.46812495
#> 29  -1.205415570 -0.23612219  1.41243419
#> 30   0.472254520 -0.55931478  1.03325799
#> 31   1.211177072 -0.33639091  0.22878520
#> 32  -0.536647323 -0.89034879 -0.07774284
#> 33   0.793427316 -0.32343359 -0.58096979
#> 34   0.304598426 -0.82203369  0.34888399
#> 35   1.243018627 -0.39039663 -0.25241420
#> 36   0.153340320 -0.48426853  0.96147752
#> 37   0.185093188 -0.89468244 -0.04696583
#> 38  -0.597509530  0.04752441  0.40892942
#> 39   0.138348528 -0.82507460  0.78376354
#> 40   0.618567645 -0.17184673  0.84585878
#> 41  -0.148791972 -0.72118613  0.77919034
#> 42  -2.330329700 -0.88979276 -0.21219211
#> 43   0.041375224 -0.67744333 -1.03287623
#> 44   1.502085523 -0.76270024  0.08750213
#> 45  -0.559020572 -1.30754140  0.34007039
#> 46   0.545912943 -1.01207174  0.75825758
#> 47  -0.484312546 -0.46556257 -0.06628784
#> 48   0.407057764 -0.11212625 -0.40834835
#> 49  -1.915801517 -0.85239297 -0.55890749
#> 50  -0.974153664 -1.00845924 -0.42437162
#> 51   0.511258464 -0.75373846 -1.45204779
#> 52  -0.316764797 -0.59070923  0.63209040
#> 53   0.904000765 -0.44994620 -1.01763462
#> 54   0.191748236  0.10282283 -1.55676591
#> 55  -0.698034448 -1.63457783 -0.57210750
#> 56   1.145980571 -1.03542593 -1.09945142
#> 57  -1.101982588 -1.17803214 -0.70199277
#> 58  -0.589056751 -0.54194017 -0.76196050
#> 59  -0.197964484 -0.74458552  0.90455740
#> 60  -0.715937269 -0.17750760 -1.11915026
#> 61   0.449758973 -1.15713329  0.48851909
#> 62  -0.611511816  0.32879764  1.44495545
#> 63  -0.670128499 -1.07125582  0.38563290
#> 64   0.302032378 -0.58582688  0.58750277
#> 65   0.223510625 -0.57950534  2.09984441
#> 66  -1.585470559 -0.82062456  0.96175765
#> 67  -0.255653042 -0.11704522  0.68944448
#> 68  -0.947272461 -0.27466197  1.61338008
#> 69   0.247558301 -0.26133064  0.61097251
#> 70  -0.088861330  0.01688488  1.32050120
#> 71   1.497709903 -1.03009553 -2.23778858
#> 72   0.459704046 -0.14872598  0.28191763
#> 73  -0.260822520 -0.51489387  0.05130185
#> 74  -0.685353123 -0.96507069  2.69297148
#> 75  -0.660893807 -0.13986326  2.33289153
#> 76   0.729952287 -0.98568173  1.96421696
#> 77   0.665246653 -0.77179610  0.06257606
#> 78  -0.964635100 -0.96192475  1.11476435
#> 79  -0.628486734 -0.38438966  1.58795105
#> 80   0.565414032 -0.88695047 -0.65209067
#> 81  -1.043824364 -1.04676629  0.57145558
#> 82  -1.624476605 -1.06249467 -1.14691917
#> 83   0.408801050 -0.93566122  0.93281915
#> 84  -0.224667536 -0.58075958  0.61610267
#> 85   0.177827761 -0.41627753 -0.54778962
#> 86  -1.320993171 -1.13371575 -0.34330520
#> 87  -0.375948958 -0.89961647  1.39471584
#> 88  -0.110270362 -0.65178270 -0.54229834
#> 89  -0.824503284 -0.32999976  1.69085438
#> 90  -0.083421332 -1.05494408  0.16282822
#> 91   1.864540500 -0.48271429 -0.92871289
#> 92  -0.984525782 -1.28216476 -0.90588078
#> 93   0.221408207 -1.00902410  1.09312692
#> 94   0.661116467 -0.61929700  1.66523176
#> 95  -1.355031254 -0.79129935  0.31210400
#> 96   1.442483798  0.02242611  0.19905445
#> 97  -0.073248170 -1.09466110 -0.46883346
#> 98  -0.460817715 -0.46282628  0.98628203
#> 99  -0.005062841 -0.54659204  0.16047408
#> 100  1.040625127 -1.15102110  0.35378125
# For posterior inference:
# Get 100 samples of the group-level mean (the default)
parameters(samples_LNR, N = 100)
#>              m      m_lMd            s         t0
#> 1   -0.9146088 -0.3850223 -0.664040270 -1.8410717
#> 2   -0.9941253 -0.2444859  0.371749568 -1.5556010
#> 3   -0.9841995 -0.5355252 -0.550698347 -1.6809134
#> 4   -0.9676262 -0.6572037 -0.832480636 -1.7133224
#> 5   -0.9190563 -0.3642625 -0.724456609 -1.8387217
#> 6   -1.0461187 -0.7078570 -0.033257254 -1.3872088
#> 7   -0.9277310 -0.3831552 -0.459160361 -1.7399542
#> 8   -0.9513839 -0.4714232 -0.480700013 -1.6556521
#> 9   -0.8572746 -0.6240926 -0.435795797 -1.4837029
#> 10  -0.9861085 -0.5954253 -0.211951138 -1.7573114
#> 11  -0.9910828 -0.4916690 -0.678255229 -1.7340745
#> 12  -0.9695477 -0.6129801 -0.337418446 -1.5400205
#> 13  -1.0042034 -0.5831939 -0.514396623 -1.5816723
#> 14  -1.0194009 -0.6145868 -0.605175468 -1.5473513
#> 15  -1.0230041 -0.5705763 -0.431953409 -1.5743094
#> 16  -1.0079332 -0.4510322 -0.605110403 -1.6922836
#> 17  -1.0257342 -0.6975737 -0.365160658 -1.4434355
#> 18  -1.1044617 -0.6385537 -0.552249500 -1.4989429
#> 19  -0.9064463 -0.4231083 -0.484520997 -1.7482803
#> 20  -0.9192528 -0.4485982 -0.529878284 -1.7007512
#> 21  -1.0988782 -0.8576586 -0.262228404 -1.5317319
#> 22  -1.0466780 -0.6232451 -0.338501417 -1.4843499
#> 23  -0.9800735 -0.4551597 -0.782513432 -1.7881077
#> 24  -0.9853338 -0.6157689 -0.343240505 -1.5195546
#> 25  -0.9293706 -0.6932581 -0.496238630 -1.4980952
#> 26  -0.8737476 -0.3825559 -0.679738078 -1.7624119
#> 27  -0.9117045 -0.4987366 -0.560940770 -1.4435876
#> 28  -0.9734405 -0.6395901 -0.456444677 -1.4762271
#> 29  -0.8998395 -0.4531850 -0.615774368 -1.6593293
#> 30  -1.0259341 -0.5212213 -0.001625044 -1.0778357
#> 31  -0.8626011 -0.5121629 -0.677280467 -1.8586670
#> 32  -0.8277997 -0.5090940 -0.610916550 -1.5666532
#> 33  -1.0414085 -0.7443050 -0.263366157 -1.2634176
#> 34  -0.9642390 -0.7225266 -0.427247600 -1.4529501
#> 35  -0.9051584 -0.5229238 -0.600904180 -1.6510283
#> 36  -0.9116982 -0.4353258 -0.430103240 -1.4469101
#> 37  -0.9161523 -0.5163525 -0.799283442 -1.7260853
#> 38  -1.0783236 -0.5366131 -0.207138886 -1.8251725
#> 39  -0.9841432 -0.3662825 -0.626329639 -1.4574196
#> 40  -0.9896126 -0.8263568 -0.093874850 -1.5855461
#> 41  -0.9137349 -0.6119360 -0.452247371 -1.7007970
#> 42  -1.0558198 -0.1662891 -0.553654947 -1.6245024
#> 43  -1.0510783 -0.5140499 -0.726001050 -1.5891954
#> 44  -0.9866593 -0.3869422 -0.643268863 -1.5137897
#> 45  -1.0264733 -0.6369525  0.555814552 -1.2753606
#> 46  -1.0065390 -0.5352194 -0.453555418 -1.5240538
#> 47  -0.9300713 -0.4170714 -0.862412694 -1.9710515
#> 48  -0.8534653 -0.3924058 -0.770217238 -2.0287861
#> 49  -1.0059523 -0.4712395 -0.543727371 -1.7189973
#> 50  -0.8581448 -0.4307437 -0.670850520 -1.6182677
#> 51  -0.9886138 -0.6179968 -0.513010225 -1.3857140
#> 52  -1.0406302 -0.5720488 -0.496739142 -0.9967075
#> 53  -0.9181078 -0.4369147 -0.530992473 -1.6685095
#> 54  -0.9730604 -0.5105100 -0.362186636 -1.4173826
#> 55  -0.8274918 -0.4120862 -0.681308635 -1.6754446
#> 56  -0.7603805 -0.8284783 -0.463054469 -1.3243242
#> 57  -1.0348299 -0.5973703 -0.536815454 -1.1480310
#> 58  -0.9148597 -0.3934055 -0.577226414 -1.5949066
#> 59  -0.8545427 -0.3595833 -0.811737966 -1.9455422
#> 60  -0.8698081 -0.7230220 -0.146109178 -1.3224615
#> 61  -0.9636843 -0.5873008 -0.457137808 -1.4952626
#> 62  -0.8777581 -0.4529871 -0.534055300 -1.6207601
#> 63  -0.9712346 -0.4076718 -0.492285084 -1.4988493
#> 64  -0.9767004 -0.3401362 -0.277333433 -1.3126094
#> 65  -0.9878151 -0.3798807 -0.355185002 -1.6101835
#> 66  -0.9817911 -0.4852259 -0.751998999 -1.6466184
#> 67  -1.0061947 -0.4963381 -0.595785398 -1.4777439
#> 68  -0.9512705 -0.5650219 -0.794691262 -1.6544347
#> 69  -1.0049598 -0.5363540 -0.617304258 -1.6644720
#> 70  -1.0633382 -0.5828631 -0.500995125 -1.1943876
#> 71  -0.9017430 -0.4895326 -0.811333340 -1.9966467
#> 72  -0.8799141 -0.4178881 -1.568852686 -2.2192980
#> 73  -0.9673121 -0.3700478 -0.376908541 -1.4445447
#> 74  -1.0461377 -0.5829625 -0.340333300 -1.2703171
#> 75  -1.0408817 -0.4852213 -0.530418213 -1.8654642
#> 76  -0.9186060 -0.5537903  0.036959415 -1.4770908
#> 77  -0.9980693 -0.3007075 -0.545895116 -1.5933500
#> 78  -0.9961828 -0.4783328 -0.412937534 -1.5070930
#> 79  -0.9556526 -0.4549435 -0.582801146 -1.7199749
#> 80  -0.8677118 -0.3995471 -0.779926375 -1.8312638
#> 81  -0.9099270 -0.4411030 -0.693955341 -1.5802705
#> 82  -0.8998204 -0.3898504 -0.455149396 -1.4672622
#> 83  -1.1563352 -0.5130709 -0.587314154 -1.5169530
#> 84  -0.9230388 -0.5560582 -0.711911777 -1.7103141
#> 85  -0.9980692 -0.6549247 -0.168682184 -1.4299187
#> 86  -0.8648503 -0.3675329 -0.624764059 -1.8065595
#> 87  -0.9424133 -0.6586667 -0.601234693 -1.7758015
#> 88  -1.0116358 -0.5023588 -0.625991643 -1.6793134
#> 89  -0.8850977 -0.4429208 -0.746453325 -1.8399898
#> 90  -0.9027347 -0.4281748 -0.872623651 -2.0240879
#> 91  -0.9652579 -0.6039398 -0.427186128 -1.5520007
#> 92  -0.9869371 -0.3912582 -0.356381741 -1.4254665
#> 93  -1.0288597 -0.5574943 -0.609391935 -1.6057949
#> 94  -0.8987013 -0.3435826 -1.244930436 -1.7604910
#> 95  -0.9695426 -0.3554067 -0.599545906 -1.6260407
#> 96  -0.9049838 -0.4807892 -1.030472041 -1.6646987
#> 97  -1.1078543 -0.5819023 -0.359072915 -1.3696871
#> 98  -0.9985581 -0.6150816 -0.582057222 -1.4671384
#> 99  -0.9744157 -0.5304345 -0.863804777 -1.7048890
#> 100 -0.9938976 -0.5874044 -0.520040537 -1.5924651
# or from the individual-level parameters and mapped
parameters(samples_LNR, selection = "alpha", map = TRUE)
#>     subjects   m_lMTRUE  m_lMFALSE         s         t0
#> 1       as1t -1.0812710 -0.7277887 0.3994819 0.14587311
#> 2       as1t -1.1927848 -0.8287467 0.4695067 0.18175759
#> 3       as1t -0.9976145 -0.7202211 0.3856415 0.12097891
#> 4       as1t -1.0978570 -0.7911282 0.4084400 0.15740867
#> 5       as1t -1.0131582 -0.7179012 0.3848449 0.13311799
#> 6       as1t -1.0505207 -0.7281037 0.3976677 0.14393812
#> 7       as1t -1.0158792 -0.7246570 0.3890709 0.13245220
#> 8       as1t -1.0575972 -0.7357514 0.4034771 0.14869928
#> 9       as1t -1.1714480 -0.7140168 0.4779474 0.17943920
#> 10      as1t -1.0944301 -0.7488639 0.4153250 0.15314360
#> 11      as1t -0.9818833 -0.6843178 0.3835597 0.11710325
#> 12      as1t -1.0627788 -0.7507021 0.4209924 0.14758228
#> 13      as1t -1.1601617 -0.7982509 0.4426350 0.17385402
#> 14      as1t -1.0066017 -0.6964161 0.3939327 0.12747945
#> 15      as1t -1.0550187 -0.7206447 0.3983786 0.14257473
#> 16      as1t -0.9356744 -0.6200347 0.3426346 0.08650444
#> 17      as1t -0.9706621 -0.6890028 0.3592614 0.11426169
#> 18      as1t -0.9396259 -0.6568684 0.3361176 0.10233290
#> 19      as1t -1.1644111 -0.8193501 0.4613872 0.18106128
#> 20      as1t -1.1235896 -0.8053578 0.4176246 0.16262940
#> 21      as1t -1.0429946 -0.7466685 0.3841315 0.13975929
#> 22      as1t -1.0140372 -0.6702062 0.4010762 0.12639932
#> 23      as1t -1.0083520 -0.6912084 0.3959483 0.13180238
#> 24      as1t -1.0508207 -0.7750991 0.4117809 0.14681329
#> 25      as1t -0.9685613 -0.6737312 0.3778300 0.11450191
#> 26      as1t -1.0816797 -0.8043654 0.3981443 0.15466963
#> 27      as1t -0.9557355 -0.6432555 0.3595859 0.10444925
#> 28      as1t -1.0860453 -0.7282699 0.4082023 0.14714103
#> 29      as1t -1.0453288 -0.7184219 0.3966287 0.14244458
#> 30      as1t -1.0850409 -0.7838004 0.4111004 0.15236180
#> 31      as1t -1.0782371 -0.7432715 0.4177492 0.14934405
#> 32      as1t -1.1481693 -0.7978975 0.4465416 0.17403224
#> 33      as1t -1.0536801 -0.7118105 0.3888430 0.13737296
#> 34      as1t -0.9973113 -0.6827778 0.3714453 0.12894100
#> 35      as1t -1.0204699 -0.7236076 0.3871919 0.12995738
#> 36      as1t -1.0533768 -0.7282748 0.3986877 0.13789124
#> 37      as1t -1.0671334 -0.7040465 0.3944528 0.14234769
#> 38      as1t -1.0917418 -0.7685141 0.4057850 0.15267277
#> 39      as1t -1.0165614 -0.6921395 0.3667630 0.12093129
#> 40      as1t -0.9180315 -0.6247794 0.3436162 0.09376994
#> 41      as1t -1.1444651 -0.7873265 0.4484405 0.17473209
#> 42      as1t -1.1179900 -0.7584687 0.4154498 0.15999723
#> 43      as1t -1.0324668 -0.7211060 0.3760694 0.13375470
#> 44      as1t -1.0503993 -0.6921217 0.3978372 0.13801015
#> 45      as1t -1.0162114 -0.6892200 0.4012955 0.13176902
#> 46      as1t -0.9374605 -0.6582357 0.3469012 0.09959523
#> 47      as1t -1.0732859 -0.7654202 0.4153920 0.15647599
#> 48      as1t -0.9690308 -0.7301443 0.3445304 0.10726287
#> 49      as1t -1.0628829 -0.7235195 0.4125247 0.14573471
#> 50      as1t -1.0444543 -0.7647819 0.3772501 0.13547468
#> 51      as1t -1.0208620 -0.7287484 0.3745667 0.12611457
#> 52      as1t -1.0500208 -0.7123625 0.4007812 0.13837912
#> 53      as1t -1.0861170 -0.7106083 0.4130547 0.14735731
#> 54      as1t -1.0498222 -0.7086953 0.3675154 0.13035664
#> 55      as1t -1.0851692 -0.7227198 0.3968962 0.14317122
#> 56      as1t -0.8949335 -0.6115256 0.3247409 0.07692765
#> 57      as1t -1.1592364 -0.7834605 0.4622757 0.17700043
#> 58      as1t -1.1658688 -0.7946043 0.4507693 0.17798725
#> 59      as1t -1.0284283 -0.6991404 0.3727660 0.12824823
#> 60      as1t -0.9420525 -0.6328271 0.3498989 0.10194889
#> 61      as1t -1.0607266 -0.7232650 0.3926420 0.13640477
#> 62      as1t -1.0421362 -0.7486740 0.4004270 0.14028148
#> 63      as1t -1.2187227 -0.7856960 0.4676040 0.18559627
#> 64      as1t -1.1069825 -0.7402199 0.4109709 0.15306672
#> 65      as1t -1.0329794 -0.7005542 0.4035069 0.13413558
#> 66      as1t -1.1480770 -0.7760505 0.4409898 0.17543798
#> 67      as1t -1.1560718 -0.8202254 0.4292083 0.17320960
#> 68      as1t -1.1731374 -0.8282618 0.4658326 0.17942241
#> 69      as1t -0.9769407 -0.6736908 0.3628789 0.10916996
#> 70      as1t -1.0359703 -0.7408705 0.3788462 0.13509319
#> 71      as1t -0.9386223 -0.6346069 0.3661566 0.10788376
#> 72      as1t -1.0020604 -0.7080344 0.3510957 0.11769333
#> 73      as1t -1.0930019 -0.7860969 0.4026536 0.15570998
#> 74      as1t -1.0163103 -0.7158047 0.3970278 0.13288601
#> 75      as1t -0.9968537 -0.6849150 0.3717647 0.12145594
#> 76      as1t -1.0207142 -0.7092770 0.3896155 0.13340852
#> 77      as1t -0.9542438 -0.7006770 0.3492797 0.10954666
#> 78      as1t -1.1297477 -0.7223443 0.4626894 0.16897945
#> 79      as1t -1.0724400 -0.7624341 0.4015088 0.14685965
#> 80      as1t -1.0113343 -0.6567222 0.3830724 0.12754083
#> 81      as1t -1.0331425 -0.7453712 0.3760792 0.13308455
#> 82      as1t -0.9710671 -0.6803620 0.3669163 0.11668716
#> 83      as1t -1.0718857 -0.7264634 0.4007969 0.14660430
#> 84      as1t -1.0528358 -0.6988100 0.4008862 0.13232296
#> 85      as1t -0.9501292 -0.6559179 0.3475990 0.10776290
#> 86      as1t -1.0863682 -0.7495679 0.3863830 0.14093833
#> 87      as1t -0.9351924 -0.6208877 0.3447824 0.10073039
#> 88      as1t -1.1745143 -0.7787650 0.4556282 0.17575478
#> 89      as1t -0.9819871 -0.6727870 0.3740123 0.11024015
#> 90      as1t -1.1549062 -0.7996615 0.4067229 0.16504777
#> 91      as1t -1.0281752 -0.6960212 0.3785786 0.13424936
#> 92      as1t -1.1337259 -0.8114865 0.4494761 0.16743546
#> 93      as1t -1.1337259 -0.8114865 0.4494761 0.16743546
#> 94      as1t -1.1678195 -0.7953612 0.4371556 0.16763370
#> 95      as1t -1.1572961 -0.8084883 0.4289485 0.17208949
#> 96      as1t -1.1883865 -0.7939351 0.4547475 0.17355990
#> 97      as1t -1.1372676 -0.8191813 0.4405777 0.17691940
#> 98      as1t -1.1753062 -0.7848172 0.4412000 0.17414633
#> 99      as1t -1.1922522 -0.7827473 0.4562374 0.17577460
#> 100     as1t -1.1077914 -0.7514521 0.4471635 0.15985947
#> 101     as1t -0.9935922 -0.6703935 0.3635143 0.11671386
#> 102     as1t -1.0828127 -0.7322248 0.4049918 0.14700268
#> 103     as1t -1.1000054 -0.7467652 0.4096444 0.14611890
#> 104     as1t -0.9960951 -0.7356487 0.3696039 0.12368104
#> 105     as1t -1.0323188 -0.6983393 0.3916277 0.12821545
#> 106     as1t -1.0798551 -0.7145960 0.3902051 0.14418389
#> 107     as1t -1.1601794 -0.8116817 0.4301767 0.17309218
#> 108     as1t -0.9850241 -0.7175758 0.3661881 0.12040850
#> 109     as1t -1.1101408 -0.7679338 0.4411550 0.15450213
#> 110     as1t -0.9479930 -0.6945563 0.3661834 0.10956569
#> 111     as1t -0.9934473 -0.6827318 0.3640672 0.11466511
#> 112     as1t -0.9483408 -0.6628064 0.3735361 0.10521025
#> 113     as1t -1.1072280 -0.7596641 0.4200675 0.16208188
#> 114     as1t -1.1300659 -0.7890900 0.4242124 0.15643705
#> 115     as1t -0.9881760 -0.6885431 0.3602998 0.12165847
#> 116     as1t -1.0567540 -0.7339784 0.3813022 0.13540515
#> 117     as1t -0.9670183 -0.6873840 0.3678462 0.11464068
#> 118     as1t -0.9988525 -0.6956590 0.3710419 0.12384036
#> 119     as1t -1.1215135 -0.7954880 0.4272628 0.15999523
#> 120     as1t -1.0464739 -0.7470904 0.4145611 0.14428545
#> 121     as1t -1.0231805 -0.6901702 0.3698878 0.12646087
#> 122     as1t -1.2303328 -0.8337390 0.4799129 0.19203996
#> 123     as1t -1.0323814 -0.6947140 0.3874125 0.13656153
#> 124     as1t -1.0952481 -0.7425671 0.4050687 0.14583631
#> 125     as1t -1.0864901 -0.8006652 0.4077471 0.15404738
#> 126     as1t -0.9063947 -0.6446063 0.3515569 0.09396048
#> 127     as1t -0.9330937 -0.6693489 0.3578522 0.10615114
#> 128     as1t -0.8718164 -0.5815281 0.3247371 0.07729316
#> 129     as1t -1.0532008 -0.7315988 0.4023418 0.14468188
#> 130     as1t -1.1215224 -0.7973755 0.4312994 0.16894391
#> 131     as1t -1.0239165 -0.6963886 0.3932775 0.12979807
#> 132     as1t -1.0152930 -0.7086967 0.3819640 0.12743300
#> 133     as1t -0.9957312 -0.6723801 0.3854970 0.12301466
#> 134     as1t -0.9856176 -0.6655143 0.3544599 0.11273507
#> 135     as1t -1.0558522 -0.7473564 0.3789434 0.13504399
#> 136     as1t -1.0623647 -0.7299662 0.3914896 0.14296763
#> 137     as1t -1.0277614 -0.6744785 0.3720031 0.12897746
#> 138     as1t -1.0203903 -0.7325920 0.3794600 0.13104978
#> 139     as1t -1.0741290 -0.7518204 0.3953798 0.13814659
#> 140     as1t -0.9584139 -0.7072029 0.3446776 0.10192151
#> 141     as1t -0.9810713 -0.6906118 0.3623128 0.11504129
#> 142     as1t -1.0629195 -0.7204998 0.4098348 0.14106851
#> 143     as1t -1.1057832 -0.7904642 0.4017326 0.15350338
#> 144     as1t -1.0946099 -0.7826974 0.4318957 0.16236756
#> 145     as1t -1.0278411 -0.7304833 0.3921882 0.13248452
#> 146     as1t -1.1586397 -0.7821041 0.4358731 0.16994692
#> 147     as1t -1.0460045 -0.7315587 0.4118222 0.13724928
#> 148     as1t -1.1149438 -0.7683923 0.4526474 0.17037940
#> 149     as1t -1.1296334 -0.7839532 0.4547433 0.16893229
#> 150     as1t -1.0158417 -0.7213339 0.3859374 0.12545104
#> 151     bd6t -1.2787707 -0.6888034 0.5922870 0.20525100
#> 152     bd6t -1.3332639 -0.6489162 0.6708464 0.22460003
#> 153     bd6t -1.3450362 -0.6738379 0.6507270 0.22204939
#> 154     bd6t -1.3294378 -0.7029984 0.6092945 0.21513859
#> 155     bd6t -1.4150410 -0.6739876 0.6386051 0.22637219
#> 156     bd6t -1.3368203 -0.7052648 0.6255463 0.22118470
#> 157     bd6t -1.2362299 -0.6550899 0.5933846 0.20160141
#> 158     bd6t -1.3066813 -0.6581912 0.6435090 0.22201210
#> 159     bd6t -1.2043834 -0.6807109 0.5427426 0.18147099
#> 160     bd6t -1.2974578 -0.6565538 0.6008721 0.20814503
#> 161     bd6t -1.3943704 -0.7345107 0.6354063 0.22434849
#> 162     bd6t -1.2654977 -0.6568431 0.5487216 0.19181556
#> 163     bd6t -1.2552318 -0.6587647 0.6079179 0.21270414
#> 164     bd6t -1.3103407 -0.6532264 0.6470255 0.21777298
#> 165     bd6t -1.2907618 -0.7093575 0.5861076 0.20322729
#> 166     bd6t -1.3342311 -0.7369939 0.6408531 0.22915748
#> 167     bd6t -1.3341482 -0.7047991 0.6008951 0.21333186
#> 168     bd6t -1.3536018 -0.6690501 0.6815088 0.22399306
#> 169     bd6t -1.2716782 -0.7212715 0.5872606 0.20277462
#> 170     bd6t -1.3441046 -0.6830915 0.6156575 0.21466182
#> 171     bd6t -1.2682563 -0.6725788 0.6378940 0.21165535
#> 172     bd6t -1.2915300 -0.6091518 0.6433718 0.21083692
#> 173     bd6t -1.2775918 -0.7267691 0.6003284 0.21053920
#> 174     bd6t -1.2714974 -0.6309265 0.6089958 0.20271635
#> 175     bd6t -1.3592425 -0.6692291 0.6246146 0.21625071
#> 176     bd6t -1.2663177 -0.6336787 0.5977787 0.20041854
#> 177     bd6t -1.3039529 -0.6679095 0.6674803 0.22164678
#> 178     bd6t -1.3001698 -0.6982881 0.5675402 0.20508593
#> 179     bd6t -1.2861197 -0.6350910 0.5810211 0.20641565
#> 180     bd6t -1.2811095 -0.7434912 0.6167167 0.21137466
#> 181     bd6t -1.2476165 -0.6378831 0.5913761 0.20531316
#> 182     bd6t -1.3841622 -0.6392072 0.6218352 0.21886127
#> 183     bd6t -1.1768401 -0.6636981 0.5379758 0.18138348
#> 184     bd6t -1.2924681 -0.7051875 0.5930504 0.19873271
#> 185     bd6t -1.3439873 -0.6921052 0.6494111 0.21878487
#> 186     bd6t -1.3027483 -0.6856826 0.6221828 0.21273948
#> 187     bd6t -1.2663810 -0.7031601 0.6160648 0.21454157
#> 188     bd6t -1.3416791 -0.6456038 0.6783019 0.22457594
#> 189     bd6t -1.2358846 -0.7083081 0.5815813 0.19965355
#> 190     bd6t -1.2930115 -0.6945231 0.6317825 0.21284016
#> 191     bd6t -1.3348146 -0.6572744 0.6486147 0.21923917
#> 192     bd6t -1.2764132 -0.6755755 0.6476800 0.21621000
#> 193     bd6t -1.2015365 -0.6186125 0.5392264 0.18126747
#> 194     bd6t -1.3095492 -0.6554551 0.6362480 0.21213353
#> 195     bd6t -1.2402131 -0.6732548 0.6329675 0.21219777
#> 196     bd6t -1.2381729 -0.6384377 0.5841601 0.19890426
#> 197     bd6t -1.3555587 -0.6127782 0.6160116 0.20932832
#> 198     bd6t -1.3275125 -0.7142134 0.6666365 0.22083474
#> 199     bd6t -1.2633399 -0.6683181 0.6059331 0.20790260
#> 200     bd6t -1.3672325 -0.6439884 0.6786843 0.22844612
#>  [ reached 'max' / getOption("max.print") -- omitted 250 rows ]