Posterior quantiles

Returns the quantiles of the selected parameter type. Full range of possible samples manipulations described in get_pars.

# S3 method for class 'emc'
posterior_summary(
  emc,
  selection = "mu",
  probs = c(0.025, 0.5, 0.975),
  digits = 3,
  ...
)

posterior_summary(emc, ...)

Arguments

emc

An emc object

selection

A Character vector. Indicates which parameter types to check (e.g., alpha, mu, sigma2, correlation).

probs

A vector. Indicates which quantiles to return from the posterior.

digits

Integer. How many digits to round the output to

...

Optional additional arguments that can be passed to get_pars

Value

A list of posterior quantiles for each parameter group in the selected parameter type.

Examples

posterior_summary(samples_LNR)
#> $mu
#>         2.5%    50%  97.5%
#> m     -1.132 -0.975 -0.849
#> m_lMd -0.862 -0.548 -0.250
#> s     -0.882 -0.513 -0.008
#> t0    -2.020 -1.594 -1.232
#>