Generate Posterior/Prior Predictives

Simulate n_post data sets using the posterior/prior parameter estimates

Usage

# S3 method for class 'emc.prior'
predict(object, data = NULL, n_post = 50, n_cores = 1, n_trials = NULL, ...)

# S3 method for class 'emc'
predict(
  object,
  hyper = FALSE,
  n_post = 50,
  n_cores = 1,
  stat = c("random", "mean", "median")[1],
  ...
)

Arguments

object

An emc or emc.prior object from which to generate predictives

data

A data frame needed to exactly match the original design

n_post

Integer. Number of generated datasets

n_cores

Integer. Number of cores across which there should be parallellized

n_trials

An integer. If data isn't provided (although preferred), can generate data based on n_trials per cell of design

...

Optional additional arguments passed to get_pars or make_data

hyper

Boolean. Defaults to FALSE. If TRUE, simulates from the group-level (hyper) parameters instead of the subject-level parameters.

stat

Character. Can be mean, median or random (i.e., the default). Will take either random samples from the chain(s) or use the mean or median of the parameter estimates.

Value

A list of simulated data sets of length n_post

Examples

# \donttest{
# based on an emc object ran by fit() we can generate posterior predictives
predict(samples_LNR, n_cores = 1, n_post = 10)
#>     postn subjects        E     S     R        rt trials
#> 1       1     as1t accuracy right right 0.4467349      1
#> 3       1     as1t    speed right right 0.3924724      2
#> 5       1     as1t    speed  left right 0.4627158      3
#> 7       1     as1t accuracy right right 0.4157872      4
#> 9       1     as1t accuracy  left  left 0.7261787      5
#> 11      1     as1t accuracy  left  left 0.3226543      6
#> 13      1     as1t  neutral  left  left 0.3248951      7
#> 15      1     as1t  neutral right right 0.3710702      8
#> 17      1     as1t  neutral  left right 0.5078810      9
#> 19      1     as1t accuracy right right 0.5985371     10
#> 21      1     as1t accuracy right right 0.4784350     11
#> 23      1     as1t  neutral  left right 0.4857227     12
#> 25      1     as1t accuracy  left  left 0.3832396     13
#> 27      1     as1t accuracy  left  left 0.4871498     14
#> 29      1     as1t  neutral right right 0.3716643     15
#> 31      1     as1t    speed right right 0.4781147     16
#> 33      1     as1t accuracy right right 0.4408188     17
#> 35      1     as1t    speed right right 0.7214501     18
#> 37      1     as1t accuracy right right 0.4230714     19
#> 39      1     as1t    speed  left  left 0.3618097     20
#> 41      1     as1t accuracy right right 0.4480875     21
#> 43      1     as1t  neutral  left  left 0.4926888     22
#> 45      1     as1t  neutral  left  left 0.4814469     23
#> 47      1     as1t    speed  left  left 0.4239899     24
#> 49      1     as1t accuracy right right 0.4494411     25
#> 51      1     as1t  neutral right right 0.2845162     26
#> 53      1     as1t accuracy  left  left 0.4103552     27
#> 55      1     as1t  neutral  left  left 0.4888979     28
#> 57      1     as1t  neutral right right 0.5702429     29
#> 59      1     as1t    speed  left  left 0.5434171     30
#> 61      1     as1t    speed  left right 0.4420215     31
#> 63      1     as1t  neutral right  left 0.4352314     32
#> 65      1     as1t    speed  left  left 0.3629440     33
#> 67      1     as1t accuracy  left  left 0.4363704     34
#> 69      1     as1t accuracy  left  left 0.5165859     35
#> 71      1     as1t    speed right right 0.2329370     36
#> 73      1     as1t  neutral right  left 0.4209916     37
#> 75      1     as1t accuracy right right 0.3697773     38
#> 77      1     as1t    speed  left  left 0.3620859     39
#> 79      1     as1t  neutral right right 0.4218489     40
#> 81      1     as1t    speed right right 0.4582760     41
#> 83      1     as1t    speed  left  left 0.6222035     42
#> 85      1     as1t  neutral right right 0.5501514     43
#> 87      1     as1t  neutral right right 0.5562173     44
#> 89      1     as1t  neutral right  left 0.3308578     45
#> 91      1     as1t accuracy  left  left 0.4291387     46
#> 93      1     as1t    speed  left  left 0.4106609     47
#> 95      1     as1t accuracy right  left 0.3973300     48
#> 97      1     as1t  neutral right  left 0.4200284     49
#> 99      1     as1t    speed right right 0.4578367     50
#> 101     1     as1t    speed  left  left 0.6801057     51
#> 103     1     as1t  neutral right right 0.4261262     52
#> 105     1     as1t accuracy right  left 0.4516258     53
#> 107     1     as1t  neutral  left  left 0.4514092     54
#> 109     1     as1t accuracy right right 0.4627352     55
#> 111     1     as1t    speed right right 0.4137714     56
#> 113     1     as1t    speed  left  left 0.4359555     57
#> 115     1     as1t    speed right right 0.5833709     58
#> 117     1     as1t accuracy right  left 0.5416928     59
#> 119     1     as1t accuracy right right 0.4477701     60
#> 121     1     as1t    speed right right 0.3521652     61
#> 123     1     as1t accuracy right right 0.3923629     62
#> 125     1     as1t    speed  left  left 0.7071289     63
#> 127     1     as1t  neutral  left  left 0.2635930     64
#> 129     1     as1t  neutral  left right 0.3942585     65
#> 131     1     as1t accuracy  left  left 0.4041241     66
#> 133     1     as1t    speed  left  left 0.3925475     67
#> 135     1     as1t  neutral right right 0.4267222     68
#> 137     1     as1t  neutral  left  left 0.6106971     69
#> 139     1     as1t  neutral  left  left 0.2856575     70
#> 141     1     as1t    speed right right 0.3661951     71
#> 143     1     as1t    speed  left  left 0.4539706     72
#> 145     1     as1t  neutral right right 0.6073888     73
#> 147     1     as1t  neutral right right 0.3937585     74
#> 149     1     as1t    speed  left  left 0.8532253     75
#> 151     1     as1t    speed  left  left 0.4771818     76
#> 153     1     as1t    speed right  left 0.4780734     77
#> 155     1     as1t  neutral  left  left 0.3576665     78
#> 157     1     as1t accuracy right  left 0.5329628     79
#> 159     1     as1t    speed right right 0.5859997     80
#> 161     1     as1t    speed  left right 0.3894043     81
#> 163     1     as1t accuracy right right 0.5642233     82
#> 165     1     as1t accuracy  left  left 0.7080549     83
#> 167     1     as1t  neutral right right 0.5806419     84
#> 169     1     as1t    speed right right 0.4765308     85
#> 171     1     as1t    speed right right 0.4278950     86
#> 173     1     as1t accuracy  left right 0.6930289     87
#> 175     1     as1t    speed  left  left 0.3558048     88
#> 177     1     as1t accuracy right right 0.3852970     89
#> 179     1     as1t accuracy  left right 0.6808507     90
#> 181     1     as1t accuracy right right 0.4010744     91
#> 183     1     as1t    speed  left right 0.4898041     92
#> 185     1     as1t    speed  left  left 0.3571478     93
#> 187     1     as1t accuracy right right 0.5354656     94
#> 189     1     as1t    speed  left  left 0.7466936     95
#> 191     1     as1t  neutral right right 0.3976331     96
#> 193     1     as1t  neutral  left  left 0.3756732     97
#> 195     1     as1t    speed  left  left 0.3050619     98
#> 197     1     as1t accuracy  left  left 0.4066049     99
#> 199     1     as1t  neutral right right 0.4417062    100
#> 201     1     as1t accuracy right  left 0.4312144    101
#> 203     1     as1t accuracy right right 0.3898702    102
#> 205     1     as1t  neutral right  left 0.4510897    103
#> 207     1     as1t accuracy right right 0.6247644    104
#> 209     1     as1t    speed right right 0.3660733    105
#> 211     1     as1t accuracy right  left 0.5081766    106
#> 213     1     as1t  neutral  left  left 0.3892200    107
#> 215     1     as1t  neutral right right 0.4890932    108
#> 217     1     as1t  neutral right right 0.3420377    109
#> 219     1     as1t accuracy right  left 0.6429250    110
#> 221     1     as1t  neutral right right 0.4056906    111
#> 223     1     as1t    speed  left right 0.7025811    112
#> 225     1     as1t  neutral  left  left 0.4051002    113
#> 227     1     as1t  neutral right right 0.4943891    114
#> 229     1     as1t accuracy  left  left 0.5882688    115
#> 231     1     as1t  neutral right  left 0.3699440    116
#> 233     1     as1t    speed right right 0.3704505    117
#> 235     1     as1t accuracy  left  left 0.6705428    118
#> 237     1     as1t  neutral right right 0.3159059    119
#> 239     1     as1t  neutral  left right 0.4847010    120
#> 241     1     as1t accuracy right  left 0.4324297    121
#> 243     1     as1t  neutral  left right 0.5664530    122
#> 245     1     as1t accuracy  left  left 0.4293301    123
#> 247     1     as1t accuracy right right 0.3012615    124
#> 249     1     as1t accuracy right right 0.4869081    125
#> 251     1     as1t  neutral  left right 0.5618266    126
#> 253     1     as1t accuracy right right 0.3788505    127
#> 255     1     as1t  neutral  left  left 0.3272770    128
#> 257     1     as1t    speed  left  left 0.4169653    129
#> 259     1     as1t accuracy right right 0.5691731    130
#> 261     1     as1t  neutral right right 0.4331212    131
#> 263     1     as1t  neutral right right 0.3371760    132
#> 265     1     as1t  neutral  left right 0.4359665    133
#> 267     1     as1t accuracy right right 0.3687194    134
#> 269     1     as1t    speed right  left 0.4088934    135
#> 271     1     as1t accuracy right right 0.4429542    136
#> 273     1     as1t  neutral right right 0.4764527    137
#> 275     1     as1t accuracy right right 0.4514915    138
#> 277     1     as1t  neutral  left  left 0.4647433    139
#> 279     1     as1t  neutral  left  left 0.3372161    140
#> 281     1     as1t accuracy right right 0.6395568    141
#> 283     1     as1t accuracy right right 0.5097707    142
#>  [ reached 'max' / getOption("max.print") -- omitted 24878 rows ]
# }