Runs IS2 from Tran et al. 2021 on a list of emc

Runs IS2 on a list of emc, only works for types standard, factor and diagonal yet.

Usage

run_IS2(
  emc,
  stage = "sample",
  filter = 0,
  IS_samples = 1000,
  stepsize_particles = 500,
  max_particles = 5000,
  n_cores = 1,
  df = 5
)

Arguments

emc

A list of emc

stage

A string. Indicates which stage to take samples from

filter

An integer or vector. If integer specifies how many samples to remove from within that stage. If vector used as index for samples to keep.

IS_samples

An integer. Specifies how many IS2 samples to collect

stepsize_particles

An integer. It will increase particles till optimal variance with this stepsize.

max_particles

An integer. Specifies the maximum number of particles to collect before stopping one IS iteration.

n_cores

An integer. Specifies how many cores to run IS_2 on.

df

An integer. The degrees of freedom used in the t-distribution used as IS distribution for the group-level proposals.

Value

emc, with IS2 attribute