Standardized factor loadings

Returns a set of standardized factor loadings. The standardization considers the residual error as well as described in Stevenson, Heathcote, Forstmann & Matzke, 2024.

Usage

standardize_loadings(
  emc = NULL,
  loadings = NULL,
  sig_err_inv = NULL,
  stage = "sample",
  merge_chains = TRUE
)

Arguments

emc

An emc object with samples from a hierarchical factor analysis model

loadings

Array of pars by factors by iters. Can also specify loadings instead of emc

sig_err_inv

Array of pars by iters. Can also specify sig_err_inv instead of emc

stage

Character. From which stage to take samples

merge_chains

Return the loadings for each chain separately or merged together.

Value

standardized loadings