Prior specification or prior sampling for SEM estimation.

Prior specification or prior sampling for SEM estimation.

Usage

get_prior_SEM(
  prior = NULL,
  n_pars = NULL,
  sample = TRUE,
  N = 1e+05,
  selection = "mu",
  design = NULL,
  Lambda_mat = NULL,
  B_mat = NULL,
  K_mat = NULL,
  G_mat = NULL,
  covariates = NULL
)

Arguments

prior

A named list containing the prior mean on group-level mean (theta_mu_mean), variance of group-level mean (theta_mu_var), variance of the loadings and G (lambda_var), variance of the latent regressions and (B_var), shape and rate prior on the factor variances (a_p and b_p), and shape and rate prior on the residual variances (a_e and b_e)

n_pars

Argument used by the sampler, best left NULL. In user case inferred from the design

sample

Whether to sample from the prior. Default is TRUE. If not returns a prior list

N

How many samples to draw from the prior, default 1e5

selection

Which parameter type to select e.g. alpha

design

The design obtained from design, required when map = TRUE

Lambda_mat

The loadings constraint matrix

B_mat

The latent regressions constraint matrix

K_mat

The regression on the parameters by the included covariates

G_mat

The regression on the latent factors by the included covariates

covariates

The included covariates

Value

A list with a single entry of type of samples from the prior (if sample = TRUE) or else a prior object