Make random effects

Simulates subject-level parameters in the format required by make_data().

Usage

make_random_effects(
  design,
  group_means,
  n_subj = NULL,
  variance_proportion = 0.2,
  covariances = NULL
)

Arguments

design

A design list. The design as specified by design()

group_means

A numeric vector. The group level means for each parameter, in the same order as sampled_p_vector(design)

n_subj

An integer. The number of subjects to generate parameters for. If NULL will be inferred from design

variance_proportion

A double. Optional. If covariances are not specified, the variances will be created by multiplying the means by this number. The covariances will be 0.

covariances

A covariance matrix. Optional. Specify the intended covariance matrix.

Value

A matrix of subject-level parameters.

Examples

# First create a design
design_DDMaE <- design(data = forstmann,model=DDM,
                            formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                            constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then create a group-level means vector:
group_means =c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
               t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
# Now we can create subject-level parameters
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 5)

# We can also define a covariance matrix to simulate from
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 5,
             covariances = diag(.1, length(group_means)))

# The subject level parameters can be used to generate data
make_data(subj_pars, design_DDMaE, n_trials = 10)
#>     subjects        E     S trials     R        rt
#> 1          1    speed  left      1  left 0.5626181
#> 6          1  neutral  left      1  left 0.5411673
#> 11         1 accuracy  left      1  left 0.6495204
#> 16         1    speed right      1 right 0.7249488
#> 21         1  neutral right      1 right 0.6884800
#> 26         1 accuracy right      1 right 1.0131084
#> 31         1    speed  left      2  left 0.4690735
#> 36         1  neutral  left      2  left 0.7245124
#> 41         1 accuracy  left      2  left 1.0132000
#> 46         1    speed right      2 right 0.7993010
#> 51         1  neutral right      2 right 0.7573755
#> 56         1 accuracy right      2 right 1.1834122
#> 61         1    speed  left      3  left 0.4433425
#> 66         1  neutral  left      3  left 0.6674801
#> 71         1 accuracy  left      3  left 0.9943378
#> 76         1    speed right      3  left 0.4451863
#> 81         1  neutral right      3  left 0.7960488
#> 86         1 accuracy right      3 right 1.7599061
#> 91         1    speed  left      4  left 0.5174590
#> 96         1  neutral  left      4  left 0.5285849
#> 101        1 accuracy  left      4  left 1.9909396
#> 106        1    speed right      4 right 0.5660095
#> 111        1  neutral right      4 right 1.0851174
#> 116        1 accuracy right      4 right 1.5035254
#> 121        1    speed  left      5  left 0.4822708
#> 126        1  neutral  left      5  left 0.4667385
#> 131        1 accuracy  left      5  left 0.6852064
#> 136        1    speed right      5 right 0.5051167
#> 141        1  neutral right      5  left 0.7307264
#> 146        1 accuracy right      5 right 0.8797451
#> 151        1    speed  left      6  left 1.5199362
#> 156        1  neutral  left      6  left 0.4508802
#> 161        1 accuracy  left      6  left 1.3091526
#> 166        1    speed right      6  left 0.6973122
#> 171        1  neutral right      6  left 0.7050387
#> 176        1 accuracy right      6 right 0.7856961
#> 181        1    speed  left      7  left 0.4766122
#> 186        1  neutral  left      7  left 0.4392853
#> 191        1 accuracy  left      7  left 0.9418472
#> 196        1    speed right      7 right 0.7523084
#> 201        1  neutral right      7 right 0.6127213
#> 206        1 accuracy right      7 right 0.9280032
#> 211        1    speed  left      8 right 0.6054324
#> 216        1  neutral  left      8  left 0.5457527
#> 221        1 accuracy  left      8  left 0.7612754
#> 226        1    speed right      8 right 0.5967081
#> 231        1  neutral right      8 right 0.6961190
#> 236        1 accuracy right      8 right 1.4483854
#> 241        1    speed  left      9  left 0.6320414
#> 246        1  neutral  left      9  left 0.6161114
#> 251        1 accuracy  left      9  left 1.0083369
#> 256        1    speed right      9 right 0.4786244
#> 261        1  neutral right      9 right 0.5712876
#> 266        1 accuracy right      9 right 1.0699900
#> 271        1    speed  left     10  left 0.4423910
#> 276        1  neutral  left     10  left 0.4915659
#> 281        1 accuracy  left     10  left 0.7073477
#> 286        1    speed right     10  left 0.5322527
#> 291        1  neutral right     10 right 0.4762882
#> 296        1 accuracy right     10 right 2.0545576
#> 2          2    speed  left      1  left 0.4631756
#> 7          2  neutral  left      1  left 1.4783181
#> 12         2 accuracy  left      1  left 0.6275947
#> 17         2    speed right      1  left 0.2750890
#> 22         2  neutral right      1 right 0.6886427
#> 27         2 accuracy right      1 right 0.6469094
#> 32         2    speed  left      2  left 0.6883951
#> 37         2  neutral  left      2  left 2.2581988
#> 42         2 accuracy  left      2  left 0.5878458
#> 47         2    speed right      2 right 0.3258761
#> 52         2  neutral right      2 right 0.8258317
#> 57         2 accuracy right      2 right 0.9354880
#> 62         2    speed  left      3 right 0.4356737
#> 67         2  neutral  left      3  left 0.8311204
#> 72         2 accuracy  left      3  left 0.4157404
#> 77         2    speed right      3 right 0.3132607
#> 82         2  neutral right      3 right 1.2670568
#> 87         2 accuracy right      3 right 1.2890464
#> 92         2    speed  left      4  left 0.4829956
#> 97         2  neutral  left      4  left 1.1337004
#> 102        2 accuracy  left      4  left 0.5913989
#> 107        2    speed right      4 right 0.2042694
#> 112        2  neutral right      4 right 1.1910506
#> 117        2 accuracy right      4  left 1.5535024
#> 122        2    speed  left      5  left 0.3883921
#> 127        2  neutral  left      5  left 0.9423985
#> 132        2 accuracy  left      5  left 0.6868684
#> 137        2    speed right      5 right 0.2631076
#> 142        2  neutral right      5 right 3.3936232
#> 147        2 accuracy right      5 right 0.7603863
#> 152        2    speed  left      6  left 0.2286603
#> 157        2  neutral  left      6  left 0.7466556
#> 162        2 accuracy  left      6  left 0.7400910
#> 167        2    speed right      6  left 0.3022313
#> 172        2  neutral right      6 right 0.6171688
#> 177        2 accuracy right      6 right 0.6311913
#> 182        2    speed  left      7  left 0.3183669
#> 187        2  neutral  left      7  left 0.6464847
#> 192        2 accuracy  left      7  left 0.3061423
#> 197        2    speed right      7 right 0.4239314
#> 202        2  neutral right      7 right 0.9482026
#> 207        2 accuracy right      7 right 0.6888344
#> 212        2    speed  left      8 right 0.3389831
#> 217        2  neutral  left      8 right 1.0199087
#> 222        2 accuracy  left      8  left 1.1010195
#> 227        2    speed right      8  left 0.4676308
#> 232        2  neutral right      8 right 1.3473553
#> 237        2 accuracy right      8  left 0.8941457
#> 242        2    speed  left      9  left 0.2030876
#> 247        2  neutral  left      9  left 1.6296898
#> 252        2 accuracy  left      9  left 0.8445647
#> 257        2    speed right      9 right 0.2579778
#> 262        2  neutral right      9 right 0.9998721
#> 267        2 accuracy right      9 right 4.9878070
#> 272        2    speed  left     10  left 0.5600739
#> 277        2  neutral  left     10  left 1.0725499
#> 282        2 accuracy  left     10  left 0.9547408
#> 287        2    speed right     10 right 0.3586468
#> 292        2  neutral right     10 right 1.6063006
#> 297        2 accuracy right     10 right 1.6015356
#> 3          3    speed  left      1  left 0.4445484
#> 8          3  neutral  left      1  left 0.2708726
#> 13         3 accuracy  left      1  left 0.3302341
#> 18         3    speed right      1 right 0.2591530
#> 23         3  neutral right      1 right 0.5554001
#> 28         3 accuracy right      1 right 0.4123365
#> 33         3    speed  left      2  left 0.3136867
#> 38         3  neutral  left      2 right 0.4109299
#> 43         3 accuracy  left      2  left 0.5805048
#> 48         3    speed right      2 right 0.2575735
#> 53         3  neutral right      2 right 0.2528485
#> 58         3 accuracy right      2 right 0.3218992
#> 63         3    speed  left      3  left 0.2725933
#> 68         3  neutral  left      3 right 0.7483228
#> 73         3 accuracy  left      3  left 0.3869763
#> 78         3    speed right      3  left 0.3434596
#> 83         3  neutral right      3 right 0.4083256
#> 88         3 accuracy right      3  left 0.2832851
#> 93         3    speed  left      4  left 0.2742501
#> 98         3  neutral  left      4 right 0.3058212
#> 103        3 accuracy  left      4  left 0.3408514
#> 108        3    speed right      4  left 0.2679872
#> 113        3  neutral right      4 right 0.3558533
#> 118        3 accuracy right      4  left 0.2469108
#> 123        3    speed  left      5  left 0.3845923
#> 128        3  neutral  left      5  left 0.5676488
#> 133        3 accuracy  left      5  left 0.3979271
#> 138        3    speed right      5  left 0.4023204
#> 143        3  neutral right      5 right 0.3970271
#> 148        3 accuracy right      5  left 0.2632367
#> 153        3    speed  left      6  left 0.3529369
#> 158        3  neutral  left      6  left 0.4452637
#> 163        3 accuracy  left      6  left 0.6062095
#> 168        3    speed right      6 right 0.3021751
#> 173        3  neutral right      6 right 0.5833643
#> 178        3 accuracy right      6 right 0.5679838
#> 183        3    speed  left      7  left 0.2718014
#> 188        3  neutral  left      7  left 0.3648663
#> 193        3 accuracy  left      7  left 0.2691552
#> 198        3    speed right      7  left 0.2448503
#> 203        3  neutral right      7 right 0.3377176
#> 208        3 accuracy right      7 right 0.3492104
#> 213        3    speed  left      8  left 0.2608445
#> 218        3  neutral  left      8  left 0.5737739
#> 223        3 accuracy  left      8 right 0.2451819
#> 228        3    speed right      8 right 0.2670261
#> 233        3  neutral right      8 right 0.2698834
#> 238        3 accuracy right      8 right 0.2624289
#> 243        3    speed  left      9  left 0.2634149
#> 248        3  neutral  left      9  left 0.3249518
#> 253        3 accuracy  left      9  left 0.4313342
#> 258        3    speed right      9  left 0.2701562
#> 263        3  neutral right      9 right 0.5169923
#> 268        3 accuracy right      9 right 0.5049647
#> 273        3    speed  left     10 right 0.3141882
#> 278        3  neutral  left     10  left 0.3971430
#> 283        3 accuracy  left     10  left 0.3598972
#> 288        3    speed right     10 right 0.2531382
#> 293        3  neutral right     10 right 0.3797161
#> 298        3 accuracy right     10 right 0.8425145
#> 4          4    speed  left      1  left 0.3320816
#> 9          4  neutral  left      1  left 0.4344852
#> 14         4 accuracy  left      1  left 0.2456145
#> 19         4    speed right      1 right 0.3265257
#> 24         4  neutral right      1 right 0.8307979
#> 29         4 accuracy right      1 right 0.4055589
#> 34         4    speed  left      2  left 0.3852241
#> 39         4  neutral  left      2  left 0.5803159
#> 44         4 accuracy  left      2 right 0.2937234
#> 49         4    speed right      2  left 0.2959690
#> 54         4  neutral right      2 right 0.2760100
#> 59         4 accuracy right      2 right 0.6281494
#> 64         4    speed  left      3  left 0.2830853
#> 69         4  neutral  left      3  left 0.4108235
#> 74         4 accuracy  left      3  left 0.9801867
#> 79         4    speed right      3  left 0.2628161
#> 84         4  neutral right      3 right 0.3746757
#> 89         4 accuracy right      3 right 0.2417244
#> 94         4    speed  left      4  left 0.2553360
#> 99         4  neutral  left      4  left 0.7036442
#> 104        4 accuracy  left      4  left 0.2480213
#> 109        4    speed right      4 right 0.3106226
#> 114        4  neutral right      4 right 0.3853486
#> 119        4 accuracy right      4 right 0.3351961
#> 124        4    speed  left      5  left 0.4275156
#> 129        4  neutral  left      5  left 0.3501885
#> 134        4 accuracy  left      5 right 0.3142702
#> 139        4    speed right      5 right 0.3128429
#> 144        4  neutral right      5 right 0.3736941
#> 149        4 accuracy right      5 right 0.4196393
#> 154        4    speed  left      6  left 0.4175108
#> 159        4  neutral  left      6  left 0.8289732
#> 164        4 accuracy  left      6  left 0.3038673
#> 169        4    speed right      6 right 0.3551306
#> 174        4  neutral right      6 right 0.6734445
#> 179        4 accuracy right      6 right 0.4049142
#> 184        4    speed  left      7  left 0.4450398
#> 189        4  neutral  left      7  left 0.5323653
#> 194        4 accuracy  left      7  left 0.3365786
#> 199        4    speed right      7 right 0.3114103
#> 204        4  neutral right      7 right 0.2836989
#> 209        4 accuracy right      7 right 0.2688734
#> 214        4    speed  left      8  left 0.2333799
#> 219        4  neutral  left      8  left 0.3403717
#> 224        4 accuracy  left      8  left 0.4648283
#> 229        4    speed right      8  left 0.3612722
#> 234        4  neutral right      8 right 0.3362041
#> 239        4 accuracy right      8 right 0.2636439
#> 244        4    speed  left      9  left 0.2313794
#> 249        4  neutral  left      9  left 0.3779410
#> 254        4 accuracy  left      9  left 0.3201440
#> 259        4    speed right      9 right 0.4984733
#> 264        4  neutral right      9 right 0.8040191
#> 269        4 accuracy right      9 right 0.4650286
#> 274        4    speed  left     10 right 0.2753914
#> 279        4  neutral  left     10  left 0.3034761
#> 284        4 accuracy  left     10  left 0.3095156
#> 289        4    speed right     10 right 0.2866325
#> 294        4  neutral right     10 right 0.3052985
#> 299        4 accuracy right     10 right 0.4719027
#> 5          5    speed  left      1  left 0.3165900
#> 10         5  neutral  left      1  left 3.6504155
#> 15         5 accuracy  left      1  left 2.8370911
#> 20         5    speed right      1 right 0.4479773
#> 25         5  neutral right      1 right 1.8555810
#> 30         5 accuracy right      1 right 1.1473168
#> 35         5    speed  left      2 right 1.2519028
#> 40         5  neutral  left      2  left 1.6653343
#> 45         5 accuracy  left      2  left 1.2776202
#> 50         5    speed right      2 right 1.2129976
#> 55         5  neutral right      2 right 1.5800334
#> 60         5 accuracy right      2 right 3.4042266
#> 65         5    speed  left      3  left 0.8515260
#> 70         5  neutral  left      3  left 3.4335352
#> 75         5 accuracy  left      3  left 2.6275929
#> 80         5    speed right      3 right 0.5533217
#> 85         5  neutral right      3 right 1.6803612
#> 90         5 accuracy right      3 right 0.7377166
#> 95         5    speed  left      4  left 0.5455720
#> 100        5  neutral  left      4  left 2.8930476
#> 105        5 accuracy  left      4  left 0.3696072
#> 110        5    speed right      4 right 1.0054773
#> 115        5  neutral right      4 right 2.6719195
#> 120        5 accuracy right      4 right 1.4149659
#> 125        5    speed  left      5  left 0.7938840
#> 130        5  neutral  left      5  left 2.8554688
#> 135        5 accuracy  left      5  left 3.1632208
#> 140        5    speed right      5 right 0.2910965
#> 145        5  neutral right      5 right 1.9879338
#> 150        5 accuracy right      5 right 0.9224281
#> 155        5    speed  left      6 right 0.4922105
#> 160        5  neutral  left      6  left 1.6194068
#> 165        5 accuracy  left      6  left 0.9981338
#> 170        5    speed right      6 right 1.1633593
#> 175        5  neutral right      6 right 2.3455226
#> 180        5 accuracy right      6 right 0.9878246
#> 185        5    speed  left      7  left 0.6106098
#> 190        5  neutral  left      7 right 2.9994828
#> 195        5 accuracy  left      7  left 0.7278127
#> 200        5    speed right      7 right 0.9753499
#> 205        5  neutral right      7 right 3.1286420
#> 210        5 accuracy right      7 right 0.4782711
#> 215        5    speed  left      8  left 0.7735626
#> 220        5  neutral  left      8  left 2.6016903
#> 225        5 accuracy  left      8  left 0.6836499
#> 230        5    speed right      8 right 1.7102651
#> 235        5  neutral right      8 right 2.6271065
#> 240        5 accuracy right      8 right 2.2555576
#> 245        5    speed  left      9  left 2.2643000
#> 250        5  neutral  left      9  left 3.5676857
#> 255        5 accuracy  left      9  left 0.5936393
#> 260        5    speed right      9 right 2.9667016
#> 265        5  neutral right      9 right 2.2337854
#> 270        5 accuracy right      9 right 0.7881788
#> 275        5    speed  left     10  left 0.6298306
#> 280        5  neutral  left     10  left 1.7598721
#> 285        5 accuracy  left     10  left 0.4725505
#> 290        5    speed right     10 right 0.7762505
#> 295        5  neutral right     10 right 2.0022955
#> 300        5 accuracy right     10 right 0.4084644