Generate Subject-Level Parameters

Simulates subject-level parameters in the format required by make_data().

Usage

make_random_effects(
  design,
  group_means,
  n_subj = NULL,
  variance_proportion = 0.2,
  covariances = NULL
)

Arguments

design

A design list. The design as specified by design()

group_means

A numeric vector. The group level means for each parameter, in the same order as sampled_pars(design)

n_subj

An integer. The number of subjects to generate parameters for. If NULL will be inferred from design

variance_proportion

A double. Optional. If covariances are not specified, the variances will be created by multiplying the means by this number. The covariances will be 0.

covariances

A covariance matrix. Optional. Specify the intended covariance matrix.

Value

A matrix of subject-level parameters.

Examples

# First create a design
design_DDMaE <- design(data = forstmann,model=DDM,
                            formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                            constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then create a group-level means vector:
group_means =c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
               t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
# Now we can create subject-level parameters
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 19)

# We can also define a covariance matrix to simulate from
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 19,
             covariances = diag(.1, length(group_means)))

# The subject level parameters can be used to generate data
make_data(subj_pars, design_DDMaE, n_trials = 10)
#>      subjects        E     S trials     R        rt
#> 1        as1t    speed  left      1  left 0.1739289
#> 20       as1t  neutral  left      1  left 0.1857003
#> 39       as1t accuracy  left      1  left 0.5122115
#> 58       as1t    speed right      1  left 0.2127294
#> 77       as1t  neutral right      1 right 0.2512599
#> 96       as1t accuracy right      1 right 0.7732646
#> 115      as1t    speed  left      2  left 0.1458726
#> 134      as1t  neutral  left      2  left 0.1654491
#> 153      as1t accuracy  left      2  left 0.3368329
#> 172      as1t    speed right      2  left 0.4247824
#> 191      as1t  neutral right      2 right 0.1897626
#> 210      as1t accuracy right      2 right 1.0105605
#> 229      as1t    speed  left      3  left 0.2193947
#> 248      as1t  neutral  left      3 right 0.2874850
#> 267      as1t accuracy  left      3  left 0.5245214
#> 286      as1t    speed right      3 right 0.3509188
#> 305      as1t  neutral right      3 right 0.1495483
#> 324      as1t accuracy right      3 right 0.8246675
#> 343      as1t    speed  left      4  left 0.2364307
#> 362      as1t  neutral  left      4  left 0.1666933
#> 381      as1t accuracy  left      4  left 0.7700385
#> 400      as1t    speed right      4 right 0.1921199
#> 419      as1t  neutral right      4 right 0.5071156
#> 438      as1t accuracy right      4 right 1.1669174
#> 457      as1t    speed  left      5  left 0.4437099
#> 476      as1t  neutral  left      5  left 0.2393762
#> 495      as1t accuracy  left      5  left 0.4166828
#> 514      as1t    speed right      5  left 0.1987408
#> 533      as1t  neutral right      5 right 0.1828121
#> 552      as1t accuracy right      5 right 0.9692529
#> 571      as1t    speed  left      6  left 0.2384956
#> 590      as1t  neutral  left      6  left 0.1219610
#> 609      as1t accuracy  left      6  left 0.9338652
#> 628      as1t    speed right      6 right 0.3004748
#> 647      as1t  neutral right      6 right 0.2649877
#> 666      as1t accuracy right      6 right 0.3732264
#> 685      as1t    speed  left      7  left 0.2451327
#> 704      as1t  neutral  left      7  left 0.2403125
#> 723      as1t accuracy  left      7  left 1.0741274
#> 742      as1t    speed right      7 right 0.1673596
#> 761      as1t  neutral right      7 right 0.5069372
#> 780      as1t accuracy right      7 right 0.8375290
#> 799      as1t    speed  left      8  left 0.1386146
#> 818      as1t  neutral  left      8  left 0.4088141
#> 837      as1t accuracy  left      8  left 0.4529004
#> 856      as1t    speed right      8  left 0.1524746
#> 875      as1t  neutral right      8 right 0.2790089
#> 894      as1t accuracy right      8 right 0.6341080
#> 913      as1t    speed  left      9  left 0.1554695
#> 932      as1t  neutral  left      9  left 0.1879992
#> 951      as1t accuracy  left      9  left 1.4475942
#> 970      as1t    speed right      9 right 0.1578595
#> 989      as1t  neutral right      9 right 0.3620401
#> 1008     as1t accuracy right      9 right 0.5271057
#> 1027     as1t    speed  left     10  left 0.1389430
#> 1046     as1t  neutral  left     10 right 0.5749928
#> 1065     as1t accuracy  left     10  left 1.3962296
#> 1084     as1t    speed right     10 right 0.2340322
#> 1103     as1t  neutral right     10 right 0.5576933
#> 1122     as1t accuracy right     10  left 0.7994647
#> 2        bd6t    speed  left      1  left 0.4713983
#> 21       bd6t  neutral  left      1  left 0.2987067
#> 40       bd6t accuracy  left      1  left 1.3923496
#> 59       bd6t    speed right      1 right 0.1939515
#> 78       bd6t  neutral right      1 right 0.3400629
#> 97       bd6t accuracy right      1 right 1.0640051
#> 116      bd6t    speed  left      2 right 0.2746947
#> 135      bd6t  neutral  left      2  left 0.4954209
#> 154      bd6t accuracy  left      2  left 1.0532222
#> 173      bd6t    speed right      2  left 0.3009342
#> 192      bd6t  neutral right      2 right 1.0166187
#> 211      bd6t accuracy right      2 right 0.9897111
#> 230      bd6t    speed  left      3  left 0.5083403
#> 249      bd6t  neutral  left      3  left 0.3583790
#> 268      bd6t accuracy  left      3  left 0.8675884
#> 287      bd6t    speed right      3 right 0.2277594
#> 306      bd6t  neutral right      3 right 0.2947048
#> 325      bd6t accuracy right      3 right 0.6948552
#> 344      bd6t    speed  left      4  left 0.3793804
#> 363      bd6t  neutral  left      4  left 0.2977988
#> 382      bd6t accuracy  left      4  left 0.4178160
#> 401      bd6t    speed right      4 right 0.2224853
#> 420      bd6t  neutral right      4 right 0.7395757
#> 439      bd6t accuracy right      4 right 1.7623632
#> 458      bd6t    speed  left      5  left 0.2710349
#> 477      bd6t  neutral  left      5  left 0.5462134
#> 496      bd6t accuracy  left      5  left 1.1705083
#> 515      bd6t    speed right      5  left 0.2147056
#> 534      bd6t  neutral right      5 right 0.3577717
#> 553      bd6t accuracy right      5 right 0.3407533
#> 572      bd6t    speed  left      6  left 0.2502133
#> 591      bd6t  neutral  left      6  left 1.2994841
#> 610      bd6t accuracy  left      6  left 0.6708505
#> 629      bd6t    speed right      6 right 0.2607135
#> 648      bd6t  neutral right      6 right 0.2533927
#> 667      bd6t accuracy right      6 right 0.5549022
#> 686      bd6t    speed  left      7  left 0.2466625
#> 705      bd6t  neutral  left      7  left 0.6153619
#> 724      bd6t accuracy  left      7 right 0.4179532
#> 743      bd6t    speed right      7 right 0.2351906
#> 762      bd6t  neutral right      7 right 0.2716931
#> 781      bd6t accuracy right      7 right 0.5050941
#> 800      bd6t    speed  left      8  left 0.2450574
#> 819      bd6t  neutral  left      8  left 0.3935660
#> 838      bd6t accuracy  left      8  left 0.4750647
#> 857      bd6t    speed right      8 right 0.1912268
#> 876      bd6t  neutral right      8 right 0.9844539
#> 895      bd6t accuracy right      8 right 0.5795268
#> 914      bd6t    speed  left      9  left 0.2624711
#> 933      bd6t  neutral  left      9  left 0.7337030
#> 952      bd6t accuracy  left      9  left 0.8479317
#> 971      bd6t    speed right      9  left 0.3416137
#> 990      bd6t  neutral right      9 right 0.7615374
#> 1009     bd6t accuracy right      9 right 0.4473530
#> 1028     bd6t    speed  left     10  left 0.2220383
#> 1047     bd6t  neutral  left     10  left 0.4529179
#> 1066     bd6t accuracy  left     10  left 1.8410419
#> 1085     bd6t    speed right     10 right 0.3136252
#> 1104     bd6t  neutral right     10 right 0.3131521
#> 1123     bd6t accuracy right     10 right 0.4272405
#> 3        bl1t    speed  left      1 right 0.2622108
#> 22       bl1t  neutral  left      1  left 0.5173025
#> 41       bl1t accuracy  left      1  left 0.7713296
#> 60       bl1t    speed right      1 right 0.2571113
#> 79       bl1t  neutral right      1 right 0.3729965
#> 98       bl1t accuracy right      1 right 0.7526458
#> 117      bl1t    speed  left      2 right 0.6373499
#> 136      bl1t  neutral  left      2  left 1.3909282
#> 155      bl1t accuracy  left      2  left 0.4099769
#> 174      bl1t    speed right      2 right 0.4568756
#> 193      bl1t  neutral right      2 right 1.3672298
#> 212      bl1t accuracy right      2 right 0.4094694
#> 231      bl1t    speed  left      3  left 0.3123721
#> 250      bl1t  neutral  left      3  left 0.7101259
#> 269      bl1t accuracy  left      3  left 0.4502619
#> 288      bl1t    speed right      3 right 0.2512112
#> 307      bl1t  neutral right      3 right 0.3461823
#> 326      bl1t accuracy right      3 right 0.2716901
#> 345      bl1t    speed  left      4 right 0.2310379
#> 364      bl1t  neutral  left      4  left 0.8900620
#> 383      bl1t accuracy  left      4  left 1.4054401
#> 402      bl1t    speed right      4 right 0.2397503
#> 421      bl1t  neutral right      4 right 0.5212424
#> 440      bl1t accuracy right      4 right 0.4004643
#> 459      bl1t    speed  left      5  left 0.4522427
#> 478      bl1t  neutral  left      5  left 0.3721917
#> 497      bl1t accuracy  left      5 right 0.2582251
#> 516      bl1t    speed right      5 right 0.3522924
#> 535      bl1t  neutral right      5 right 0.5599462
#> 554      bl1t accuracy right      5 right 0.2468689
#> 573      bl1t    speed  left      6  left 0.3471776
#> 592      bl1t  neutral  left      6  left 0.6379985
#> 611      bl1t accuracy  left      6 right 0.8491322
#> 630      bl1t    speed right      6 right 0.4050914
#> 649      bl1t  neutral right      6 right 0.3029159
#> 668      bl1t accuracy right      6 right 0.2840156
#> 687      bl1t    speed  left      7 right 0.2854414
#> 706      bl1t  neutral  left      7  left 0.8823628
#> 725      bl1t accuracy  left      7  left 1.2623431
#> 744      bl1t    speed right      7 right 0.3738304
#> 763      bl1t  neutral right      7 right 0.4044063
#> 782      bl1t accuracy right      7 right 0.5923090
#> 801      bl1t    speed  left      8  left 0.4801990
#> 820      bl1t  neutral  left      8  left 0.5510942
#> 839      bl1t accuracy  left      8  left 0.7893112
#> 858      bl1t    speed right      8 right 0.2376798
#>  [ reached 'max' / getOption("max.print") -- omitted 974 rows ]