Generate Subject-Level Parameters

Simulates subject-level parameters in the format required by make_data().

Usage

make_random_effects(
  design,
  group_means,
  n_subj = NULL,
  variance_proportion = 0.2,
  covariances = NULL
)

Arguments

design

A design list. The design as specified by design()

group_means

A numeric vector. The group level means for each parameter, in the same order as sampled_pars(design)

n_subj

An integer. The number of subjects to generate parameters for. If NULL will be inferred from design

variance_proportion

A double. Optional. If covariances are not specified, the variances will be created by multiplying the means by this number. The covariances will be 0.

covariances

A covariance matrix. Optional. Specify the intended covariance matrix.

Value

A matrix of subject-level parameters.

Examples

# First create a design
design_DDMaE <- design(data = forstmann,model=DDM,
                            formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                            constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then create a group-level means vector:
group_means =c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
               t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
# Now we can create subject-level parameters
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 19)

# We can also define a covariance matrix to simulate from
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 19,
             covariances = diag(.1, length(group_means)))

# The subject level parameters can be used to generate data
make_data(subj_pars, design_DDMaE, n_trials = 10)
#>      subjects        E     S trials     R        rt
#> 1        as1t    speed  left      1  left 0.4780454
#> 20       as1t  neutral  left      1  left 0.4759104
#> 39       as1t accuracy  left      1  left 1.1718632
#> 58       as1t    speed right      1 right 0.5667276
#> 77       as1t  neutral right      1  left 0.6212035
#> 96       as1t accuracy right      1 right 2.0467163
#> 115      as1t    speed  left      2  left 0.4371223
#> 134      as1t  neutral  left      2  left 0.5120237
#> 153      as1t accuracy  left      2  left 0.4844987
#> 172      as1t    speed right      2 right 0.5619360
#> 191      as1t  neutral right      2 right 0.7386668
#> 210      as1t accuracy right      2 right 1.0041775
#> 229      as1t    speed  left      3  left 0.3953372
#> 248      as1t  neutral  left      3  left 0.4852143
#> 267      as1t accuracy  left      3  left 0.6379164
#> 286      as1t    speed right      3 right 0.6223246
#> 305      as1t  neutral right      3 right 0.5921608
#> 324      as1t accuracy right      3 right 0.8493992
#> 343      as1t    speed  left      4  left 0.4153444
#> 362      as1t  neutral  left      4  left 0.3891956
#> 381      as1t accuracy  left      4  left 0.8956678
#> 400      as1t    speed right      4 right 0.6548569
#> 419      as1t  neutral right      4 right 0.6363210
#> 438      as1t accuracy right      4 right 1.2458434
#> 457      as1t    speed  left      5  left 0.4523636
#> 476      as1t  neutral  left      5  left 0.4935807
#> 495      as1t accuracy  left      5  left 0.5039341
#> 514      as1t    speed right      5 right 0.5935295
#> 533      as1t  neutral right      5 right 0.4185438
#> 552      as1t accuracy right      5 right 1.5641369
#> 571      as1t    speed  left      6  left 0.7528991
#> 590      as1t  neutral  left      6 right 0.6255229
#> 609      as1t accuracy  left      6  left 0.7792230
#> 628      as1t    speed right      6 right 0.4568317
#> 647      as1t  neutral right      6 right 0.5242045
#> 666      as1t accuracy right      6 right 1.2483275
#> 685      as1t    speed  left      7  left 0.4178923
#> 704      as1t  neutral  left      7  left 0.3625771
#> 723      as1t accuracy  left      7  left 1.0340046
#> 742      as1t    speed right      7 right 0.6359046
#> 761      as1t  neutral right      7 right 0.6236284
#> 780      as1t accuracy right      7 right 1.4248961
#> 799      as1t    speed  left      8  left 0.3808151
#> 818      as1t  neutral  left      8  left 0.5400785
#> 837      as1t accuracy  left      8  left 1.3479134
#> 856      as1t    speed right      8  left 0.3632198
#> 875      as1t  neutral right      8 right 0.6922662
#> 894      as1t accuracy right      8 right 0.9886152
#> 913      as1t    speed  left      9  left 0.3685773
#> 932      as1t  neutral  left      9  left 0.6736121
#> 951      as1t accuracy  left      9  left 0.7993001
#> 970      as1t    speed right      9 right 0.3778616
#> 989      as1t  neutral right      9 right 0.7164222
#> 1008     as1t accuracy right      9 right 0.9590042
#> 1027     as1t    speed  left     10  left 0.4464240
#> 1046     as1t  neutral  left     10  left 0.8121620
#> 1065     as1t accuracy  left     10  left 1.3765594
#> 1084     as1t    speed right     10 right 0.9804360
#> 1103     as1t  neutral right     10  left 0.3814415
#> 1122     as1t accuracy right     10 right 2.0605514
#> 2        bd6t    speed  left      1  left 0.2169299
#> 21       bd6t  neutral  left      1 right 0.2536343
#> 40       bd6t accuracy  left      1  left 0.3581891
#> 59       bd6t    speed right      1 right 0.1761560
#> 78       bd6t  neutral right      1 right 0.2073588
#> 97       bd6t accuracy right      1 right 0.4364506
#> 116      bd6t    speed  left      2  left 0.2241493
#> 135      bd6t  neutral  left      2  left 0.2935668
#> 154      bd6t accuracy  left      2  left 0.6370671
#> 173      bd6t    speed right      2 right 0.2519634
#> 192      bd6t  neutral right      2 right 0.3276139
#> 211      bd6t accuracy right      2 right 0.5274194
#> 230      bd6t    speed  left      3  left 0.3947749
#> 249      bd6t  neutral  left      3  left 0.2321832
#> 268      bd6t accuracy  left      3  left 0.5757675
#> 287      bd6t    speed right      3 right 0.2324804
#> 306      bd6t  neutral right      3 right 0.1798221
#> 325      bd6t accuracy right      3 right 0.2828630
#> 344      bd6t    speed  left      4  left 0.2247630
#> 363      bd6t  neutral  left      4  left 0.4866059
#> 382      bd6t accuracy  left      4  left 0.7331885
#> 401      bd6t    speed right      4 right 0.2609684
#> 420      bd6t  neutral right      4 right 0.3000633
#> 439      bd6t accuracy right      4 right 0.3962064
#> 458      bd6t    speed  left      5  left 0.2513452
#> 477      bd6t  neutral  left      5 right 0.2048150
#> 496      bd6t accuracy  left      5  left 1.8523516
#> 515      bd6t    speed right      5  left 0.2498163
#> 534      bd6t  neutral right      5 right 0.2626347
#> 553      bd6t accuracy right      5 right 0.3552079
#> 572      bd6t    speed  left      6  left 0.3643340
#> 591      bd6t  neutral  left      6  left 0.3213013
#> 610      bd6t accuracy  left      6  left 0.6692881
#> 629      bd6t    speed right      6  left 0.2453622
#> 648      bd6t  neutral right      6 right 0.2272334
#> 667      bd6t accuracy right      6 right 0.8136129
#> 686      bd6t    speed  left      7 right 0.4093781
#> 705      bd6t  neutral  left      7  left 0.2054793
#> 724      bd6t accuracy  left      7  left 0.7556253
#> 743      bd6t    speed right      7 right 0.2429538
#> 762      bd6t  neutral right      7 right 0.4891264
#> 781      bd6t accuracy right      7 right 0.5773466
#> 800      bd6t    speed  left      8  left 0.3037794
#> 819      bd6t  neutral  left      8  left 0.5253532
#> 838      bd6t accuracy  left      8  left 0.5075520
#> 857      bd6t    speed right      8 right 0.1952327
#> 876      bd6t  neutral right      8 right 0.2317708
#> 895      bd6t accuracy right      8 right 0.2735095
#> 914      bd6t    speed  left      9  left 0.2176899
#> 933      bd6t  neutral  left      9  left 0.4191783
#> 952      bd6t accuracy  left      9  left 0.9260078
#> 971      bd6t    speed right      9  left 0.1981778
#> 990      bd6t  neutral right      9 right 0.2386534
#> 1009     bd6t accuracy right      9 right 0.4040468
#> 1028     bd6t    speed  left     10 right 0.4973090
#> 1047     bd6t  neutral  left     10  left 0.4087200
#> 1066     bd6t accuracy  left     10  left 0.7689248
#> 1085     bd6t    speed right     10 right 0.3019322
#> 1104     bd6t  neutral right     10 right 0.3419265
#> 1123     bd6t accuracy right     10 right 0.2853044
#> 3        bl1t    speed  left      1 right 0.6386645
#> 22       bl1t  neutral  left      1  left 0.3553747
#> 41       bl1t accuracy  left      1  left 1.0711982
#> 60       bl1t    speed right      1 right 0.3912733
#> 79       bl1t  neutral right      1 right 0.9768283
#> 98       bl1t accuracy right      1 right 0.8130434
#> 117      bl1t    speed  left      2  left 0.2899668
#> 136      bl1t  neutral  left      2  left 0.3094705
#> 155      bl1t accuracy  left      2  left 2.3864856
#> 174      bl1t    speed right      2 right 0.4462745
#> 193      bl1t  neutral right      2  left 0.3408112
#> 212      bl1t accuracy right      2 right 1.6156192
#> 231      bl1t    speed  left      3 right 1.2855407
#> 250      bl1t  neutral  left      3  left 0.6346183
#> 269      bl1t accuracy  left      3  left 0.4320356
#> 288      bl1t    speed right      3 right 0.6160627
#> 307      bl1t  neutral right      3 right 0.3145602
#> 326      bl1t accuracy right      3 right 1.0630710
#> 345      bl1t    speed  left      4  left 0.4763713
#> 364      bl1t  neutral  left      4 right 0.9182186
#> 383      bl1t accuracy  left      4  left 0.6491324
#> 402      bl1t    speed right      4 right 0.7845827
#> 421      bl1t  neutral right      4  left 0.3713295
#> 440      bl1t accuracy right      4 right 0.5309697
#> 459      bl1t    speed  left      5  left 1.1861633
#> 478      bl1t  neutral  left      5  left 0.3876705
#> 497      bl1t accuracy  left      5  left 1.3615999
#> 516      bl1t    speed right      5 right 0.3460466
#> 535      bl1t  neutral right      5 right 0.3879045
#> 554      bl1t accuracy right      5 right 0.8039241
#> 573      bl1t    speed  left      6  left 0.2950044
#> 592      bl1t  neutral  left      6  left 0.5834715
#> 611      bl1t accuracy  left      6  left 0.7628668
#> 630      bl1t    speed right      6 right 0.5552273
#> 649      bl1t  neutral right      6  left 0.6543523
#> 668      bl1t accuracy right      6 right 0.7472938
#> 687      bl1t    speed  left      7  left 0.3454584
#> 706      bl1t  neutral  left      7 right 0.2906755
#> 725      bl1t accuracy  left      7  left 1.9073226
#> 744      bl1t    speed right      7 right 0.7232619
#> 763      bl1t  neutral right      7 right 0.7012316
#> 782      bl1t accuracy right      7 right 1.5447804
#> 801      bl1t    speed  left      8 right 0.8013341
#> 820      bl1t  neutral  left      8  left 0.6742297
#> 839      bl1t accuracy  left      8  left 0.9464433
#> 858      bl1t    speed right      8 right 0.3363074
#>  [ reached 'max' / getOption("max.print") -- omitted 974 rows ]