Make random effects

Simulates subject-level parameters in the format required by make_data().

make_random_effects(
  design,
  group_means,
  n_subj = NULL,
  variance_proportion = 0.2,
  covariances = NULL
)

Arguments

design

A design list. The design as specified by design()

group_means

A numeric vector. The group level means for each parameter, in the same order as sampled_p_vector(design)

n_subj

An integer. The number of subjects to generate parameters for. If NULL will be inferred from design

variance_proportion

A double. Optional. If covariances are not specified, the variances will be created by multiplying the means by this number. The covariances will be 0.

covariances

A covariance matrix. Optional. Specify the intended covariance matrix.

Value

A matrix of subject-level parameters.

Examples

# First create a design
design_DDMaE <- design(data = forstmann,model=DDM,
                            formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                            constants=c(s=log(1)))
#> Parameter(s) DP, st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $DP
#>  DP
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then create a group-level means vector:
group_means =c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
               t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
# Now we can create subject-level parameters
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 5)

# We can also define a covariance matrix to simulate from
subj_pars <- make_random_effects(design_DDMaE, group_means, n_subj = 5,
             covariances = diag(.1, length(group_means)))

# The subject level parameters can be used to generate data
make_data(subj_pars, design_DDMaE, n_trials = 10)
#>     subjects        E     S trials     R        rt
#> 1          1    speed  left      1  left 0.6232310
#> 6          1  neutral  left      1  left 1.3394881
#> 11         1 accuracy  left      1  left 0.5715535
#> 16         1    speed right      1 right 0.3034187
#> 21         1  neutral right      1 right 0.3791782
#> 26         1 accuracy right      1 right 0.4021788
#> 31         1    speed  left      2  left 0.6615325
#> 36         1  neutral  left      2  left 0.8765126
#> 41         1 accuracy  left      2  left 0.5093003
#> 46         1    speed right      2 right 0.7704243
#> 51         1  neutral right      2 right 0.6980752
#> 56         1 accuracy right      2 right 0.6083576
#> 61         1    speed  left      3  left 0.4076057
#> 66         1  neutral  left      3 right 0.7427931
#> 71         1 accuracy  left      3  left 0.4890775
#> 76         1    speed right      3 right 0.2362027
#> 81         1  neutral right      3 right 0.5075064
#> 86         1 accuracy right      3 right 0.4662732
#> 91         1    speed  left      4 right 0.1944917
#> 96         1  neutral  left      4  left 0.8942550
#> 101        1 accuracy  left      4  left 0.6366451
#> 106        1    speed right      4 right 0.2101953
#> 111        1  neutral right      4 right 0.4594132
#> 116        1 accuracy right      4 right 0.3589235
#> 121        1    speed  left      5  left 0.3940245
#> 126        1  neutral  left      5  left 1.0738969
#> 131        1 accuracy  left      5  left 1.2781015
#> 136        1    speed right      5 right 0.4357317
#> 141        1  neutral right      5 right 0.8884564
#> 146        1 accuracy right      5 right 0.4318380
#> 151        1    speed  left      6  left 0.4340867
#> 156        1  neutral  left      6  left 1.0742550
#> 161        1 accuracy  left      6 right 0.3279504
#> 166        1    speed right      6 right 0.3376950
#> 171        1  neutral right      6 right 0.3337525
#> 176        1 accuracy right      6 right 0.2679895
#> 181        1    speed  left      7  left 0.7224251
#> 186        1  neutral  left      7  left 1.0789464
#> 191        1 accuracy  left      7  left 1.0966319
#> 196        1    speed right      7 right 0.2141433
#> 201        1  neutral right      7 right 0.3618501
#> 206        1 accuracy right      7 right 0.3505812
#> 211        1    speed  left      8  left 0.5560162
#> 216        1  neutral  left      8  left 0.9359425
#> 221        1 accuracy  left      8  left 1.3948838
#> 226        1    speed right      8 right 0.6517508
#> 231        1  neutral right      8 right 0.7398173
#> 236        1 accuracy right      8 right 0.3849400
#> 241        1    speed  left      9  left 0.4462445
#> 246        1  neutral  left      9  left 0.7554662
#> 251        1 accuracy  left      9  left 0.4602529
#> 256        1    speed right      9 right 0.7118884
#> 261        1  neutral right      9 right 0.7449842
#> 266        1 accuracy right      9 right 0.7130970
#> 271        1    speed  left     10  left 0.3385228
#> 276        1  neutral  left     10  left 0.4893777
#> 281        1 accuracy  left     10  left 0.4701247
#> 286        1    speed right     10 right 0.2933959
#> 291        1  neutral right     10 right 0.4904249
#> 296        1 accuracy right     10 right 0.4459100
#> 2          2    speed  left      1  left 0.3224898
#> 7          2  neutral  left      1  left 0.2878762
#> 12         2 accuracy  left      1  left 2.0832921
#> 17         2    speed right      1 right 0.3946247
#> 22         2  neutral right      1 right 0.4943496
#> 27         2 accuracy right      1 right 1.0673639
#> 32         2    speed  left      2  left 0.3291044
#> 37         2  neutral  left      2  left 0.3266233
#> 42         2 accuracy  left      2  left 0.4867050
#> 47         2    speed right      2 right 0.3172491
#> 52         2  neutral right      2 right 0.5818169
#> 57         2 accuracy right      2 right 0.6175698
#> 62         2    speed  left      3  left 0.3180867
#> 67         2  neutral  left      3  left 0.2970501
#> 72         2 accuracy  left      3  left 0.4423423
#> 77         2    speed right      3 right 0.4321752
#> 82         2  neutral right      3 right 0.4194488
#> 87         2 accuracy right      3 right 0.4944659
#> 92         2    speed  left      4 right 0.3636470
#> 97         2  neutral  left      4  left 0.3011572
#> 102        2 accuracy  left      4  left 0.5242720
#> 107        2    speed right      4 right 0.4589835
#> 112        2  neutral right      4 right 1.2087834
#> 117        2 accuracy right      4 right 0.5164137
#> 122        2    speed  left      5  left 0.3895086
#> 127        2  neutral  left      5  left 0.7172295
#> 132        2 accuracy  left      5  left 1.0012334
#> 137        2    speed right      5 right 0.3592915
#> 142        2  neutral right      5 right 0.5685196
#> 147        2 accuracy right      5 right 0.6679990
#> 152        2    speed  left      6  left 0.3281427
#> 157        2  neutral  left      6  left 0.4246552
#> 162        2 accuracy  left      6  left 0.4907052
#> 167        2    speed right      6 right 1.1704374
#> 172        2  neutral right      6 right 0.9339395
#> 177        2 accuracy right      6 right 0.9187463
#> 182        2    speed  left      7  left 0.4674661
#> 187        2  neutral  left      7  left 0.3993660
#> 192        2 accuracy  left      7  left 0.4953250
#> 197        2    speed right      7  left 0.3257030
#> 202        2  neutral right      7 right 0.6198535
#> 207        2 accuracy right      7 right 0.5865298
#> 212        2    speed  left      8  left 0.4356695
#> 217        2  neutral  left      8  left 0.3113361
#> 222        2 accuracy  left      8  left 0.3334570
#> 227        2    speed right      8  left 0.2655953
#> 232        2  neutral right      8 right 0.6558630
#> 237        2 accuracy right      8 right 0.6771777
#> 242        2    speed  left      9  left 0.3905088
#> 247        2  neutral  left      9  left 0.4314375
#> 252        2 accuracy  left      9  left 0.6076621
#> 257        2    speed right      9 right 0.4295611
#> 262        2  neutral right      9  left 1.1280433
#> 267        2 accuracy right      9 right 1.0700524
#> 272        2    speed  left     10  left 0.3277739
#> 277        2  neutral  left     10  left 0.3182598
#> 282        2 accuracy  left     10  left 0.4929671
#> 287        2    speed right     10 right 0.5393919
#> 292        2  neutral right     10 right 0.4835602
#> 297        2 accuracy right     10 right 0.7095419
#> 3          3    speed  left      1 right 0.4036788
#> 8          3  neutral  left      1  left 0.5288299
#> 13         3 accuracy  left      1  left 0.7304312
#> 18         3    speed right      1 right 0.4165724
#> 23         3  neutral right      1  left 0.6131407
#> 28         3 accuracy right      1 right 0.8179621
#> 33         3    speed  left      2  left 0.3969236
#> 38         3  neutral  left      2 right 0.4977916
#> 43         3 accuracy  left      2  left 0.5632411
#> 48         3    speed right      2 right 0.4858900
#> 53         3  neutral right      2 right 0.3981415
#> 58         3 accuracy right      2 right 0.8156557
#> 63         3    speed  left      3  left 0.3594882
#> 68         3  neutral  left      3  left 0.5064802
#> 73         3 accuracy  left      3  left 0.8697025
#> 78         3    speed right      3  left 0.3392199
#> 83         3  neutral right      3  left 0.3648745
#> 88         3 accuracy right      3 right 0.6909819
#> 93         3    speed  left      4  left 0.3824697
#> 98         3  neutral  left      4  left 0.4798217
#> 103        3 accuracy  left      4  left 0.4561040
#> 108        3    speed right      4 right 0.4836050
#> 113        3  neutral right      4  left 0.3579754
#> 118        3 accuracy right      4 right 0.6850402
#> 123        3    speed  left      5 right 0.4047670
#> 128        3  neutral  left      5  left 0.3486053
#> 133        3 accuracy  left      5  left 0.5067245
#> 138        3    speed right      5  left 0.4269210
#> 143        3  neutral right      5  left 0.3414722
#> 148        3 accuracy right      5 right 0.5144289
#> 153        3    speed  left      6 right 0.3882210
#> 158        3  neutral  left      6  left 0.3849189
#> 163        3 accuracy  left      6 right 0.6661821
#> 168        3    speed right      6  left 0.5122174
#> 173        3  neutral right      6 right 0.3572419
#> 178        3 accuracy right      6  left 0.6733951
#> 183        3    speed  left      7 right 0.4440470
#> 188        3  neutral  left      7  left 0.4127052
#> 193        3 accuracy  left      7  left 0.4900228
#> 198        3    speed right      7 right 0.5260270
#> 203        3  neutral right      7  left 0.3473449
#> 208        3 accuracy right      7 right 1.0838413
#> 213        3    speed  left      8  left 0.3736533
#> 218        3  neutral  left      8  left 0.3909694
#> 223        3 accuracy  left      8 right 0.6222264
#> 228        3    speed right      8 right 0.7439781
#> 233        3  neutral right      8  left 0.4256317
#> 238        3 accuracy right      8 right 0.8249680
#> 243        3    speed  left      9  left 0.3351626
#> 248        3  neutral  left      9  left 0.3564101
#> 253        3 accuracy  left      9  left 0.5860979
#> 258        3    speed right      9 right 0.3980379
#> 263        3  neutral right      9  left 0.3944753
#> 268        3 accuracy right      9 right 0.7593190
#> 273        3    speed  left     10  left 0.3737880
#> 278        3  neutral  left     10 right 0.5637285
#> 283        3 accuracy  left     10  left 0.4285307
#> 288        3    speed right     10 right 0.3904484
#> 293        3  neutral right     10 right 0.4620540
#> 298        3 accuracy right     10 right 0.9214697
#> 4          4    speed  left      1  left 0.3133570
#> 9          4  neutral  left      1  left 0.4010461
#> 14         4 accuracy  left      1  left 0.6246272
#> 19         4    speed right      1 right 0.3508173
#> 24         4  neutral right      1 right 0.3299097
#> 29         4 accuracy right      1 right 0.4091849
#> 34         4    speed  left      2  left 0.3744009
#> 39         4  neutral  left      2  left 0.7239268
#> 44         4 accuracy  left      2  left 0.5373519
#> 49         4    speed right      2 right 0.3264315
#> 54         4  neutral right      2  left 0.9077303
#> 59         4 accuracy right      2 right 0.7364248
#> 64         4    speed  left      3  left 0.6342779
#> 69         4  neutral  left      3  left 0.5292843
#> 74         4 accuracy  left      3  left 0.5709252
#> 79         4    speed right      3 right 0.3178529
#> 84         4  neutral right      3 right 0.4035677
#> 89         4 accuracy right      3 right 0.2932927
#> 94         4    speed  left      4  left 0.3753706
#> 99         4  neutral  left      4 right 0.3935587
#> 104        4 accuracy  left      4  left 0.4202015
#> 109        4    speed right      4 right 0.6621198
#> 114        4  neutral right      4 right 0.5986903
#> 119        4 accuracy right      4 right 0.2877960
#> 124        4    speed  left      5  left 0.2843373
#> 129        4  neutral  left      5  left 0.4381422
#> 134        4 accuracy  left      5  left 0.6714608
#> 139        4    speed right      5 right 0.5683002
#> 144        4  neutral right      5 right 0.4921846
#> 149        4 accuracy right      5 right 0.3418452
#> 154        4    speed  left      6  left 0.3236530
#> 159        4  neutral  left      6  left 0.4250542
#> 164        4 accuracy  left      6  left 0.2834482
#> 169        4    speed right      6 right 0.9425174
#> 174        4  neutral right      6 right 1.0818889
#> 179        4 accuracy right      6 right 0.6084931
#> 184        4    speed  left      7  left 0.2899794
#> 189        4  neutral  left      7  left 0.3365515
#> 194        4 accuracy  left      7 right 1.1361410
#> 199        4    speed right      7 right 0.4618909
#> 204        4  neutral right      7 right 1.1889319
#> 209        4 accuracy right      7 right 0.5679557
#> 214        4    speed  left      8  left 0.3318778
#> 219        4  neutral  left      8  left 0.3442991
#> 224        4 accuracy  left      8  left 0.3420465
#> 229        4    speed right      8 right 0.3668227
#> 234        4  neutral right      8  left 0.3317381
#> 239        4 accuracy right      8 right 0.4530142
#> 244        4    speed  left      9  left 0.3342054
#> 249        4  neutral  left      9  left 0.3856291
#> 254        4 accuracy  left      9  left 0.4252648
#> 259        4    speed right      9 right 0.2705984
#> 264        4  neutral right      9 right 0.3533202
#> 269        4 accuracy right      9 right 0.8424095
#> 274        4    speed  left     10  left 0.5242166
#> 279        4  neutral  left     10  left 0.7267518
#> 284        4 accuracy  left     10  left 0.2960526
#> 289        4    speed right     10 right 0.4705045
#> 294        4  neutral right     10  left 0.4910336
#> 299        4 accuracy right     10 right 0.4821738
#> 5          5    speed  left      1  left 0.7304023
#> 10         5  neutral  left      1 right 0.7008863
#> 15         5 accuracy  left      1 right 0.2524029
#> 20         5    speed right      1  left 0.3294391
#> 25         5  neutral right      1 right 0.3112870
#> 30         5 accuracy right      1 right 0.3812034
#> 35         5    speed  left      2  left 0.2296821
#> 40         5  neutral  left      2  left 0.6173114
#> 45         5 accuracy  left      2  left 0.2677633
#> 50         5    speed right      2  left 0.2972232
#> 55         5  neutral right      2 right 0.9527664
#> 60         5 accuracy right      2 right 0.3842647
#> 65         5    speed  left      3  left 0.2666569
#> 70         5  neutral  left      3  left 0.3949535
#> 75         5 accuracy  left      3  left 0.4150665
#> 80         5    speed right      3  left 0.2297503
#> 85         5  neutral right      3 right 0.6701494
#> 90         5 accuracy right      3 right 0.5309529
#> 95         5    speed  left      4  left 0.4321995
#> 100        5  neutral  left      4  left 0.4198532
#> 105        5 accuracy  left      4  left 0.6651938
#> 110        5    speed right      4  left 0.5628292
#> 115        5  neutral right      4 right 0.4930514
#> 120        5 accuracy right      4 right 0.4777111
#> 125        5    speed  left      5  left 0.2873530
#> 130        5  neutral  left      5  left 0.3994602
#> 135        5 accuracy  left      5  left 0.4568569
#> 140        5    speed right      5  left 0.3248368
#> 145        5  neutral right      5 right 0.5892313
#> 150        5 accuracy right      5 right 0.5397438
#> 155        5    speed  left      6  left 0.2666245
#> 160        5  neutral  left      6  left 0.4140946
#> 165        5 accuracy  left      6  left 0.4441016
#> 170        5    speed right      6 right 0.4074366
#> 175        5  neutral right      6 right 0.3801889
#> 180        5 accuracy right      6 right 0.4405495
#> 185        5    speed  left      7  left 0.2528614
#> 190        5  neutral  left      7  left 0.7661201
#> 195        5 accuracy  left      7  left 0.2550960
#> 200        5    speed right      7 right 0.3559816
#> 205        5  neutral right      7 right 0.9652962
#> 210        5 accuracy right      7 right 0.4542395
#> 215        5    speed  left      8  left 0.5101407
#> 220        5  neutral  left      8  left 0.2707033
#> 225        5 accuracy  left      8  left 0.3684850
#> 230        5    speed right      8 right 0.6278082
#> 235        5  neutral right      8 right 0.8335842
#> 240        5 accuracy right      8 right 0.7289186
#> 245        5    speed  left      9  left 0.6449421
#> 250        5  neutral  left      9  left 0.2658692
#> 255        5 accuracy  left      9  left 0.3124679
#> 260        5    speed right      9 right 0.6232054
#> 265        5  neutral right      9 right 0.3722640
#> 270        5 accuracy right      9 right 0.4281642
#> 275        5    speed  left     10  left 0.5416106
#> 280        5  neutral  left     10  left 0.4509849
#> 285        5 accuracy  left     10 right 0.5580123
#> 290        5    speed right     10 right 0.6406308
#> 295        5  neutral right     10 right 0.4519287
#> 300        5 accuracy right     10 right 0.6349761