Return Data Frame of Parameters

Return Data Frame of Parameters

Usage

# S3 method for class 'emc.prior'
parameters(x, selection = "mu", N = 1000, covariates = NULL, ...)

# S3 method for class 'emc'
parameters(x, selection = "mu", N = NULL, resample = FALSE, ...)

parameters(x, ...)

Arguments

x

An emc or emc.prior object

selection

String designating parameter type (e.g. mu, sigma2, correlation, alpha)

N

Integer. How many samples to take from the posterior/prior. If NULL will return the full posterior

covariates

For priors, possible covariates in the design

...

Optional arguments that can be passed to get_pars

resample

Boolean. If TRUE will sample N samples from the posterior with replacement

Value

A data frame with one row for each sample (with a subjects column if selection = "alpha" and using draws from the posterior)

Examples

# For prior inference:
# First set up a prior
design_DDMaE <- design(data = forstmann,model=DDM,
                       formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                       constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then set up a prior using make_prior
p_vector=c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
           t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
psd <- c(v_Sleft=1,v_Sright=1,a=.3,a_Eneutral=.3,a_Eaccuracy=.3,
         t0=.4,Z=1,sv=.4,SZ=1)
# Here we left the variance prior at default
prior_DDMaE <- prior(design_DDMaE,mu_mean=p_vector,mu_sd=psd)
# Get our prior samples
parameters(prior_DDMaE, N = 100)
#>        v_Sleft   v_Sright           a   a_Eneutral a_Eaccuracy         t0
#> 1   -1.1416915  2.3173229 -0.51109224  0.283614185   1.0073462 -1.0540564
#> 2   -2.0566665  1.6554879 -0.13735483  0.605445360   0.8541901 -1.3716687
#> 3   -1.6281480  2.8599310 -0.02132921  0.891669494   0.6878761 -1.9548654
#> 4   -2.5412000  3.6703744  0.11986715  0.690212066   0.4456369 -2.5335320
#> 5   -2.1995980  0.6452539  0.04106949  0.426628317   0.7130734 -2.0994492
#> 6   -0.8141854  1.2915389 -0.25417236  0.911065482   0.7429473 -1.6394843
#> 7   -1.9554243  3.3820785  0.20488387  0.418939618   0.4583172 -1.7382465
#> 8   -4.0340593  2.1924282  0.34122036  0.652897555   1.2411294 -1.2017252
#> 9   -2.3568818  2.6255882 -0.53881663  0.357768614   0.6026777 -1.8935285
#> 10  -2.0865426  0.8777483 -0.06901595  0.295388054   0.8160726 -0.8294440
#> 11  -2.5895421  3.2893817  0.08508197  0.162953952   0.8914797 -1.4201901
#> 12  -1.8185852  2.8973622  0.23942102  0.774481995   0.5997079 -1.6051723
#> 13  -1.8888197  3.2456440  0.14252325  0.528961543   0.5562231 -1.8630478
#> 14  -0.8560396  2.0068901  0.14442980  1.055508552   0.6973099 -2.2733649
#> 15  -2.7906216  2.1750554  0.20961167  0.593896800   0.6235540 -1.9081904
#> 16   0.0376485  3.1345184  0.34589278  0.574924217   0.8340946 -1.9854354
#> 17  -2.0438247  0.8972661 -0.27798783  0.125991064   1.1895932 -2.0513701
#> 18  -2.2271236  0.7910470 -0.02321202  0.928104230   0.7115112 -1.5114151
#> 19  -2.3697410  3.0693997  0.10105016  0.565681665   0.8363711 -1.8212252
#> 20  -2.7686513  1.5085714 -0.05007807  0.888654145   1.1455259 -1.4582573
#> 21  -1.6047060  2.4950762  0.38769424  0.746886193   0.4995318 -1.0564790
#> 22  -3.1601454  1.7931506  0.14199690  0.323877825   0.4837432 -1.5181904
#> 23   0.6409475  1.4205702 -0.03650553  0.301175540   0.3349149 -1.7971514
#> 24  -1.7616339  1.4291119  0.13539733  0.544494534   0.5610069 -0.8151702
#> 25  -1.0250948  1.9730051 -0.17628314  0.722972625   0.8018822 -2.0822661
#> 26  -2.9080869  3.8355001 -0.42610663  0.243553349   0.3833743 -1.4373460
#> 27  -1.9026307  3.3950825  0.18024621  0.462172758   1.4841587 -2.1687452
#> 28  -1.5688603  2.6781407 -0.26020893  0.891305958   0.4962590 -1.6962286
#> 29  -2.9316472  0.5952880  0.20849653  0.574157574   0.4307204 -1.4764877
#> 30  -1.3349026  0.8750278 -0.08144960  0.214935605   1.0646573 -1.6059263
#> 31  -2.6222337  1.2684292 -0.05799295  0.592956299   0.6685207 -1.9434007
#> 32  -0.8890701  1.9980422 -0.44259114  0.408966813   0.5995436 -1.5115197
#> 33  -1.6238678  1.2660863  0.11730688  0.644268234   0.1235126 -1.2206128
#> 34  -3.1883196  1.4136532 -0.08760609  0.262894755   0.8273366 -1.8724887
#> 35  -1.4975294  2.5954290  0.27013346  0.556901832   0.5394953 -1.7567228
#> 36  -0.2033707  1.9215451 -0.15425167  0.505211358   0.8114512 -1.5317360
#> 37  -2.3847509  0.9972227 -0.33752041  0.205507504   0.9549056 -1.9435339
#> 38  -2.1828633  2.3420169 -0.24769638  0.215072011   0.9131628 -1.6782900
#> 39  -1.9801230 -0.4759292 -0.33342033  0.543297289   1.0931955 -1.8537433
#> 40  -3.8649923  2.4544444  0.13292260  0.096045100  -0.3185924 -1.6194341
#> 41  -1.8856117  0.4498653  0.15930145  0.315243238   0.1534734 -1.8538199
#> 42  -2.8549388  2.7477832 -0.28954649  0.069185701   0.7896863 -1.8667256
#> 43  -1.2832605  0.4145366 -0.54334041  0.644411102   0.3989040 -1.7381261
#> 44  -1.0499905  3.6605168  0.02640421  0.738416842   0.5472555 -2.3774747
#> 45  -1.3059009  1.6998868  0.15124344  0.618944057   0.9547638 -2.1899438
#> 46  -2.3498661  1.5638604  0.22483213  0.320838370   0.5199257 -1.7080390
#> 47  -3.3114075  2.7656255 -0.06268072  0.597764908   0.5878953 -2.3539272
#> 48  -2.8095152  2.8657220 -0.29964437  0.090596021   1.0138830 -1.9785056
#> 49  -1.9982267  2.0716273 -0.23840262  0.130075268  -0.1327131 -1.0517353
#> 50  -0.2617311  2.1644574  0.16814749  0.216270117   1.0151343 -1.7290031
#> 51  -2.8739061  3.0078807 -0.15836225  0.967615364   1.1045430 -1.6399890
#> 52  -2.9250205  3.7094431 -0.20125600  0.848949454   1.2603597 -1.7119251
#> 53  -2.5277487 -0.3040231 -0.44265429  0.737247964   0.1916349 -1.6145837
#> 54  -1.2195199  0.8157127  0.88850350  0.493593390   0.1902356 -0.7115241
#> 55  -3.2376238  4.0386419  0.09963984  0.565036908   0.7113416 -1.7776959
#> 56  -1.0439273  4.5144808 -0.29448265  0.124887261   0.6701893 -1.7347686
#> 57  -2.4786681  3.1151709 -0.19544434  0.389369875   0.4207758 -2.1150494
#> 58  -3.1276351  0.8505346  0.08141339  0.842318220   0.4291288 -1.0753760
#> 59  -2.4405056  0.8158104 -0.05884755  0.247592466   0.1154314 -1.6195021
#> 60  -2.7422779  3.2692950 -0.17021101  0.677222628   0.8956888 -1.4540610
#> 61  -0.7885207  2.5878075  0.19903316  0.245373674   0.2872312 -1.6854770
#> 62  -1.6240920  3.6563656 -0.28856380  0.602719484   0.5828059 -1.6285180
#> 63   0.4031052  0.6837324  0.47854901  0.174947394   0.5100035 -2.2982667
#> 64  -2.9565757  1.6677840  0.51201357  0.926172188   0.7935087 -2.4204652
#> 65  -1.7674074  1.4694188 -0.44426992  0.405195609   0.7837816 -1.3007891
#> 66  -0.6562257  3.2451656  0.65513183  0.480893527   0.5880621 -1.5345498
#> 67  -3.1505486  2.0933621  0.03228507  0.349563240   0.6230344 -1.6779733
#> 68  -4.1242073  2.8975280  0.04613808  0.134598458   0.8184197 -1.7778080
#> 69  -1.7152922  2.3196024 -0.21326146  0.587213727   0.8313903 -2.0425338
#> 70  -1.6906975  2.8452565  0.07813253  0.567685786   1.0282315 -1.8828072
#> 71   0.3024903  1.7154407 -0.05683903  1.156390220   0.7555864 -1.4139788
#> 72  -0.1853858  2.0825635  0.22030254  0.309008921   1.1666010 -1.1134159
#> 73  -2.3411469  1.3978240  0.19407482  0.095942454   0.8171683 -1.6408790
#> 74  -1.4945621  2.7712067 -0.16040983  0.297704720   0.6038956 -2.1169825
#> 75  -2.8552585  0.8579448  0.12895644  0.561823091   0.8683300 -1.4031968
#> 76  -3.3810884  1.3214145 -0.16516327  0.381713684   1.3422764 -1.7770926
#> 77  -1.6689946  1.3671982  0.31236175  0.505201758   0.3784053 -1.8305258
#> 78  -0.7991418  3.1860677 -0.07996855  0.468144054   0.9115151 -1.9145682
#> 79  -2.0410327  2.7419804 -0.05317741  0.207509546   0.9562150 -1.2625552
#> 80  -2.0393918  2.0296093 -0.01566881  0.502849427   0.2325515 -1.2548116
#> 81  -2.0397487  1.1112150  0.08309410  0.397361290   0.1227970 -1.9770988
#> 82  -2.7286654  1.2443704  0.14995361  0.140781706   0.8215604 -0.6262587
#> 83  -1.2665372  1.4667500  0.08913807  0.252919125   0.8600286 -1.4567351
#> 84  -2.8802783  1.1133775  0.06209916  0.465259324   0.9500902 -2.0429451
#> 85  -1.0154143  1.7264945  0.51034765  0.394529477   0.9310619 -1.3359902
#> 86  -3.2730017  2.5241377 -0.01997865  0.008852041   0.9078903 -2.0077633
#> 87  -2.8592207  3.5935873 -0.07105755  0.739838873   0.4659275 -1.4608585
#> 88  -2.1092097  1.9621016  0.23793416  0.697217024   0.9129896 -1.4498128
#> 89  -3.3374927  1.7384664  0.21257283  0.347217833   0.9114250 -0.5578739
#> 90  -2.2203090  1.4083945  0.18384121  0.713374380   0.3001783 -2.0770775
#> 91  -1.3951535  1.8512623 -0.24950835  0.927903848   0.7739906 -1.7977597
#> 92  -1.2702873  3.3266684 -0.22441444  0.096822574   0.5722895 -2.1882282
#> 93  -1.7368854  1.1043003  0.43448268  0.223757610   0.9532635 -1.8919066
#> 94  -3.2258873  1.2403986 -0.17581065  0.760861261   0.8225156 -1.7595372
#> 95  -0.5812340  1.6861554  0.02577961  0.461645619   0.4750593 -1.8213233
#> 96  -3.5304591  2.8200715 -0.54236213  0.020940152   0.5249138 -2.0304269
#> 97  -1.3765682  1.7634634 -0.10106752  0.638528214   0.3242086 -1.7844081
#> 98  -1.1710940  2.4888038 -0.55434429  0.228327120   0.1223080 -1.7759106
#> 99  -3.1622648  1.3191302 -0.20590958  0.730612513   1.2007045 -1.7966911
#> 100 -3.3088759  1.9938748 -0.18475475 -0.056211919   0.6214544 -1.3341724
#>                 Z           sv          SZ
#> 1   -0.8638984677 -0.909371320  0.30912023
#> 2    2.0541233894 -1.228311346 -0.20698596
#> 3    0.4818826695 -0.751215005  0.34576620
#> 4    0.5812650786 -0.442979084 -0.34008215
#> 5   -0.8349137129 -0.626396434 -0.63870871
#> 6   -0.2940999086 -0.477507377 -1.19315297
#> 7    0.6469289788 -0.337871040  0.83475227
#> 8   -1.1583939008 -0.256256463  0.19462606
#> 9    0.5007380060 -0.648034353 -0.33784912
#> 10   0.2849057156 -0.152774115 -0.30155192
#> 11  -0.7272324655 -0.918984695  1.51484310
#> 12  -0.5048216481 -0.255134068  0.26594711
#> 13   0.7220844178 -0.848887369  0.84813665
#> 14  -0.4629128021 -0.353215496 -1.73772160
#> 15   0.4581189437 -1.333436375 -0.11223163
#> 16  -0.1612423423 -0.567502297 -0.52564693
#> 17  -0.3831147620 -0.866858865  0.07107247
#> 18  -1.4324146155 -0.805316861  0.69583712
#> 19   2.1598291863 -0.734658238 -2.12067400
#> 20   0.7414332990 -0.055720798 -0.44659829
#> 21   1.7027552077 -0.828804380  1.16365125
#> 22  -0.2768808441 -1.448527347 -0.34418790
#> 23  -0.5971408184 -0.451246855  2.81696686
#> 24  -1.3307133416 -1.048439750  1.35177482
#> 25  -0.6257084146 -0.636405259 -0.74574092
#> 26   0.3958563903 -0.842546809  1.05060506
#> 27   0.7482083194 -0.789226074  2.09586065
#> 28  -2.6826756871 -1.061956970  0.07001966
#> 29  -2.4466638153 -0.466534095  0.90232769
#> 30   0.1150310111 -1.150354752 -0.59549193
#> 31   2.0715876276 -1.154112720 -0.01154834
#> 32  -0.9973483616 -0.329135201  0.17134517
#> 33  -0.7991550579 -1.606503520 -1.17503841
#> 34  -0.4565154115 -0.427815134 -0.23577554
#> 35   0.6097541968 -0.090105628  1.12474844
#> 36  -0.3377671548 -0.624018219 -1.94128380
#> 37  -0.3933288248 -1.052598353 -0.56348365
#> 38  -1.2790509883 -0.257693583 -0.27548747
#> 39  -0.8274300323 -0.958862912  0.80554582
#> 40   1.4027819973 -0.008635554  0.47106360
#> 41   0.2822113170 -1.099700451  0.72419423
#> 42   0.7390214346 -1.064820747 -0.05483303
#> 43  -0.7725700253 -0.549010449  0.86075826
#> 44   0.1977044068 -0.505258399 -0.89772447
#> 45  -0.1681122005 -0.881227292 -0.64754983
#> 46  -0.5226948297 -0.273800236  1.87535630
#> 47  -0.2153262063 -0.337785115 -0.21783185
#> 48  -0.4551347847 -0.431205096  0.82570116
#> 49   1.0121446687 -0.235702802 -2.16713233
#> 50   0.7213795110 -1.221944062  0.42394430
#> 51   1.2023072743 -0.590634026 -1.83412295
#> 52   0.6016990391 -0.975344630  3.16484811
#> 53   0.0687062481 -0.736444413 -0.65632144
#> 54   0.7804625333 -0.513406554 -0.94430963
#> 55   0.1210753424 -0.811701674 -1.60684904
#> 56  -1.3825311316 -1.119824866  0.38400137
#> 57  -0.9736121984 -0.668213206  0.74463427
#> 58   0.8136732781 -0.313922378 -0.56041921
#> 59  -0.3204286284 -1.186178352 -0.86948709
#> 60  -1.4447626430 -0.856992183 -0.71669118
#> 61  -0.3777597057 -0.320116572 -0.12122252
#> 62   0.2961113252 -0.978552909 -0.54688690
#> 63  -1.1282767089 -0.665297995  0.51448392
#> 64  -0.6709656892  0.025893659 -1.53385004
#> 65   0.9946237408 -1.121115797  0.77911140
#> 66   1.2947682402 -0.415690472  1.45826878
#> 67   1.5709801482 -0.559598551 -2.11453498
#> 68  -1.1648465954 -0.920517165  0.50156792
#> 69   0.9310756904 -0.642558527  0.40323712
#> 70   1.5754611505 -0.096358045 -1.02581103
#> 71  -0.6048014601 -0.952330364  0.26935382
#> 72  -0.6798219703 -0.631267408  0.73444558
#> 73  -0.0465357459 -0.548941622  0.59161425
#> 74  -0.3227936747 -1.139063104  0.22721919
#> 75  -0.9791328537 -0.563174553  1.06066105
#> 76  -2.4254665125 -1.591466273 -1.00358611
#> 77   0.7462236864 -0.637525948 -0.34820200
#> 78  -0.7819528836 -0.308052716  0.85558687
#> 79   0.8603516691 -0.511686346  2.21299477
#> 80   0.8198922967 -0.923429945  1.06324533
#> 81   0.0414974265 -0.339236317 -0.26264955
#> 82  -0.7441305343 -0.726185342 -0.73352957
#> 83  -1.4192382240 -0.742498303 -0.09990369
#> 84   0.6703804593 -0.839400886 -0.29477192
#> 85   1.4634651594 -0.607086497 -0.66550652
#> 86  -2.1991449070 -0.017175182  1.51279877
#> 87   1.2422185623 -1.047430860 -0.22430903
#> 88  -0.3564424355 -0.053592729 -1.05687427
#> 89   0.5273989037 -0.545298838 -1.69687950
#> 90  -0.5640147448 -0.713092139  1.03284693
#> 91  -1.9652068513  0.157719588  0.21848312
#> 92  -0.0844647605 -1.548516219 -0.62586517
#> 93  -1.1445743775 -0.399329028  2.20581415
#> 94  -0.3647812769 -0.634099945 -0.34023025
#> 95  -0.7333905262 -0.244225605 -0.17941927
#> 96   0.0004011817 -1.247099222 -0.34479792
#> 97   1.8078397557 -0.788827583  0.46925979
#> 98   0.5323025688  0.081027484  0.20871042
#> 99   0.1340172073 -0.167527862  0.12017831
#> 100 -1.8981095013 -0.283828771  2.23016676
# For posterior inference:
# Get 100 samples of the group-level mean (the default)
parameters(samples_LNR, N = 100)
#>              m      m_lMd            s         t0
#> 1   -0.9146088 -0.3850223 -0.664040270 -1.8410717
#> 2   -0.9941253 -0.2444859  0.371749568 -1.5556010
#> 3   -0.9841995 -0.5355252 -0.550698347 -1.6809134
#> 4   -0.9676262 -0.6572037 -0.832480636 -1.7133224
#> 5   -0.9190563 -0.3642625 -0.724456609 -1.8387217
#> 6   -1.0461187 -0.7078570 -0.033257254 -1.3872088
#> 7   -0.9277310 -0.3831552 -0.459160361 -1.7399542
#> 8   -0.9513839 -0.4714232 -0.480700013 -1.6556521
#> 9   -0.8572746 -0.6240926 -0.435795797 -1.4837029
#> 10  -0.9861085 -0.5954253 -0.211951138 -1.7573114
#> 11  -0.9910828 -0.4916690 -0.678255229 -1.7340745
#> 12  -0.9695477 -0.6129801 -0.337418446 -1.5400205
#> 13  -1.0042034 -0.5831939 -0.514396623 -1.5816723
#> 14  -1.0194009 -0.6145868 -0.605175468 -1.5473513
#> 15  -1.0230041 -0.5705763 -0.431953409 -1.5743094
#> 16  -1.0079332 -0.4510322 -0.605110403 -1.6922836
#> 17  -1.0257342 -0.6975737 -0.365160658 -1.4434355
#> 18  -1.1044617 -0.6385537 -0.552249500 -1.4989429
#> 19  -0.9064463 -0.4231083 -0.484520997 -1.7482803
#> 20  -0.9192528 -0.4485982 -0.529878284 -1.7007512
#> 21  -1.0988782 -0.8576586 -0.262228404 -1.5317319
#> 22  -1.0466780 -0.6232451 -0.338501417 -1.4843499
#> 23  -0.9800735 -0.4551597 -0.782513432 -1.7881077
#> 24  -0.9853338 -0.6157689 -0.343240505 -1.5195546
#> 25  -0.9293706 -0.6932581 -0.496238630 -1.4980952
#> 26  -0.8737476 -0.3825559 -0.679738078 -1.7624119
#> 27  -0.9117045 -0.4987366 -0.560940770 -1.4435876
#> 28  -0.9734405 -0.6395901 -0.456444677 -1.4762271
#> 29  -0.8998395 -0.4531850 -0.615774368 -1.6593293
#> 30  -1.0259341 -0.5212213 -0.001625044 -1.0778357
#> 31  -0.8626011 -0.5121629 -0.677280467 -1.8586670
#> 32  -0.8277997 -0.5090940 -0.610916550 -1.5666532
#> 33  -1.0414085 -0.7443050 -0.263366157 -1.2634176
#> 34  -0.9642390 -0.7225266 -0.427247600 -1.4529501
#> 35  -0.9051584 -0.5229238 -0.600904180 -1.6510283
#> 36  -0.9116982 -0.4353258 -0.430103240 -1.4469101
#> 37  -0.9161523 -0.5163525 -0.799283442 -1.7260853
#> 38  -1.0783236 -0.5366131 -0.207138886 -1.8251725
#> 39  -0.9841432 -0.3662825 -0.626329639 -1.4574196
#> 40  -0.9896126 -0.8263568 -0.093874850 -1.5855461
#> 41  -0.9137349 -0.6119360 -0.452247371 -1.7007970
#> 42  -1.0558198 -0.1662891 -0.553654947 -1.6245024
#> 43  -1.0510783 -0.5140499 -0.726001050 -1.5891954
#> 44  -0.9866593 -0.3869422 -0.643268863 -1.5137897
#> 45  -1.0264733 -0.6369525  0.555814552 -1.2753606
#> 46  -1.0065390 -0.5352194 -0.453555418 -1.5240538
#> 47  -0.9300713 -0.4170714 -0.862412694 -1.9710515
#> 48  -0.8534653 -0.3924058 -0.770217238 -2.0287861
#> 49  -1.0059523 -0.4712395 -0.543727371 -1.7189973
#> 50  -0.8581448 -0.4307437 -0.670850520 -1.6182677
#> 51  -0.9886138 -0.6179968 -0.513010225 -1.3857140
#> 52  -1.0406302 -0.5720488 -0.496739142 -0.9967075
#> 53  -0.9181078 -0.4369147 -0.530992473 -1.6685095
#> 54  -0.9730604 -0.5105100 -0.362186636 -1.4173826
#> 55  -0.8274918 -0.4120862 -0.681308635 -1.6754446
#> 56  -0.7603805 -0.8284783 -0.463054469 -1.3243242
#> 57  -1.0348299 -0.5973703 -0.536815454 -1.1480310
#> 58  -0.9148597 -0.3934055 -0.577226414 -1.5949066
#> 59  -0.8545427 -0.3595833 -0.811737966 -1.9455422
#> 60  -0.8698081 -0.7230220 -0.146109178 -1.3224615
#> 61  -0.9636843 -0.5873008 -0.457137808 -1.4952626
#> 62  -0.8777581 -0.4529871 -0.534055300 -1.6207601
#> 63  -0.9712346 -0.4076718 -0.492285084 -1.4988493
#> 64  -0.9767004 -0.3401362 -0.277333433 -1.3126094
#> 65  -0.9878151 -0.3798807 -0.355185002 -1.6101835
#> 66  -0.9817911 -0.4852259 -0.751998999 -1.6466184
#> 67  -1.0061947 -0.4963381 -0.595785398 -1.4777439
#> 68  -0.9512705 -0.5650219 -0.794691262 -1.6544347
#> 69  -1.0049598 -0.5363540 -0.617304258 -1.6644720
#> 70  -1.0633382 -0.5828631 -0.500995125 -1.1943876
#> 71  -0.9017430 -0.4895326 -0.811333340 -1.9966467
#> 72  -0.8799141 -0.4178881 -1.568852686 -2.2192980
#> 73  -0.9673121 -0.3700478 -0.376908541 -1.4445447
#> 74  -1.0461377 -0.5829625 -0.340333300 -1.2703171
#> 75  -1.0408817 -0.4852213 -0.530418213 -1.8654642
#> 76  -0.9186060 -0.5537903  0.036959415 -1.4770908
#> 77  -0.9980693 -0.3007075 -0.545895116 -1.5933500
#> 78  -0.9961828 -0.4783328 -0.412937534 -1.5070930
#> 79  -0.9556526 -0.4549435 -0.582801146 -1.7199749
#> 80  -0.8677118 -0.3995471 -0.779926375 -1.8312638
#> 81  -0.9099270 -0.4411030 -0.693955341 -1.5802705
#> 82  -0.8998204 -0.3898504 -0.455149396 -1.4672622
#> 83  -1.1563352 -0.5130709 -0.587314154 -1.5169530
#> 84  -0.9230388 -0.5560582 -0.711911777 -1.7103141
#> 85  -0.9980692 -0.6549247 -0.168682184 -1.4299187
#> 86  -0.8648503 -0.3675329 -0.624764059 -1.8065595
#> 87  -0.9424133 -0.6586667 -0.601234693 -1.7758015
#> 88  -1.0116358 -0.5023588 -0.625991643 -1.6793134
#> 89  -0.8850977 -0.4429208 -0.746453325 -1.8399898
#> 90  -0.9027347 -0.4281748 -0.872623651 -2.0240879
#> 91  -0.9652579 -0.6039398 -0.427186128 -1.5520007
#> 92  -0.9869371 -0.3912582 -0.356381741 -1.4254665
#> 93  -1.0288597 -0.5574943 -0.609391935 -1.6057949
#> 94  -0.8987013 -0.3435826 -1.244930436 -1.7604910
#> 95  -0.9695426 -0.3554067 -0.599545906 -1.6260407
#> 96  -0.9049838 -0.4807892 -1.030472041 -1.6646987
#> 97  -1.1078543 -0.5819023 -0.359072915 -1.3696871
#> 98  -0.9985581 -0.6150816 -0.582057222 -1.4671384
#> 99  -0.9744157 -0.5304345 -0.863804777 -1.7048890
#> 100 -0.9938976 -0.5874044 -0.520040537 -1.5924651
# or from the individual-level parameters and mapped
parameters(samples_LNR, selection = "alpha", map = TRUE)
#>     subjects   m_lMTRUE  m_lMFALSE         s         t0
#> 1       as1t -1.0812710 -0.7277887 0.3994819 0.14587311
#> 2       as1t -1.1927848 -0.8287467 0.4695067 0.18175759
#> 3       as1t -0.9976145 -0.7202211 0.3856415 0.12097891
#> 4       as1t -1.0978570 -0.7911282 0.4084400 0.15740867
#> 5       as1t -1.0131582 -0.7179012 0.3848449 0.13311799
#> 6       as1t -1.0505207 -0.7281037 0.3976677 0.14393812
#> 7       as1t -1.0158792 -0.7246570 0.3890709 0.13245220
#> 8       as1t -1.0575972 -0.7357514 0.4034771 0.14869928
#> 9       as1t -1.1714480 -0.7140168 0.4779474 0.17943920
#> 10      as1t -1.0944301 -0.7488639 0.4153250 0.15314360
#> 11      as1t -0.9818833 -0.6843178 0.3835597 0.11710325
#> 12      as1t -1.0627788 -0.7507021 0.4209924 0.14758228
#> 13      as1t -1.1601617 -0.7982509 0.4426350 0.17385402
#> 14      as1t -1.0066017 -0.6964161 0.3939327 0.12747945
#> 15      as1t -1.0550187 -0.7206447 0.3983786 0.14257473
#> 16      as1t -0.9356744 -0.6200347 0.3426346 0.08650444
#> 17      as1t -0.9706621 -0.6890028 0.3592614 0.11426169
#> 18      as1t -0.9396259 -0.6568684 0.3361176 0.10233290
#> 19      as1t -1.1644111 -0.8193501 0.4613872 0.18106128
#> 20      as1t -1.1235896 -0.8053578 0.4176246 0.16262940
#> 21      as1t -1.0429946 -0.7466685 0.3841315 0.13975929
#> 22      as1t -1.0140372 -0.6702062 0.4010762 0.12639932
#> 23      as1t -1.0083520 -0.6912084 0.3959483 0.13180238
#> 24      as1t -1.0508207 -0.7750991 0.4117809 0.14681329
#> 25      as1t -0.9685613 -0.6737312 0.3778300 0.11450191
#> 26      as1t -1.0816797 -0.8043654 0.3981443 0.15466963
#> 27      as1t -0.9557355 -0.6432555 0.3595859 0.10444925
#> 28      as1t -1.0860453 -0.7282699 0.4082023 0.14714103
#> 29      as1t -1.0453288 -0.7184219 0.3966287 0.14244458
#> 30      as1t -1.0850409 -0.7838004 0.4111004 0.15236180
#> 31      as1t -1.0782371 -0.7432715 0.4177492 0.14934405
#> 32      as1t -1.1481693 -0.7978975 0.4465416 0.17403224
#> 33      as1t -1.0536801 -0.7118105 0.3888430 0.13737296
#> 34      as1t -0.9973113 -0.6827778 0.3714453 0.12894100
#> 35      as1t -1.0204699 -0.7236076 0.3871919 0.12995738
#> 36      as1t -1.0533768 -0.7282748 0.3986877 0.13789124
#> 37      as1t -1.0671334 -0.7040465 0.3944528 0.14234769
#> 38      as1t -1.0917418 -0.7685141 0.4057850 0.15267277
#> 39      as1t -1.0165614 -0.6921395 0.3667630 0.12093129
#> 40      as1t -0.9180315 -0.6247794 0.3436162 0.09376994
#> 41      as1t -1.1444651 -0.7873265 0.4484405 0.17473209
#> 42      as1t -1.1179900 -0.7584687 0.4154498 0.15999723
#> 43      as1t -1.0324668 -0.7211060 0.3760694 0.13375470
#> 44      as1t -1.0503993 -0.6921217 0.3978372 0.13801015
#> 45      as1t -1.0162114 -0.6892200 0.4012955 0.13176902
#> 46      as1t -0.9374605 -0.6582357 0.3469012 0.09959523
#> 47      as1t -1.0732859 -0.7654202 0.4153920 0.15647599
#> 48      as1t -0.9690308 -0.7301443 0.3445304 0.10726287
#> 49      as1t -1.0628829 -0.7235195 0.4125247 0.14573471
#> 50      as1t -1.0444543 -0.7647819 0.3772501 0.13547468
#> 51      as1t -1.0208620 -0.7287484 0.3745667 0.12611457
#> 52      as1t -1.0500208 -0.7123625 0.4007812 0.13837912
#> 53      as1t -1.0861170 -0.7106083 0.4130547 0.14735731
#> 54      as1t -1.0498222 -0.7086953 0.3675154 0.13035664
#> 55      as1t -1.0851692 -0.7227198 0.3968962 0.14317122
#> 56      as1t -0.8949335 -0.6115256 0.3247409 0.07692765
#> 57      as1t -1.1592364 -0.7834605 0.4622757 0.17700043
#> 58      as1t -1.1658688 -0.7946043 0.4507693 0.17798725
#> 59      as1t -1.0284283 -0.6991404 0.3727660 0.12824823
#> 60      as1t -0.9420525 -0.6328271 0.3498989 0.10194889
#> 61      as1t -1.0607266 -0.7232650 0.3926420 0.13640477
#> 62      as1t -1.0421362 -0.7486740 0.4004270 0.14028148
#> 63      as1t -1.2187227 -0.7856960 0.4676040 0.18559627
#> 64      as1t -1.1069825 -0.7402199 0.4109709 0.15306672
#> 65      as1t -1.0329794 -0.7005542 0.4035069 0.13413558
#> 66      as1t -1.1480770 -0.7760505 0.4409898 0.17543798
#> 67      as1t -1.1560718 -0.8202254 0.4292083 0.17320960
#> 68      as1t -1.1731374 -0.8282618 0.4658326 0.17942241
#> 69      as1t -0.9769407 -0.6736908 0.3628789 0.10916996
#> 70      as1t -1.0359703 -0.7408705 0.3788462 0.13509319
#> 71      as1t -0.9386223 -0.6346069 0.3661566 0.10788376
#> 72      as1t -1.0020604 -0.7080344 0.3510957 0.11769333
#> 73      as1t -1.0930019 -0.7860969 0.4026536 0.15570998
#> 74      as1t -1.0163103 -0.7158047 0.3970278 0.13288601
#> 75      as1t -0.9968537 -0.6849150 0.3717647 0.12145594
#> 76      as1t -1.0207142 -0.7092770 0.3896155 0.13340852
#> 77      as1t -0.9542438 -0.7006770 0.3492797 0.10954666
#> 78      as1t -1.1297477 -0.7223443 0.4626894 0.16897945
#> 79      as1t -1.0724400 -0.7624341 0.4015088 0.14685965
#> 80      as1t -1.0113343 -0.6567222 0.3830724 0.12754083
#> 81      as1t -1.0331425 -0.7453712 0.3760792 0.13308455
#> 82      as1t -0.9710671 -0.6803620 0.3669163 0.11668716
#> 83      as1t -1.0718857 -0.7264634 0.4007969 0.14660430
#> 84      as1t -1.0528358 -0.6988100 0.4008862 0.13232296
#> 85      as1t -0.9501292 -0.6559179 0.3475990 0.10776290
#> 86      as1t -1.0863682 -0.7495679 0.3863830 0.14093833
#> 87      as1t -0.9351924 -0.6208877 0.3447824 0.10073039
#> 88      as1t -1.1745143 -0.7787650 0.4556282 0.17575478
#> 89      as1t -0.9819871 -0.6727870 0.3740123 0.11024015
#> 90      as1t -1.1549062 -0.7996615 0.4067229 0.16504777
#> 91      as1t -1.0281752 -0.6960212 0.3785786 0.13424936
#> 92      as1t -1.1337259 -0.8114865 0.4494761 0.16743546
#> 93      as1t -1.1337259 -0.8114865 0.4494761 0.16743546
#> 94      as1t -1.1678195 -0.7953612 0.4371556 0.16763370
#> 95      as1t -1.1572961 -0.8084883 0.4289485 0.17208949
#> 96      as1t -1.1883865 -0.7939351 0.4547475 0.17355990
#> 97      as1t -1.1372676 -0.8191813 0.4405777 0.17691940
#> 98      as1t -1.1753062 -0.7848172 0.4412000 0.17414633
#> 99      as1t -1.1922522 -0.7827473 0.4562374 0.17577460
#> 100     as1t -1.1077914 -0.7514521 0.4471635 0.15985947
#> 101     as1t -0.9935922 -0.6703935 0.3635143 0.11671386
#> 102     as1t -1.0828127 -0.7322248 0.4049918 0.14700268
#> 103     as1t -1.1000054 -0.7467652 0.4096444 0.14611890
#> 104     as1t -0.9960951 -0.7356487 0.3696039 0.12368104
#> 105     as1t -1.0323188 -0.6983393 0.3916277 0.12821545
#> 106     as1t -1.0798551 -0.7145960 0.3902051 0.14418389
#> 107     as1t -1.1601794 -0.8116817 0.4301767 0.17309218
#> 108     as1t -0.9850241 -0.7175758 0.3661881 0.12040850
#> 109     as1t -1.1101408 -0.7679338 0.4411550 0.15450213
#> 110     as1t -0.9479930 -0.6945563 0.3661834 0.10956569
#> 111     as1t -0.9934473 -0.6827318 0.3640672 0.11466511
#> 112     as1t -0.9483408 -0.6628064 0.3735361 0.10521025
#> 113     as1t -1.1072280 -0.7596641 0.4200675 0.16208188
#> 114     as1t -1.1300659 -0.7890900 0.4242124 0.15643705
#> 115     as1t -0.9881760 -0.6885431 0.3602998 0.12165847
#> 116     as1t -1.0567540 -0.7339784 0.3813022 0.13540515
#> 117     as1t -0.9670183 -0.6873840 0.3678462 0.11464068
#> 118     as1t -0.9988525 -0.6956590 0.3710419 0.12384036
#> 119     as1t -1.1215135 -0.7954880 0.4272628 0.15999523
#> 120     as1t -1.0464739 -0.7470904 0.4145611 0.14428545
#> 121     as1t -1.0231805 -0.6901702 0.3698878 0.12646087
#> 122     as1t -1.2303328 -0.8337390 0.4799129 0.19203996
#> 123     as1t -1.0323814 -0.6947140 0.3874125 0.13656153
#> 124     as1t -1.0952481 -0.7425671 0.4050687 0.14583631
#> 125     as1t -1.0864901 -0.8006652 0.4077471 0.15404738
#> 126     as1t -0.9063947 -0.6446063 0.3515569 0.09396048
#> 127     as1t -0.9330937 -0.6693489 0.3578522 0.10615114
#> 128     as1t -0.8718164 -0.5815281 0.3247371 0.07729316
#> 129     as1t -1.0532008 -0.7315988 0.4023418 0.14468188
#> 130     as1t -1.1215224 -0.7973755 0.4312994 0.16894391
#> 131     as1t -1.0239165 -0.6963886 0.3932775 0.12979807
#> 132     as1t -1.0152930 -0.7086967 0.3819640 0.12743300
#> 133     as1t -0.9957312 -0.6723801 0.3854970 0.12301466
#> 134     as1t -0.9856176 -0.6655143 0.3544599 0.11273507
#> 135     as1t -1.0558522 -0.7473564 0.3789434 0.13504399
#> 136     as1t -1.0623647 -0.7299662 0.3914896 0.14296763
#> 137     as1t -1.0277614 -0.6744785 0.3720031 0.12897746
#> 138     as1t -1.0203903 -0.7325920 0.3794600 0.13104978
#> 139     as1t -1.0741290 -0.7518204 0.3953798 0.13814659
#> 140     as1t -0.9584139 -0.7072029 0.3446776 0.10192151
#> 141     as1t -0.9810713 -0.6906118 0.3623128 0.11504129
#> 142     as1t -1.0629195 -0.7204998 0.4098348 0.14106851
#> 143     as1t -1.1057832 -0.7904642 0.4017326 0.15350338
#> 144     as1t -1.0946099 -0.7826974 0.4318957 0.16236756
#> 145     as1t -1.0278411 -0.7304833 0.3921882 0.13248452
#> 146     as1t -1.1586397 -0.7821041 0.4358731 0.16994692
#> 147     as1t -1.0460045 -0.7315587 0.4118222 0.13724928
#> 148     as1t -1.1149438 -0.7683923 0.4526474 0.17037940
#> 149     as1t -1.1296334 -0.7839532 0.4547433 0.16893229
#> 150     as1t -1.0158417 -0.7213339 0.3859374 0.12545104
#> 151     bd6t -1.2787707 -0.6888034 0.5922870 0.20525100
#> 152     bd6t -1.3332639 -0.6489162 0.6708464 0.22460003
#> 153     bd6t -1.3450362 -0.6738379 0.6507270 0.22204939
#> 154     bd6t -1.3294378 -0.7029984 0.6092945 0.21513859
#> 155     bd6t -1.4150410 -0.6739876 0.6386051 0.22637219
#> 156     bd6t -1.3368203 -0.7052648 0.6255463 0.22118470
#> 157     bd6t -1.2362299 -0.6550899 0.5933846 0.20160141
#> 158     bd6t -1.3066813 -0.6581912 0.6435090 0.22201210
#> 159     bd6t -1.2043834 -0.6807109 0.5427426 0.18147099
#> 160     bd6t -1.2974578 -0.6565538 0.6008721 0.20814503
#> 161     bd6t -1.3943704 -0.7345107 0.6354063 0.22434849
#> 162     bd6t -1.2654977 -0.6568431 0.5487216 0.19181556
#> 163     bd6t -1.2552318 -0.6587647 0.6079179 0.21270414
#> 164     bd6t -1.3103407 -0.6532264 0.6470255 0.21777298
#> 165     bd6t -1.2907618 -0.7093575 0.5861076 0.20322729
#> 166     bd6t -1.3342311 -0.7369939 0.6408531 0.22915748
#> 167     bd6t -1.3341482 -0.7047991 0.6008951 0.21333186
#> 168     bd6t -1.3536018 -0.6690501 0.6815088 0.22399306
#> 169     bd6t -1.2716782 -0.7212715 0.5872606 0.20277462
#> 170     bd6t -1.3441046 -0.6830915 0.6156575 0.21466182
#> 171     bd6t -1.2682563 -0.6725788 0.6378940 0.21165535
#> 172     bd6t -1.2915300 -0.6091518 0.6433718 0.21083692
#> 173     bd6t -1.2775918 -0.7267691 0.6003284 0.21053920
#> 174     bd6t -1.2714974 -0.6309265 0.6089958 0.20271635
#> 175     bd6t -1.3592425 -0.6692291 0.6246146 0.21625071
#> 176     bd6t -1.2663177 -0.6336787 0.5977787 0.20041854
#> 177     bd6t -1.3039529 -0.6679095 0.6674803 0.22164678
#> 178     bd6t -1.3001698 -0.6982881 0.5675402 0.20508593
#> 179     bd6t -1.2861197 -0.6350910 0.5810211 0.20641565
#> 180     bd6t -1.2811095 -0.7434912 0.6167167 0.21137466
#> 181     bd6t -1.2476165 -0.6378831 0.5913761 0.20531316
#> 182     bd6t -1.3841622 -0.6392072 0.6218352 0.21886127
#> 183     bd6t -1.1768401 -0.6636981 0.5379758 0.18138348
#> 184     bd6t -1.2924681 -0.7051875 0.5930504 0.19873271
#> 185     bd6t -1.3439873 -0.6921052 0.6494111 0.21878487
#> 186     bd6t -1.3027483 -0.6856826 0.6221828 0.21273948
#> 187     bd6t -1.2663810 -0.7031601 0.6160648 0.21454157
#> 188     bd6t -1.3416791 -0.6456038 0.6783019 0.22457594
#> 189     bd6t -1.2358846 -0.7083081 0.5815813 0.19965355
#> 190     bd6t -1.2930115 -0.6945231 0.6317825 0.21284016
#> 191     bd6t -1.3348146 -0.6572744 0.6486147 0.21923917
#> 192     bd6t -1.2764132 -0.6755755 0.6476800 0.21621000
#> 193     bd6t -1.2015365 -0.6186125 0.5392264 0.18126747
#> 194     bd6t -1.3095492 -0.6554551 0.6362480 0.21213353
#> 195     bd6t -1.2402131 -0.6732548 0.6329675 0.21219777
#> 196     bd6t -1.2381729 -0.6384377 0.5841601 0.19890426
#> 197     bd6t -1.3555587 -0.6127782 0.6160116 0.20932832
#> 198     bd6t -1.3275125 -0.7142134 0.6666365 0.22083474
#> 199     bd6t -1.2633399 -0.6683181 0.6059331 0.20790260
#> 200     bd6t -1.3672325 -0.6439884 0.6786843 0.22844612
#> 201     bd6t -1.3164707 -0.6924762 0.5927029 0.20955567
#> 202     bd6t -1.3458019 -0.6800813 0.6696478 0.22655574
#> 203     bd6t -1.2706587 -0.6138232 0.6151737 0.20599866
#> 204     bd6t -1.3164394 -0.6865911 0.6188437 0.21138187
#> 205     bd6t -1.3783287 -0.6256321 0.7014398 0.23231628
#> 206     bd6t -1.2552378 -0.6480347 0.6261517 0.21286774
#> 207     bd6t -1.2965887 -0.7073266 0.6133917 0.21435194
#> 208     bd6t -1.2668862 -0.6427373 0.6161358 0.20290264
#> 209     bd6t -1.2770358 -0.7280258 0.5727825 0.19906466
#> 210     bd6t -1.3298245 -0.6482123 0.6293301 0.21294162
#> 211     bd6t -1.3721488 -0.6713595 0.6700776 0.22498250
#> 212     bd6t -1.2882347 -0.6522074 0.6080511 0.21297252
#> 213     bd6t -1.3264717 -0.7112028 0.5994577 0.21520520
#> 214     bd6t -1.3278327 -0.6549422 0.6278741 0.21484978
#> 215     bd6t -1.4256570 -0.7470942 0.6822729 0.23669060
#> 216     bd6t -1.3705342 -0.7733283 0.6474366 0.23203093
#> 217     bd6t -1.3149661 -0.7232919 0.6192556 0.21860357
#> 218     bd6t -1.3047090 -0.7072893 0.5808998 0.20552319
#> 219     bd6t -1.2220836 -0.6377786 0.5560832 0.18779162
#> 220     bd6t -1.2842314 -0.7584194 0.5702897 0.20725628
#> 221     bd6t -1.2768435 -0.7207549 0.5957001 0.21104321
#> 222     bd6t -1.3012536 -0.7213678 0.6075360 0.21506127
#> 223     bd6t -1.3347800 -0.6989699 0.6290362 0.22333416
#> 224     bd6t -1.3349268 -0.6944746 0.6152582 0.21328419
#> 225     bd6t -1.2609887 -0.6396416 0.5730970 0.20017986
#> 226     bd6t -1.3468399 -0.6895191 0.6563079 0.22519659
#> 227     bd6t -1.3322803 -0.7293190 0.6565419 0.21517562
#> 228     bd6t -1.2911233 -0.5896904 0.6127454 0.20668501
#> 229     bd6t -1.3388782 -0.6503070 0.6376301 0.21587453
#> 230     bd6t -1.2517291 -0.6500285 0.5854220 0.19673584
#> 231     bd6t -1.2729970 -0.6258531 0.6410923 0.21082416
#> 232     bd6t -1.3455225 -0.7082390 0.6290493 0.21755304
#> 233     bd6t -1.2816604 -0.6612639 0.6087521 0.20366997
#> 234     bd6t -1.3524589 -0.7082947 0.6363586 0.21863317
#> 235     bd6t -1.2508438 -0.6907758 0.6066504 0.20939088
#> 236     bd6t -1.2577363 -0.6300270 0.6066546 0.20154382
#> 237     bd6t -1.3142301 -0.6895431 0.6243956 0.21362719
#> 238     bd6t -1.3832116 -0.6949295 0.6535821 0.22510810
#> 239     bd6t -1.3218506 -0.7214527 0.6153911 0.21658872
#> 240     bd6t -1.3574946 -0.6693916 0.6633372 0.22432907
#> 241     bd6t -1.3533550 -0.6853090 0.6343961 0.21907944
#> 242     bd6t -1.2549574 -0.6846228 0.6067873 0.20904192
#> 243     bd6t -1.3049240 -0.6289349 0.6665260 0.21709827
#> 244     bd6t -1.2578983 -0.6960833 0.5875969 0.21065498
#> 245     bd6t -1.3274544 -0.6510884 0.6285307 0.21680441
#> 246     bd6t -1.2864949 -0.6740695 0.6279326 0.21204618
#> 247     bd6t -1.2963489 -0.6674078 0.5989687 0.20035793
#> 248     bd6t -1.2435531 -0.7109759 0.5759983 0.20081085
#> 249     bd6t -1.3101776 -0.6914108 0.6457829 0.21150012
#> 250     bd6t -1.2775903 -0.6571818 0.6147492 0.20683428
#> 251     bd6t -1.3333227 -0.7042631 0.6273213 0.21658369
#> 252     bd6t -1.2790443 -0.7193784 0.5975228 0.21002266
#> 253     bd6t -1.3088738 -0.7138781 0.6208870 0.21379220
#> 254     bd6t -1.3040137 -0.6862902 0.6357878 0.21933225
#> 255     bd6t -1.3901271 -0.7018049 0.6471093 0.22228524
#> 256     bd6t -1.3564900 -0.6874401 0.6476536 0.22311148
#> 257     bd6t -1.2990411 -0.7148971 0.6341221 0.21837468
#> 258     bd6t -1.3748818 -0.7946797 0.6391142 0.22608794
#> 259     bd6t -1.2854668 -0.6450292 0.5921127 0.20745076
#> 260     bd6t -1.3033580 -0.6588308 0.6646456 0.22271400
#> 261     bd6t -1.2283151 -0.6372127 0.5524746 0.19042395
#> 262     bd6t -1.3206940 -0.7245047 0.6395085 0.22065261
#> 263     bd6t -1.3307687 -0.7325706 0.6261568 0.21747673
#> 264     bd6t -1.3656661 -0.6408257 0.6673152 0.22175268
#> 265     bd6t -1.2615584 -0.6325541 0.6315852 0.20512565
#> 266     bd6t -1.2876420 -0.7110494 0.6553283 0.21943288
#> 267     bd6t -1.2942879 -0.6820519 0.6249094 0.21771792
#> 268     bd6t -1.3228688 -0.6754648 0.6323515 0.21668605
#> 269     bd6t -1.3739248 -0.6704170 0.6189016 0.22342298
#> 270     bd6t -1.3778874 -0.6846226 0.7016985 0.23165372
#> 271     bd6t -1.3308329 -0.6917110 0.6432118 0.21815244
#> 272     bd6t -1.3160769 -0.6634903 0.6080792 0.21241497
#> 273     bd6t -1.3160196 -0.6635224 0.6082410 0.21240846
#> 274     bd6t -1.2388846 -0.6274549 0.5806919 0.19372559
#> 275     bd6t -1.2560920 -0.6373651 0.5977147 0.20049137
#> 276     bd6t -1.2780284 -0.6717883 0.6144308 0.20859791
#> 277     bd6t -1.3479535 -0.7436869 0.6185246 0.21832627
#> 278     bd6t -1.2373215 -0.6268181 0.5791170 0.19243429
#> 279     bd6t -1.3247366 -0.7130154 0.6172905 0.21861400
#> 280     bd6t -1.3144065 -0.7414015 0.6234641 0.21799166
#> 281     bd6t -1.1573718 -0.5936337 0.5311777 0.17027059
#> 282     bd6t -1.2805618 -0.6648304 0.5960387 0.20363146
#> 283     bd6t -1.2586499 -0.6644287 0.5735356 0.19734894
#> 284     bd6t -1.3389384 -0.6721080 0.6320540 0.21975963
#> 285     bd6t -1.2516124 -0.6939657 0.5841359 0.20224819
#> 286     bd6t -1.2724848 -0.5936002 0.6220333 0.21070394
#> 287     bd6t -1.3231584 -0.7184248 0.6449293 0.21862959
#> 288     bd6t -1.3127104 -0.6698919 0.6008923 0.21342185
#> 289     bd6t -1.2781097 -0.6909704 0.6501021 0.22028696
#> 290     bd6t -1.3290657 -0.7386704 0.6253370 0.21957821
#> 291     bd6t -1.3779953 -0.6568560 0.6392121 0.22189256
#> 292     bd6t -1.3829551 -0.6979308 0.6700324 0.23129876
#> 293     bd6t -1.2784309 -0.6758400 0.6208659 0.21819472
#> 294     bd6t -1.3381282 -0.6608440 0.6805733 0.22642662
#> 295     bd6t -1.3729316 -0.7072142 0.6215278 0.22088856
#> 296     bd6t -1.3551844 -0.6953212 0.6334015 0.22119405
#> 297     bd6t -1.2677102 -0.6593327 0.5778759 0.19941169
#> 298     bd6t -1.2319532 -0.7374743 0.5934226 0.20876072
#> 299     bd6t -1.2794617 -0.6613397 0.6002639 0.20593270
#> 300     bd6t -1.3317407 -0.6278929 0.6787567 0.22625347
#> 301     bl1t -1.2434933 -0.6332466 0.6827868 0.20947061
#> 302     bl1t -1.2793669 -0.6609414 0.7183452 0.22326559
#> 303     bl1t -1.2666391 -0.7243237 0.7707108 0.22688634
#> 304     bl1t -1.3055366 -0.7452084 0.7129430 0.22099472
#> 305     bl1t -1.2976357 -0.7203877 0.7525492 0.23004351
#> 306     bl1t -1.3157475 -0.7133610 0.7656402 0.22933867
#> 307     bl1t -1.3264046 -0.6839673 0.7544832 0.22817989
#> 308     bl1t -1.2909580 -0.6859283 0.7006088 0.21891724
#> 309     bl1t -1.2576410 -0.6696175 0.7026994 0.21871924
#> 310     bl1t -1.2676144 -0.7058424 0.7215742 0.21868960
#> 311     bl1t -1.4280242 -0.6933086 0.7983082 0.23234348
#> 312     bl1t -1.2755951 -0.7282972 0.6969836 0.21557466
#> 313     bl1t -1.2267656 -0.6205297 0.6969232 0.21367776
#> 314     bl1t -1.2807580 -0.7010400 0.7179641 0.22680650
#> 315     bl1t -1.3297354 -0.7423592 0.7888751 0.24248462
#> 316     bl1t -1.4108204 -0.7109884 0.8067827 0.23857669
#> 317     bl1t -1.2935209 -0.6412166 0.7352457 0.21967041
#> 318     bl1t -1.3245835 -0.6677735 0.7443384 0.22936938
#> 319     bl1t -1.3401011 -0.7085068 0.7682270 0.23100014
#> 320     bl1t -1.3600215 -0.7123765 0.7955628 0.23069160
#> 321     bl1t -1.2778614 -0.6604978 0.7943567 0.22847507
#> 322     bl1t -1.3045006 -0.7229364 0.7290650 0.22607560
#> 323     bl1t -1.2923390 -0.6877623 0.7254666 0.21829755
#> 324     bl1t -1.3146918 -0.6590739 0.7390842 0.21989613
#> 325     bl1t -1.2619774 -0.6319224 0.6985151 0.21349259
#> 326     bl1t -1.4252015 -0.6896814 0.7974777 0.23887480
#> 327     bl1t -1.3564351 -0.6734381 0.7596367 0.23355613
#> 328     bl1t -1.3269897 -0.7095284 0.7299619 0.22993514
#> 329     bl1t -1.2996260 -0.6774278 0.6719268 0.20427727
#> 330     bl1t -1.2584841 -0.7155495 0.6852537 0.21012192
#> 331     bl1t -1.3248568 -0.7001511 0.7669749 0.23159966
#> 332     bl1t -1.3858183 -0.6967369 0.7564035 0.23211629
#> 333     bl1t -1.2806406 -0.7114492 0.6848734 0.22555406
#> 334     bl1t -1.4207007 -0.7140347 0.8038873 0.23808329
#> 335     bl1t -1.3645432 -0.6811738 0.8323710 0.23952276
#> 336     bl1t -1.2895791 -0.7026920 0.7902398 0.22965519
#> 337     bl1t -1.3012556 -0.6767149 0.7222775 0.21723211
#> 338     bl1t -1.4030613 -0.6876464 0.7881708 0.23631697
#> 339     bl1t -1.3469834 -0.7376487 0.7924327 0.23492684
#> 340     bl1t -1.2208720 -0.6325014 0.6426565 0.20010763
#> 341     bl1t -1.3634809 -0.6964974 0.8437159 0.23697738
#> 342     bl1t -1.3858167 -0.6907393 0.7703703 0.23310112
#> 343     bl1t -1.2902589 -0.7053677 0.7373904 0.22392431
#> 344     bl1t -1.3149365 -0.7305074 0.7309455 0.23161278
#> 345     bl1t -1.2686826 -0.6546874 0.7253867 0.21812038
#> 346     bl1t -1.3231345 -0.6762943 0.8145095 0.23544332
#> 347     bl1t -1.3024628 -0.6766816 0.7393345 0.23228968
#> 348     bl1t -1.2810738 -0.6996991 0.7384431 0.22665865
#> 349     bl1t -1.3665864 -0.6700537 0.7716249 0.23200124
#> 350     bl1t -1.3598180 -0.7329004 0.7128067 0.22775577
#> 351     bl1t -1.3112078 -0.7429241 0.7322572 0.22576949
#> 352     bl1t -1.3415684 -0.6897130 0.7681605 0.23172348
#> 353     bl1t -1.2535385 -0.6681875 0.6810852 0.21197924
#> 354     bl1t -1.2783953 -0.7141276 0.7060803 0.21710092
#> 355     bl1t -1.2779954 -0.7349529 0.7043858 0.21735499
#> 356     bl1t -1.3891716 -0.7188716 0.7755726 0.23496388
#> 357     bl1t -1.3332576 -0.7274506 0.7883486 0.23416926
#> 358     bl1t -1.3316812 -0.6432623 0.7270803 0.21536799
#> 359     bl1t -1.2979926 -0.6414927 0.7393212 0.22158097
#> 360     bl1t -1.3071516 -0.6879567 0.7296420 0.22013975
#> 361     bl1t -1.2478897 -0.7581518 0.7112517 0.22393265
#> 362     bl1t -1.3063540 -0.6565323 0.7500707 0.22662844
#> 363     bl1t -1.3178335 -0.7806466 0.6962184 0.22598990
#> 364     bl1t -1.3055499 -0.7299029 0.7235843 0.22253910
#> 365     bl1t -1.3549746 -0.7813589 0.7845437 0.23803147
#> 366     bl1t -1.3408680 -0.7302503 0.7732815 0.23133268
#> 367     bl1t -1.2927391 -0.7247046 0.7289629 0.22549012
#> 368     bl1t -1.3254701 -0.7506223 0.7750028 0.23765736
#> 369     bl1t -1.2306138 -0.6279130 0.6984505 0.20885162
#> 370     bl1t -1.3573990 -0.6989486 0.7836291 0.22795835
#> 371     bl1t -1.3919637 -0.7302135 0.7719645 0.23589983
#> 372     bl1t -1.3786835 -0.7460174 0.7089401 0.22827013
#> 373     bl1t -1.3438899 -0.6560739 0.7741618 0.22867210
#> 374     bl1t -1.3489424 -0.6926259 0.7818146 0.22824716
#> 375     bl1t -1.3365476 -0.7296961 0.7407262 0.23255540
#> 376     bl1t -1.3631056 -0.7117577 0.7741993 0.23570602
#> 377     bl1t -1.2904239 -0.7694505 0.7314697 0.22191338
#> 378     bl1t -1.2136172 -0.6361783 0.7445790 0.21624953
#> 379     bl1t -1.3603953 -0.7598451 0.7575868 0.23401877
#> 380     bl1t -1.2901156 -0.7474398 0.6845913 0.21268040
#> 381     bl1t -1.3240894 -0.6893170 0.7566717 0.22618171
#> 382     bl1t -1.3606224 -0.7048074 0.7872779 0.23034577
#> 383     bl1t -1.3232977 -0.7038333 0.7232087 0.22939709
#> 384     bl1t -1.3706753 -0.7442373 0.7471203 0.23110851
#> 385     bl1t -1.3184325 -0.6595051 0.7380128 0.22398572
#> 386     bl1t -1.2600298 -0.6871770 0.7209717 0.21595189
#> 387     bl1t -1.2804555 -0.6847649 0.6904280 0.21175139
#> 388     bl1t -1.3346840 -0.7642069 0.7563894 0.23228882
#> 389     bl1t -1.3639181 -0.6781765 0.7690696 0.22887366
#> 390     bl1t -1.3642626 -0.7015881 0.7903438 0.23130378
#> 391     bl1t -1.2463672 -0.6861299 0.6928322 0.21692102
#> 392     bl1t -1.2270511 -0.7019193 0.6802286 0.21013644
#> 393     bl1t -1.2179640 -0.7252477 0.7000075 0.21771662
#> 394     bl1t -1.2732189 -0.6863680 0.7229505 0.21624145
#> 395     bl1t -1.2947367 -0.6808635 0.7047988 0.21506297
#> 396     bl1t -1.2885247 -0.6831837 0.7091061 0.21961559
#> 397     bl1t -1.3260130 -0.6493912 0.7077598 0.21856479
#> 398     bl1t -1.2582505 -0.7063759 0.6885809 0.21415670
#> 399     bl1t -1.2910168 -0.7493225 0.7231603 0.22607445
#> 400     bl1t -1.2609718 -0.6538031 0.7070955 0.21777324
#> 401     bl1t -1.3397535 -0.7080449 0.7410047 0.22832167
#> 402     bl1t -1.2771670 -0.7079242 0.7425833 0.22397667
#> 403     bl1t -1.3222578 -0.7137122 0.7381023 0.22446759
#> 404     bl1t -1.3011697 -0.6700873 0.7254504 0.21328302
#> 405     bl1t -1.4136038 -0.6867183 0.7897125 0.23188917
#> 406     bl1t -1.3076416 -0.7127251 0.7176119 0.22303607
#> 407     bl1t -1.2494984 -0.6998492 0.6657140 0.20833324
#> 408     bl1t -1.4682249 -0.7234198 0.7848039 0.24300820
#> 409     bl1t -1.3224148 -0.6788078 0.7586225 0.22754175
#> 410     bl1t -1.3700971 -0.6849725 0.7658003 0.22906250
#> 411     bl1t -1.3445491 -0.6946112 0.7684304 0.23238604
#> 412     bl1t -1.2945445 -0.7241975 0.7535613 0.22781958
#> 413     bl1t -1.3822024 -0.7328086 0.7441370 0.23070530
#> 414     bl1t -1.3257877 -0.6706681 0.7766167 0.23393734
#> 415     bl1t -1.2526781 -0.6080733 0.7489955 0.22351023
#> 416     bl1t -1.3058642 -0.7042191 0.7230058 0.22375108
#> 417     bl1t -1.2997529 -0.6031637 0.7140330 0.21716587
#> 418     bl1t -1.3709648 -0.7149853 0.7744844 0.22910348
#> 419     bl1t -1.3295884 -0.7368625 0.7795997 0.23465122
#> 420     bl1t -1.3432540 -0.7307390 0.7814007 0.23204302
#> 421     bl1t -1.3258184 -0.6915881 0.6698837 0.21547077
#> 422     bl1t -1.3311609 -0.7450557 0.7945539 0.23420374
#> 423     bl1t -1.3343170 -0.6717950 0.7427024 0.22913600
#> 424     bl1t -1.2463499 -0.7152679 0.6783588 0.21238829
#> 425     bl1t -1.3623235 -0.7227178 0.7539973 0.22986888
#> 426     bl1t -1.3080173 -0.6887558 0.7038037 0.21692170
#> 427     bl1t -1.3697082 -0.7139714 0.8057576 0.23165202
#> 428     bl1t -1.3356323 -0.7026354 0.7381467 0.22684017
#> 429     bl1t -1.3105110 -0.7172023 0.7448045 0.22640042
#> 430     bl1t -1.2870598 -0.7463522 0.7446741 0.22798007
#> 431     bl1t -1.1927727 -0.6541213 0.7001136 0.20488545
#> 432     bl1t -1.2436681 -0.7293009 0.7060990 0.21183264
#> 433     bl1t -1.2896835 -0.7074892 0.7374614 0.22542358
#> 434     bl1t -1.2928345 -0.6764384 0.7040123 0.21116495
#> 435     bl1t -1.2873202 -0.6910717 0.7712573 0.22434827
#> 436     bl1t -1.2460661 -0.6373156 0.6945361 0.21325508
#> 437     bl1t -1.3756917 -0.6994980 0.7645753 0.23338468
#> 438     bl1t -1.3396645 -0.6669766 0.7347241 0.22457703
#> 439     bl1t -1.3073737 -0.7272788 0.7247706 0.22322923
#> 440     bl1t -1.3753485 -0.7210929 0.7782386 0.23755806
#> 441     bl1t -1.3450078 -0.8011649 0.7415437 0.23125406
#> 442     bl1t -1.2248644 -0.6576869 0.7029910 0.21708621
#> 443     bl1t -1.2787640 -0.7566740 0.6841263 0.21276676
#> 444     bl1t -1.3235862 -0.7302735 0.7521979 0.22761151
#> 445     bl1t -1.3280233 -0.6715349 0.7570860 0.23042091
#> 446     bl1t -1.3296795 -0.7140465 0.7351072 0.22480547
#> 447     bl1t -1.3539731 -0.7276418 0.7711582 0.23500419
#> 448     bl1t -1.2874242 -0.7478965 0.7270216 0.22440121
#> 449     bl1t -1.3472665 -0.6983385 0.7815849 0.22977328
#> 450     bl1t -1.2872784 -0.6543579 0.7177511 0.21921326