Return Data Frame of Parameters

Return Data Frame of Parameters

Usage

# S3 method for class 'emc.prior'
parameters(x, selection = "mu", N = 1000, covariates = NULL, ...)

# S3 method for class 'emc'
parameters(x, selection = "mu", N = NULL, resample = FALSE, ...)

parameters(x, ...)

Arguments

x

An emc or emc.prior object

selection

String designating parameter type (e.g. mu, sigma2, correlation, alpha)

N

Integer. How many samples to take from the posterior/prior. If NULL will return the full posterior

covariates

For priors, possible covariates in the design

...

Optional arguments that can be passed to get_pars

resample

Boolean. If TRUE will sample N samples from the posterior with replacement

Value

A data frame with one row for each sample (with a subjects column if selection = "alpha" and using draws from the posterior)

Examples

# For prior inference:
# First set up a prior
design_DDMaE <- design(data = forstmann,model=DDM,
                       formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                       constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then set up a prior using make_prior
p_vector=c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
           t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
psd <- c(v_Sleft=1,v_Sright=1,a=.3,a_Eneutral=.3,a_Eaccuracy=.3,
         t0=.4,Z=1,sv=.4,SZ=1)
# Here we left the variance prior at default
prior_DDMaE <- prior(design_DDMaE,mu_mean=p_vector,mu_sd=psd)
# Get our prior samples
parameters(prior_DDMaE, N = 100)
#>        v_Sleft    v_Sright            a  a_Eneutral a_Eaccuracy         t0
#> 1   -2.0143473  1.78891874 -0.261933308  0.67291215  1.05183887 -1.9073356
#> 2   -1.4920929  4.38879007  0.366179650 -0.38161813  0.63465539 -1.3839875
#> 3   -1.6328156  1.99761939 -0.653497420  1.02898161  0.70464278 -1.6807870
#> 4   -2.1356472  1.21686627  0.148646518  0.10976533  1.05564610 -1.1632452
#> 5   -3.0129731  2.19832175 -0.051958008  0.13019161  0.30377074 -1.9009357
#> 6    0.1045036  2.50421575  0.208772122  0.18641907  0.76272526 -1.7093602
#> 7   -2.9753893  1.93035268  0.306900232  1.00526243  1.08944611 -1.4841607
#> 8   -2.0121300  2.62866158  0.140683450  0.81067334  0.75426995 -0.7763488
#> 9   -1.1135732  1.59763392 -0.015118507  0.73629775  1.01042205 -1.8351078
#> 10  -0.7081644  1.48925534  0.106587112 -0.20269143  1.00790680 -1.4713140
#> 11  -1.3083241  1.04251640 -0.069995656 -0.01120208  0.67992973 -1.6358492
#> 12  -1.7609404  1.50390491 -0.032066995  0.26185780  0.67641547 -1.4190669
#> 13  -2.5757008  1.61125711  0.055954704  0.39296454  1.25939131 -2.2844338
#> 14  -2.8197489  3.83452578  0.799668875  0.38129145  0.94927458 -0.7505563
#> 15  -1.5765657  1.86447628 -0.437822001 -0.26268756  0.68978357 -1.2512241
#> 16  -1.5261317  3.23703979  0.319168690  0.00215092  1.10314698 -1.6419070
#> 17  -0.9811881  2.02022515 -0.487662423  0.80993170  0.50704526 -1.9898548
#> 18  -2.6537607  1.47810689  0.361387347  0.40417719  0.56568479 -1.9324431
#> 19  -2.4147664  0.17366614 -0.186058325  0.48252091  0.73549159 -1.8110409
#> 20  -1.7907836  3.36346565  0.089008657  0.75834686  0.79613702 -1.5899029
#> 21  -3.0362220  2.60699035  0.004878018  0.53227471  0.66022568 -2.3216600
#> 22  -1.7180383  3.41903170 -0.246475780  0.19374901  0.50105881 -1.1770062
#> 23  -1.3414607  1.42611077 -0.432534848  0.06513992  0.30043654 -1.2920012
#> 24  -2.5990365  2.59835922  0.162190243  0.24643496  0.31564972 -1.0655531
#> 25  -2.4904163  2.23565132  0.448761643  0.32001586  1.03628769 -1.6879314
#> 26  -3.2222596  0.82895036  0.617402180  0.49166638  0.22657421 -2.2470308
#> 27  -0.4540213  1.91858817 -0.034867440  0.35992412  1.43366607 -0.8930882
#> 28  -0.6061965  2.33676969  0.049436458  0.47119745  0.82056985 -2.2069569
#> 29  -1.8190485  0.46871925 -0.245640654 -0.31060723  1.12895332 -1.8495090
#> 30  -0.6606711  0.16859230  0.070632220  0.35030844  0.87000614 -1.7273717
#> 31  -1.7480227  2.46943001 -0.072197078  0.16494739  0.63673064 -1.4873194
#> 32  -2.0717568  0.51743571  0.172151924  0.10712461  0.55077378 -2.1905759
#> 33  -2.2463177  1.51277562  0.299982797  0.50563806  0.57599072 -1.4371626
#> 34  -2.1375424  1.92551504  0.182561192  0.45840489  0.29932999 -1.0709574
#> 35  -3.5945949  0.41530733 -0.666468354  0.48240859  0.45244897 -1.2286038
#> 36  -0.6443133  0.94553323 -0.194578969  0.30901464  0.50824370 -1.3573647
#> 37  -3.0745248  3.63901981  0.496661572  0.59588438  0.64760420 -1.6677018
#> 38  -1.3263277  0.81380201  0.226381859  0.22848618  0.86814600 -2.1170268
#> 39  -2.6002265  2.75404973  0.439440688  0.35551414  0.16930247 -1.1585471
#> 40  -1.8902167  2.00633338  0.187233922  0.15531870  0.71638111 -1.6584367
#> 41  -0.3506815  0.65712326  0.124590202  0.14602453  0.83267864 -1.2821551
#> 42  -1.7837295  1.76090501  0.139878565  0.64983773  0.86290101 -1.4694361
#> 43  -3.6142509  0.66046938  0.089001213  0.64494967  0.36536972 -1.1575108
#> 44  -1.9055272  2.21560428  0.199034658  0.33432340  0.67165223 -1.1518064
#> 45  -1.8393616  2.14529086  0.288626315  0.51007543  0.84878169 -1.3868753
#> 46  -0.7396521  1.30887511  0.437796758  0.67422886  0.45740392 -1.4195329
#> 47  -1.6506564 -0.01694964 -0.195005766  0.29520024  0.59923935 -1.3074945
#> 48  -2.0556672  1.06278336 -0.108748823  0.50400573  0.88809168 -1.7700736
#> 49  -2.1491511  1.83977198 -0.447567199  1.08947178  1.12578624 -1.4761280
#> 50  -3.1185333  3.66932285  0.267315479  0.13243125  0.83256782 -1.8296790
#> 51  -4.0062619  2.80797762 -0.376837704  0.15521746  0.97644048 -1.8526641
#> 52  -3.7454473  2.79972485  0.078312679  0.43204175  0.59326672 -1.7349650
#> 53  -1.9918031  4.10376430 -0.652563644  0.12849963  0.75754783 -1.9006405
#> 54  -3.1541262 -0.59310899  0.045360038  0.11656977  0.99791596 -1.3899005
#> 55   0.2709917  1.15635791 -0.152897545  0.56752395  0.80859915 -0.7065183
#> 56  -1.8434427  1.06267931 -0.479596196  0.72756089  0.61588207 -1.8525609
#> 57  -2.3041881  0.36345683 -0.080066329  0.32171799  0.78866559 -1.5394183
#> 58  -2.3058235  1.08298495 -0.230040535  0.41759375  0.40846155 -1.9996819
#> 59  -2.2516425 -0.28696357  0.684312927  0.26534504  0.79000427 -1.4953243
#> 60  -1.4472151  1.22579742  0.386538387  0.36148356  0.49377867 -1.6232733
#> 61  -2.5227432  3.37478968 -0.234251599  1.25234090  0.54734572 -1.8409318
#> 62  -3.6310159  1.02320402 -0.439648638  0.17552012  0.70305163 -1.7905780
#> 63  -1.6893639  3.75527168  0.234418687 -0.52188084  0.67596741 -1.3903173
#> 64  -2.1409833  1.24673945  0.048071163 -0.25048316  0.88114878 -1.9164633
#> 65  -2.9969414  3.49372860 -0.097728352  0.73430679  0.71983061 -2.0451052
#> 66  -2.8949088  1.74063196  0.180747449  0.31142614  1.14131240 -0.8202211
#> 67  -2.1718411 -0.43918954  0.118671897 -0.51212332  0.56926866 -1.6471991
#> 68  -2.1609394  2.32006369  0.238384325  0.61376148  0.12238497 -2.1590119
#> 69  -2.6278425  2.93026533 -0.017108963  0.50464102  0.94598176 -1.4601512
#> 70  -0.2126303  2.15926161  0.183616732  0.56480578  0.82209618 -1.8195393
#> 71  -1.4908624  3.33614353 -0.197042356  0.50618526  0.24153308 -1.0063018
#> 72  -3.5769896  3.55296906 -0.102185834  0.26066324  0.75603730 -1.9385794
#> 73  -3.6350317  2.23777912 -0.334165268  0.23200084  0.77475785 -1.3451795
#> 74  -2.3146025  2.86650383 -0.190400366 -0.07673109  0.50099224 -2.2199757
#> 75  -1.5046980  2.81710516  0.364539940  0.29149160  0.84815080 -1.5847241
#> 76  -1.4865265  2.15355633 -0.407680111 -0.08411235  0.62322196 -1.6416483
#> 77  -0.8616512  2.24206126 -0.078002850  0.94565597  0.57242812 -1.1695833
#> 78  -3.2821607  2.03900413  0.114598086  0.46964719  0.92001305 -1.2431130
#> 79  -2.9466407  0.78128488 -0.746701856  0.18621035  1.08189644 -1.1184695
#> 80   1.4937916  1.34760234 -0.011231984  0.53588603  0.42612420 -1.7220827
#> 81  -1.0334671  0.39512177 -0.343223574  0.38329377  0.72592625 -1.5367905
#> 82  -2.2147505  3.51896713 -0.248443959  0.62456658  0.68439846 -1.6384294
#> 83  -2.9796294  2.46408059  0.181841228  0.47714306  0.70219257 -1.9567201
#> 84  -1.3003269  2.16756717 -0.292312729  0.61820006  0.72325119 -1.0724587
#> 85  -2.9683101  2.21966363  0.146670663  0.37324087  0.57782324 -2.0217419
#> 86  -1.4688865  0.79845685  0.255705848  0.04393180  0.44159995 -1.7239125
#> 87  -1.5397461  3.22482249 -0.268176131  0.37166118  0.86968259 -1.3441057
#> 88  -2.3755934  2.19696841  0.062315200  0.82503458  1.02320585 -1.4840548
#> 89  -1.6364581  1.96913920  0.356041434  0.62811848  0.91058560 -1.0908256
#> 90  -1.2487058  0.41516166 -0.514014153  0.13074164  0.83699811 -1.4348356
#> 91  -3.2887542  0.88731049  0.059591332 -0.26547745  0.30254585 -1.6531145
#> 92  -1.4423882  1.82295005 -0.471632432  0.65341336  1.05260068 -1.6523470
#> 93  -1.5424756  2.17339148  0.250735515  0.51253276  0.34095635 -1.3242183
#> 94  -1.1921800  1.09270850  0.014595889  0.75042495  0.74830554 -1.5482880
#> 95  -1.3698247  2.24915258  0.038175671  0.25015183  0.03916751 -1.9327165
#> 96  -1.7905089  0.51881828  0.417293422  0.79514155  0.34338556 -2.3454227
#> 97  -0.8432153  1.89794267 -0.068820623  0.35442899  0.48701497 -2.0793568
#> 98  -0.9092566  1.48666838 -0.337446908 -0.06185172  0.57023473 -1.7494508
#> 99   0.3119862  1.20966602 -0.668400818  0.53162758  0.46446379 -1.6040856
#> 100 -3.8486804  1.64775874 -0.357280396  0.88359300  1.35622245 -1.7724956
#>               Z          sv           SZ
#> 1   -1.22582614 -1.05554600 -0.715218998
#> 2    0.70103065 -0.31650267 -0.461237346
#> 3    0.57457987 -1.02163052  0.583482081
#> 4    0.18970537 -0.63504871 -0.059579545
#> 5   -2.95806605 -1.43882533 -1.561036115
#> 6   -1.91909720 -0.49564339 -0.333768284
#> 7    0.16176803 -1.43682238 -0.472848623
#> 8    1.70350206 -0.07644533 -0.044023779
#> 9   -1.19600214 -0.58984444 -0.310846108
#> 10   1.68941053 -0.83234814 -0.530763727
#> 11   0.14824852 -0.91280372  0.485460656
#> 12  -0.04797968 -0.87417233  1.204219269
#> 13   0.65222992 -0.29639806 -0.604044718
#> 14  -0.85046837 -0.76162046 -0.903802994
#> 15  -2.37438334 -0.51215684  0.805777826
#> 16  -0.86351945 -1.26512164  1.400852236
#> 17  -1.54990428 -0.41732998 -1.150181367
#> 18   0.11144852 -0.59540697  1.540816688
#> 19  -1.21434652 -0.75852085 -0.488598480
#> 20  -0.43830664 -1.29392267 -2.298117806
#> 21   0.55858479 -1.27295624  0.882854600
#> 22  -0.64228731 -1.06781352  2.464855080
#> 23  -0.27339526 -0.67946576  0.679806035
#> 24  -1.16347291 -0.41608439 -0.527637571
#> 25  -1.66503829 -0.94783061  2.852394689
#> 26  -0.86636598 -1.23452926  1.289062466
#> 27   1.24033955 -0.37277889  0.336996233
#> 28   1.65891412 -0.53935469  1.653519794
#> 29  -0.39107820 -1.05435668 -1.005577880
#> 30  -2.37161374 -0.27899569 -0.726065796
#> 31   0.90709531 -0.62075780 -1.616039577
#> 32  -0.71223125 -0.74793980 -0.035116848
#> 33   0.32583398 -1.04734711 -0.321182269
#> 34  -0.19190403 -0.72346487 -1.212844373
#> 35  -0.02058647 -0.40782638  0.558042306
#> 36  -0.58101340 -0.03837019 -2.031389517
#> 37  -0.62764363 -0.55004507  0.102742884
#> 38   0.31055858 -1.28575540 -0.251662980
#> 39  -1.01517378 -1.08433589  0.210649975
#> 40  -1.49772193 -0.78848026  0.587740295
#> 41   0.25160603 -0.94541644 -1.712726105
#> 42  -2.75610623 -0.22225368  1.575143352
#> 43  -0.54690822 -0.57553137 -1.540775524
#> 44  -0.62590573 -1.07687190  0.509232139
#> 45   2.43055756 -0.70114170  1.550736959
#> 46   0.81946193 -0.23299123  0.754627933
#> 47   1.26608778 -0.48027743  0.935595976
#> 48   0.30286906 -0.26941473 -0.542916133
#> 49   1.81138305 -0.90241569 -1.201113700
#> 50  -0.06032026 -0.55167362 -1.107995877
#> 51   1.34654515 -0.82391737 -0.018650438
#> 52  -0.01183713 -0.82770055 -0.586206048
#> 53  -1.04128119 -1.35559008  0.258114420
#> 54   0.78411773 -1.13979931 -0.003997843
#> 55  -2.10425840  0.14908099  0.100061206
#> 56  -0.66379082 -0.34285734  1.189663979
#> 57   0.15485019 -0.28410666 -1.177648672
#> 58  -0.60633288 -0.61584138 -0.033978848
#> 59   0.18267264 -0.90689438 -1.217229236
#> 60  -0.47970138 -0.63985061 -0.528237503
#> 61   1.25210867 -0.74736962  1.411625711
#> 62   0.32736306 -0.84071291 -0.261182086
#> 63   2.88717869 -0.80202960  0.729857785
#> 64  -0.16657264 -0.60230338  0.216178016
#> 65  -1.01019528 -0.54460827  0.246837866
#> 66   1.19270437 -0.26786847  0.268282875
#> 67  -2.09230498 -1.48619389  0.075932903
#> 68  -1.18260537 -0.93170751 -0.396179179
#> 69   0.87143586 -0.60114575 -0.230183208
#> 70  -0.09314059 -1.17151687 -0.964258521
#> 71   0.13182360 -0.43839945 -0.203668509
#> 72  -0.48832828 -0.55110712 -0.999867294
#> 73  -1.02479856 -0.59212884 -2.277350806
#> 74  -0.64588535 -0.16112568  0.169976898
#> 75  -2.20987601 -0.02123250 -0.080757637
#> 76   0.94512171 -0.45043801 -1.090049117
#> 77   0.76875085 -0.58474055  0.253551316
#> 78  -1.13362281 -0.22114761 -0.855694616
#> 79   1.61447849 -0.96295220  1.348162529
#> 80  -0.02478524 -1.06603344 -0.470736268
#> 81  -0.65144252 -0.34610095  0.210197601
#> 82  -1.64571662 -0.88028765 -0.795846079
#> 83  -0.40977768 -1.64831432 -0.034097913
#> 84  -0.21291117 -0.78014769  1.105979069
#> 85  -0.76587738 -0.72203718 -0.861374288
#> 86   0.55681791 -1.17210617  0.365565824
#> 87  -0.28829517 -0.68211278 -0.731734598
#> 88   1.54905310 -0.70498911  0.396697414
#> 89  -0.33720572 -0.51056801  0.718778901
#> 90  -0.66251271 -0.43107645 -0.097474354
#> 91  -2.04298672 -1.32298989  0.260577396
#> 92  -0.19671813 -1.01941998  1.497260227
#> 93  -1.87480902 -0.54101888  0.466305639
#> 94   0.71137683 -1.08686189 -0.058258819
#> 95  -0.11840783 -0.36662827 -0.882426039
#> 96   0.39918848  0.04704393  1.008509798
#> 97   0.06557379 -0.64460402 -0.665613094
#> 98   0.70375989 -0.66456341  0.507603042
#> 99  -0.63693774 -1.04710758  0.565451306
#> 100 -0.75014880 -1.22450786 -0.640293345
# For posterior inference:
# Get 100 samples of the group-level mean (the default)
parameters(samples_LNR, N = 100)
#>              m       m_lMd            s         t0
#> 1   -0.9811196 -0.78261637 -0.550915549 -1.5701987
#> 2   -0.9661313 -0.94391805 -0.374097927 -1.6407631
#> 3   -0.9826687 -0.45909686 -0.453095494 -1.5494807
#> 4   -0.9091190 -0.58215610 -0.583051882 -1.6288851
#> 5   -0.2347931 -0.40856077 -0.318005546 -1.8275333
#> 6   -0.9630477 -0.48595524 -0.851498474 -1.8425821
#> 7   -0.9289952 -0.56134356 -0.564343650 -1.7183571
#> 8   -0.9624463 -0.44454834 -0.393587605 -1.5837815
#> 9   -0.9526650 -0.41553245 -0.582071394 -1.7019723
#> 10  -0.9554050 -0.71563385 -0.362949261 -1.4185379
#> 11  -0.8140972 -0.26612308 -0.768833193 -1.6967899
#> 12  -0.9700553 -0.65848436 -0.452489583 -1.6002041
#> 13  -0.9133665 -0.68743892 -1.093544394 -1.7526855
#> 14  -0.9676942 -0.50167285 -0.745485585 -1.7214540
#> 15  -0.9304879 -0.55000731 -0.422394639 -1.2028143
#> 16  -0.9978517 -0.34314664 -0.416217621 -1.6497508
#> 17  -0.8820441 -0.31286621 -1.054195718 -1.8928151
#> 18  -1.0833885 -0.56306190 -0.324133381 -1.4020840
#> 19  -0.9051631 -0.72807168 -0.607005848 -1.6111211
#> 20  -0.9304912 -0.45985783 -0.336760158 -1.5513197
#> 21  -0.7404763 -0.51271531 -0.523377481 -1.6586290
#> 22  -1.0187790 -0.50733200 -0.398569682 -1.5282437
#> 23  -0.9667462 -0.50268188  0.219079443 -1.7472997
#> 24  -0.9582441 -0.52672186 -0.497889445 -1.6796394
#> 25  -0.9562441 -0.49078944 -0.436195364 -1.5368417
#> 26  -0.9663854 -0.50560942 -0.717922815 -1.5744469
#> 27  -1.0639107 -0.67385177 -0.314030962 -1.4792762
#> 28  -1.0224155 -0.56975173 -0.603173569 -1.3855251
#> 29  -0.9982484 -0.54367301 -0.491158919 -1.7003176
#> 30  -0.9903443 -0.56946833 -0.708312367 -1.6394673
#> 31  -1.0251776 -0.68233892 -0.554351915 -1.6956811
#> 32  -1.1291780 -0.75686292 -0.248538853 -1.5232408
#> 33  -0.9796944 -0.61680796 -0.620540699 -1.6505963
#> 34  -0.8113945 -0.51235466 -0.693188889 -1.7945636
#> 35  -1.2200772 -0.97315003 -0.297626169 -1.2112230
#> 36  -0.9201198 -0.57223551 -0.259050612 -1.6016039
#> 37  -0.9142194 -0.49761480 -0.302216539 -1.5773389
#> 38  -0.9616561 -0.40710868 -0.534711732 -1.7406819
#> 39  -0.9492449 -0.55375334 -0.413224737 -1.5423728
#> 40  -0.9894384 -0.39938793 -0.682662659 -1.9928827
#> 41  -0.8806886 -0.51984748 -0.573467032 -1.4341995
#> 42  -0.9679518 -0.41888690 -0.787556953 -1.9248131
#> 43  -1.0198531 -0.41597227 -0.589020660 -1.6237633
#> 44  -1.0808396 -0.43678203 -0.469510191 -1.6236085
#> 45  -0.9283611 -0.42075473 -0.634612323 -1.7550151
#> 46  -1.1426424 -0.52704506 -0.499286173 -1.5166501
#> 47  -1.3017300 -0.66544623 -0.335280194 -1.2440264
#> 48  -0.9751348 -0.47633672 -0.386779406 -1.6351779
#> 49  -0.9342233 -0.50224789 -0.620946883 -1.6489467
#> 50  -0.9985964 -0.45330474 -0.630421878 -1.5630470
#> 51  -0.9315878 -0.51966481 -0.465783536 -1.7370025
#> 52  -0.9870514 -0.89181965 -0.111344331 -1.4699460
#> 53  -0.9169195 -0.55404444 -0.353231656 -1.5539168
#> 54  -0.8786327 -0.54303675 -0.346648089 -1.5058297
#> 55  -0.9477902 -0.61595771 -0.556930477 -1.1847693
#> 56  -0.9776619 -0.40318795 -0.531056056 -1.2977443
#> 57  -0.8434763 -0.31870875 -0.789633438 -2.0120383
#> 58  -0.9119672 -0.49924716 -0.293699621 -1.5967863
#> 59  -1.0127207 -0.50425518 -0.449933515 -1.5165119
#> 60  -0.7353512 -0.32838606 -0.819073724 -2.0949358
#> 61  -0.9699761 -0.68924491 -0.484697162 -1.5171890
#> 62  -0.9522664 -0.07751097 -0.483834810 -1.6472381
#> 63  -1.0367669 -0.60408727 -0.527622950 -1.4247506
#> 64  -1.3134263 -0.93463755 -0.007397619 -1.3883584
#> 65  -1.0032964 -0.64708915 -0.614916236 -1.7384065
#> 66  -0.9177661 -0.43309888 -0.461699589 -1.6036037
#> 67  -1.0272529 -0.72496932 -0.028530760 -1.8904442
#> 68  -0.7828386 -0.36599802 -0.887909725 -1.9098846
#> 69  -0.9893040 -0.50524756 -0.622618992 -1.7356624
#> 70  -0.9539401 -0.43292314 -0.708360568 -1.9231819
#> 71  -0.7878306 -0.39772668 -0.543149953 -1.6056142
#> 72  -0.9994864 -0.69351118 -0.405460899 -1.0117419
#> 73  -1.0497741 -0.50217850 -0.474831935 -2.0087113
#> 74  -1.0139782 -0.45067914 -0.662114619 -2.1055082
#> 75  -1.1013596 -0.54601993 -0.410663264 -1.5231039
#> 76  -0.9018078 -0.36172083 -0.813482345 -1.8808036
#> 77  -0.8985171 -0.53372756 -0.516235249 -1.6973150
#> 78  -0.9947448 -0.58090808 -0.574309776 -1.6330763
#> 79  -0.9571102 -0.28786304 -0.549653244 -1.6012989
#> 80  -0.9639456 -0.46655578 -0.537054621 -1.7091922
#> 81  -0.9073231 -0.13683258 -0.458140184 -1.2824003
#> 82  -0.8228464 -0.41598224 -0.901945703 -1.8475220
#> 83  -0.9821921 -0.51472643 -0.600772843 -1.6982912
#> 84  -0.9044736 -0.70538364 -0.811337429 -1.4543041
#> 85  -0.7528521 -0.58467753 -0.738577861 -0.6768273
#> 86  -1.0532265 -0.57457383 -0.346887444 -1.3267671
#> 87  -1.0273203 -0.43054779 -0.626083765 -1.5889587
#> 88  -0.9992775 -0.73110696 -0.520787631 -1.7935877
#> 89  -1.0463057 -0.74508312 -0.113591348 -1.2245920
#> 90  -0.9786128 -0.04573991 -0.535787685 -1.7672893
#> 91  -0.9862878 -0.55763510 -0.528719731 -1.5286193
#> 92  -0.8887304 -0.31649320 -0.638359478 -1.6257021
#> 93  -0.9084803 -0.55503472 -0.687918729 -1.8579733
#> 94  -0.9313091 -0.47280438 -0.731479665 -1.8076099
#> 95  -1.0004912 -0.48937830 -0.603968635 -1.5828861
#> 96  -0.9248655 -0.54723492 -0.301369406 -1.4633857
#> 97  -0.8763110 -0.46918520 -0.650341487 -1.4310851
#> 98  -0.8957991 -0.40858245 -0.605396604 -1.7572215
#> 99  -0.9637624 -0.58609448 -0.594637143 -1.2356965
#> 100 -0.9967384 -0.72002983 -0.302294464 -1.3738543
# or from the individual-level parameters and mapped
parameters(samples_LNR, selection = "alpha", map = TRUE)
#>     subjects   m_lMTRUE  m_lMFALSE         s         t0
#> 1      alpha -1.1570009 -0.7872438 0.4293839 0.16801658
#> 2      alpha -1.0670096 -0.7817840 0.3717159 0.14567906
#> 3      alpha -1.0295195 -0.7175014 0.3760064 0.13708996
#> 4      alpha -1.1770163 -0.7872028 0.4502336 0.17448345
#> 5      alpha -1.1393071 -0.8156149 0.4427031 0.16686474
#> 6      alpha -1.0305718 -0.7005892 0.3546520 0.12159670
#> 7      alpha -1.0596583 -0.7276130 0.4041022 0.14341162
#> 8      alpha -1.1365971 -0.6935565 0.4350651 0.15456096
#> 9      alpha -1.1358369 -0.7700018 0.4360685 0.16459899
#> 10     alpha -0.9864615 -0.6749056 0.3510998 0.11346567
#> 11     alpha -1.0364228 -0.7195639 0.3797011 0.13824446
#> 12     alpha -1.0233665 -0.6886324 0.3852578 0.12726889
#> 13     alpha -1.1124558 -0.7752946 0.4240553 0.16281209
#> 14     alpha -1.1012901 -0.7462984 0.4000909 0.14921527
#> 15     alpha -0.9069335 -0.6106710 0.3531630 0.09617918
#> 16     alpha -1.0258826 -0.7053379 0.3770549 0.13104244
#> 17     alpha -1.0791789 -0.7111617 0.4077908 0.14479556
#> 18     alpha -0.9917101 -0.6936841 0.3761668 0.12990626
#> 19     alpha -0.9866716 -0.6627641 0.3776492 0.12364275
#> 20     alpha -1.0138732 -0.7133038 0.3855606 0.13554326
#> 21     alpha -1.0104788 -0.7217930 0.3803812 0.12913389
#> 22     alpha -1.0842174 -0.7242903 0.4153931 0.15084367
#> 23     alpha -1.0589072 -0.7430621 0.3734719 0.13889447
#> 24     alpha -1.0165478 -0.7216585 0.3891099 0.12822946
#> 25     alpha -0.9176045 -0.6406681 0.3260429 0.08373377
#> 26     alpha -1.0351993 -0.7306344 0.4037468 0.14109172
#> 27     alpha -1.0112245 -0.7127998 0.3723586 0.12499177
#> 28     alpha -1.0676239 -0.7520092 0.3880921 0.13833993
#> 29     alpha -1.0776054 -0.7197328 0.4354251 0.14710947
#> 30     alpha -1.0773372 -0.7508607 0.4036126 0.14451288
#> 31     alpha -1.1358475 -0.8226264 0.4509304 0.17333695
#> 32     alpha -1.0697305 -0.7310913 0.4300572 0.15202498
#> 33     alpha -1.0969674 -0.7679438 0.4221040 0.15768440
#> 34     alpha -1.0786683 -0.7487090 0.4103467 0.15054198
#> 35     alpha -1.1214514 -0.7744996 0.4273802 0.16064738
#> 36     alpha -1.1579801 -0.7964629 0.4383740 0.17228062
#> 37     alpha -1.0641653 -0.7588610 0.4163428 0.14988124
#> 38     alpha -0.9791781 -0.7009921 0.3613880 0.12306163
#> 39     alpha -1.0893286 -0.7238464 0.4301851 0.15178784
#> 40     alpha -1.1000332 -0.7874175 0.4229469 0.16014793
#> 41     alpha -1.0853779 -0.7362570 0.4086294 0.14536702
#> 42     alpha -1.1307905 -0.7485656 0.4460703 0.16851289
#> 43     alpha -1.0286610 -0.7016076 0.3867210 0.12909833
#> 44     alpha -1.1115560 -0.7480559 0.4081617 0.15523067
#> 45     alpha -0.9043945 -0.6405006 0.3300171 0.08445788
#> 46     alpha -0.9626391 -0.6532296 0.3666730 0.10895569
#> 47     alpha -1.0324869 -0.7040148 0.3857555 0.12999493
#> 48     alpha -1.2122631 -0.8352284 0.4708365 0.18429295
#> 49     alpha -1.0798260 -0.7168437 0.4053278 0.14665948
#> 50     alpha -0.9358713 -0.6070935 0.3426830 0.09980805
#> 51     alpha -1.0873505 -0.7368567 0.4186030 0.14979388
#> 52     alpha -0.9806645 -0.6763175 0.3642130 0.11359139
#> 53     alpha -0.9637161 -0.6563292 0.3392152 0.10255932
#> 54     alpha -0.9678162 -0.6533387 0.3786926 0.11947505
#> 55     alpha -1.0141181 -0.6614038 0.3805165 0.12434530
#> 56     alpha -1.1200091 -0.7628625 0.4249748 0.15856903
#> 57     alpha -1.0545759 -0.7220144 0.3871085 0.13975675
#> 58     alpha -1.0279435 -0.7109095 0.3778541 0.13365437
#> 59     alpha -1.0651132 -0.7501589 0.4012592 0.14492181
#> 60     alpha -1.0563478 -0.7381576 0.4046139 0.14626901
#> 61     alpha -1.0399877 -0.7546098 0.3718251 0.13782127
#> 62     alpha -1.0632520 -0.7144881 0.4133738 0.14429825
#> 63     alpha -1.0881057 -0.7515780 0.4403547 0.16440378
#> 64     alpha -1.1556270 -0.7798450 0.4233238 0.16633895
#> 65     alpha -1.0001065 -0.6620950 0.3842340 0.12108479
#> 66     alpha -1.0198099 -0.6893782 0.3836754 0.13544359
#> 67     alpha -0.9439693 -0.6438970 0.3623025 0.11101413
#> 68     alpha -1.0808029 -0.7697418 0.3930781 0.15268079
#> 69     alpha -0.9473575 -0.6505703 0.3612524 0.10499307
#> 70     alpha -1.0355221 -0.7703158 0.3595023 0.13163747
#> 71     alpha -1.0192035 -0.7131389 0.3764897 0.12869942
#> 72     alpha -1.1597054 -0.7582343 0.4607373 0.17565856
#> 73     alpha -1.0493961 -0.7258235 0.3710301 0.13354864
#> 74     alpha -1.0734022 -0.7536389 0.4099012 0.14678890
#> 75     alpha -0.9055053 -0.6551253 0.3138597 0.08324387
#> 76     alpha -1.0160926 -0.6970351 0.3937478 0.12578117
#> 77     alpha -1.0431399 -0.7232982 0.3814829 0.13246155
#> 78     alpha -1.1271240 -0.8060048 0.4442186 0.16721360
#> 79     alpha -1.0670625 -0.7345962 0.4141387 0.14790988
#> 80     alpha -1.1114749 -0.7428704 0.4221870 0.15689487
#> 81     alpha -0.9443908 -0.6804092 0.3542292 0.10661996
#> 82     alpha -0.9729386 -0.6949539 0.3710433 0.11339297
#> 83     alpha -0.9456366 -0.6530277 0.3498557 0.09910587
#> 84     alpha -0.9683943 -0.6868948 0.3505727 0.10914970
#> 85     alpha -1.1186404 -0.7890304 0.4038738 0.15641637
#> 86     alpha -1.0697365 -0.7200119 0.4116401 0.14756638
#> 87     alpha -1.1054970 -0.7872433 0.4214496 0.15929478
#> 88     alpha -0.9471675 -0.6662281 0.3635858 0.09885393
#> 89     alpha -1.0787862 -0.7117497 0.3971265 0.14352790
#> 90     alpha -1.0730467 -0.7362955 0.4080333 0.14282382
#> 91     alpha -0.9914230 -0.7085223 0.3706027 0.12439652
#> 92     alpha -0.9993865 -0.6873189 0.3658255 0.12085522
#> 93     alpha -1.0646996 -0.7544896 0.3937386 0.14296763
#> 94     alpha -1.0580397 -0.7358224 0.4243844 0.14750745
#> 95     alpha -1.1130749 -0.7346204 0.4230006 0.15585541
#> 96     alpha -1.0558257 -0.7138059 0.3967373 0.13531140
#> 97     alpha -1.0695557 -0.7394513 0.4269228 0.15792581
#> 98     alpha -1.1564659 -0.7780094 0.4371126 0.16932010
#> 99     alpha -1.0237728 -0.7107497 0.3842627 0.13000473
#> 100    alpha -0.9618811 -0.6560182 0.3676605 0.11373738
#> 101    alpha -1.0235771 -0.6897257 0.3662859 0.12513523
#> 102    alpha -0.9393513 -0.6448309 0.3552599 0.10750156
#> 103    alpha -1.0443786 -0.7055384 0.4001313 0.12742383
#> 104    alpha -1.0847885 -0.7928594 0.4082327 0.15822779
#> 105    alpha -1.0292111 -0.7365530 0.3648745 0.12942239
#> 106    alpha -1.0995296 -0.7177378 0.4279220 0.15486906
#> 107    alpha -1.0667479 -0.7282988 0.3957993 0.14382537
#> 108    alpha -0.9850518 -0.6581731 0.3958601 0.12613933
#> 109    alpha -0.9367062 -0.6849032 0.3523640 0.10495299
#> 110    alpha -1.1277728 -0.7819916 0.4476395 0.16881901
#> 111    alpha -1.0571453 -0.7464033 0.3996287 0.14370013
#> 112    alpha -1.0237487 -0.7394211 0.3614820 0.12631052
#> 113    alpha -1.0025288 -0.6748217 0.3877831 0.12048503
#> 114    alpha -1.1086645 -0.8301767 0.4256795 0.16050426
#> 115    alpha -1.0403600 -0.7341616 0.3852665 0.13002827
#> 116    alpha -1.0900099 -0.7806913 0.4230767 0.16333041
#> 117    alpha -1.0471909 -0.7226619 0.4004733 0.14200933
#> 118    alpha -1.0127200 -0.6909032 0.3779048 0.12319567
#> 119    alpha -0.9335804 -0.6556374 0.3592914 0.10205537
#> 120    alpha -0.9541202 -0.6773315 0.3514517 0.10763646
#> 121    alpha -1.0693791 -0.7308815 0.4089236 0.14606731
#> 122    alpha -1.1640536 -0.8084877 0.4288452 0.17343929
#> 123    alpha -1.1375975 -0.7754573 0.4076061 0.15629489
#> 124    alpha -1.0507805 -0.6994638 0.3841095 0.13423306
#> 125    alpha -1.0846929 -0.7351153 0.4029959 0.14618218
#> 126    alpha -0.9881773 -0.7053132 0.3759156 0.12005471
#> 127    alpha -1.0521620 -0.7068017 0.4113915 0.13817453
#> 128    alpha -1.0667811 -0.7201335 0.4146586 0.14560177
#> 129    alpha -0.9819987 -0.6750864 0.3853135 0.12153688
#> 130    alpha -1.1377099 -0.7736763 0.4028053 0.15886852
#> 131    alpha -1.0467212 -0.7447093 0.3831709 0.13651159
#> 132    alpha -1.0786933 -0.7338470 0.4182813 0.15261339
#> 133    alpha -1.0312194 -0.7421994 0.3958864 0.13284598
#> 134    alpha -1.1543569 -0.8002066 0.4381499 0.17128686
#> 135    alpha -1.0832061 -0.7767270 0.4038499 0.15294995
#> 136    alpha -0.9798464 -0.6775429 0.3551291 0.11139055
#> 137    alpha -0.9587767 -0.6990903 0.3635988 0.11530345
#> 138    alpha -0.9534068 -0.6852145 0.3682995 0.11259044
#> 139    alpha -1.0243337 -0.7074813 0.3674736 0.12586707
#> 140    alpha -1.1842117 -0.8186786 0.4459840 0.18868285
#> 141    alpha -1.0433496 -0.7409972 0.3704495 0.13948958
#> 142    alpha -1.1117800 -0.8193618 0.4157302 0.16859284
#> 143    alpha -0.9432500 -0.6300303 0.3442393 0.09394759
#> 144    alpha -1.0718319 -0.7176486 0.3998227 0.13953618
#> 145    alpha -0.9898864 -0.6202097 0.3695391 0.10954006
#> 146    alpha -1.0706455 -0.7162288 0.3884072 0.14315521
#> 147    alpha -0.9637710 -0.6499030 0.3661698 0.10857226
#> 148    alpha -1.1166088 -0.7311733 0.4181847 0.15157460
#> 149    alpha -1.1899999 -0.8229935 0.4918894 0.18452999
#> 150    alpha -0.9450893 -0.6669010 0.3621561 0.11003797