Return Data Frame of Parameters

Return Data Frame of Parameters

Usage

# S3 method for class 'emc.prior'
parameters(x, selection = "mu", N = 1000, covariates = NULL, ...)

# S3 method for class 'emc'
parameters(x, selection = "mu", N = NULL, resample = FALSE, ...)

parameters(x, ...)

Arguments

x

An emc or emc.prior object

selection

String designating parameter type (e.g. mu, sigma2, correlation, alpha)

N

Integer. How many samples to take from the posterior/prior. If NULL will return the full posterior

covariates

For priors, possible covariates in the design

...

Optional arguments that can be passed to get_pars

resample

Boolean. If TRUE will sample N samples from the posterior with replacement

Value

A data frame with one row for each sample (with a subjects column if selection = "alpha" and using draws from the posterior)

Examples

# For prior inference:
# First set up a prior
design_DDMaE <- design(data = forstmann,model=DDM,
                       formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                       constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then set up a prior using make_prior
p_vector=c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
           t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
psd <- c(v_Sleft=1,v_Sright=1,a=.3,a_Eneutral=.3,a_Eaccuracy=.3,
         t0=.4,Z=1,sv=.4,SZ=1)
# Here we left the variance prior at default
prior_DDMaE <- prior(design_DDMaE,mu_mean=p_vector,mu_sd=psd)
# Get our prior samples
parameters(prior_DDMaE, N = 100)
#>        v_Sleft    v_Sright            a   a_Eneutral a_Eaccuracy         t0
#> 1   -2.6584079  3.65220397 -0.005166787  0.072719169  0.81602516 -1.6179591
#> 2   -1.8867576  1.53014787  0.155351209 -0.019660088  0.75878983 -1.6119993
#> 3   -2.4345564  1.22970439  0.019748379  0.990189006  0.25437038 -1.1574368
#> 4   -1.5335649  2.91482121 -0.123785653  0.934309774  0.47771079 -2.0603358
#> 5   -3.8895549  2.61740923 -0.153278470 -0.128279698  0.57634024 -1.2887378
#> 6   -1.9999735  1.75610610  0.010247019  0.942953167  0.87599819 -2.1911478
#> 7   -1.8832968  2.68496616 -0.216441207  0.279363318  0.28645703 -2.0017443
#> 8   -3.0634455  3.07496170 -0.405091916  0.752408083  0.64632867 -1.5540073
#> 9   -3.0020532  3.09034829 -0.552219985  0.282544967  0.77804364 -2.3809199
#> 10  -2.1138233  2.80271707 -0.112463027  0.423402296  1.03684899 -1.2818841
#> 11  -0.9361219  1.26753052  0.054237950  0.834478598  0.66813323 -1.3182217
#> 12  -2.2836677  4.77325313  0.767512840  0.044726979  0.65110481 -0.9325585
#> 13  -2.4931607  1.71049217  0.383503275  1.021081595  0.79973674 -2.0470777
#> 14  -0.9542512  3.34497801 -0.142985312  0.314047158  1.05143050 -1.0293466
#> 15  -1.8633134  1.29827948  0.345800723  0.161036353  0.48136714 -1.5663974
#> 16  -0.8956950  0.53296454  0.024757216  0.363517291  1.05852944 -0.8416818
#> 17  -2.1020075  1.89001273  0.387006854  0.124065943  0.82021129 -1.9053546
#> 18  -1.4454923  2.74727733  0.315414213  0.357196737  0.20878620 -1.1974151
#> 19  -2.4764977  1.48541779 -0.437035517  0.648110889  0.78859492 -1.0371169
#> 20  -1.0418581  3.39703782  0.289171816 -0.002630336  1.08702551 -2.1529520
#> 21  -1.2075767  1.70619977 -0.449229443  0.606340620  0.58225836 -1.6022298
#> 22  -2.6780107  2.19980023 -0.095281754  0.213602880  0.70251098 -1.7072553
#> 23  -3.0093530  0.95801593  0.075386493  0.268209623  0.40223643 -1.7515170
#> 24  -2.8742346  2.97048329 -0.500788191 -0.010813300  0.98890579 -1.8290988
#> 25  -2.1797149  2.59558098  0.646086158  0.244200952  0.37369979 -0.8480419
#> 26  -2.3482740  2.47738416 -0.145911836  0.287991352  0.98153541 -1.2381078
#> 27  -2.5027703  0.95304551  0.101150681  0.522265303  0.78458620 -1.5056405
#> 28  -0.8218102  2.94292425  0.276966464  0.153792005  0.63188706 -1.3215259
#> 29  -2.8034459  0.78715895  0.617980568  0.185210247  0.91638751 -1.0118887
#> 30  -2.2245384  4.89129229 -0.393644999  0.933017646  0.60643495 -1.6886253
#> 31  -1.8526424  3.52072107  0.211456915  0.664123602  0.67088691 -2.2077865
#> 32  -3.8888508  3.31781766  0.396713093  0.570322519  0.58060895 -0.5508796
#> 33  -2.8556602  1.78594581 -0.181311313  0.147051880  0.80500224 -2.5653809
#> 34  -1.6817650  0.90221571 -0.371589475  0.747240065  0.17480375 -1.2071331
#> 35  -0.4670192  0.74178004 -0.348629669  0.062621155  0.70050862 -1.1341584
#> 36  -2.1577967  0.55150301  0.179032053  0.485289831  0.57577313 -2.0757999
#> 37  -3.2140076  3.53742999 -0.125533950  0.699993454  0.57279274 -1.4718216
#> 38  -1.0951989  3.39120487 -0.054181063 -0.155511700  0.86131118 -1.5431120
#> 39  -3.6453438  2.93352635  0.422206612  0.087160884  0.99835854 -1.6792556
#> 40  -2.0385422  3.36671700  0.156648365  0.945432005  0.64837147 -1.0794178
#> 41  -3.0990098  3.14570924  0.066362674  0.229520053  0.55062598 -1.7153489
#> 42  -0.3966296  1.84213590  0.339120114  0.162235102  0.17255746 -2.0645760
#> 43  -2.5865586  3.71352407  0.334469620  0.350690286  0.94025638 -1.9248340
#> 44  -1.4636158  1.10661455  0.157132967  0.387642751  0.65095753 -2.0441712
#> 45  -1.9954894  0.39956759 -0.239848670  0.102280607  0.82861363 -1.6380123
#> 46  -2.7967438  2.02206451  0.304328090  0.190296326  0.39935137 -1.9726457
#> 47  -2.4885430  1.37298487 -0.196932262  0.522735805  0.75383844 -1.5486607
#> 48  -2.1020567  1.58694214  0.301675850  0.248658584  0.83930099 -1.5029382
#> 49  -1.7667185  0.66196595 -0.321552147  0.328116462  0.80300825 -1.1367820
#> 50  -2.9847732  2.40462655  0.018606341  0.248417862  0.59578333 -1.1899782
#> 51  -1.9548961  2.54764006 -0.387410679  0.406960672  0.37649340 -0.8622188
#> 52  -0.4672095 -0.01103156 -0.083046577  0.630860728  0.76550931 -1.5926346
#> 53  -1.0663647  1.16224234 -0.025205265  0.402333171  1.29396633 -1.7542413
#> 54  -2.9225057  1.20538200 -0.205803120  0.107940872  0.43046749 -1.0118145
#> 55  -2.4721754  2.22967161 -0.321600899  0.610252260  0.81709210 -1.1050638
#> 56  -1.9539686  2.06222681  0.149196952  0.393008706  0.31551880 -1.1358753
#> 57  -2.1842107  0.73457595  0.415307483  0.287384089  0.48258935 -2.2344389
#> 58  -2.0605395  2.91451182 -0.143159673  0.555267058 -0.01324243 -2.1960312
#> 59  -4.1776227  2.08555328  0.041848754  0.641315268  0.78811810 -0.9411556
#> 60  -1.4594342  1.99767636 -0.035932761  0.208676318  0.60099901 -2.0300943
#> 61  -0.2104120  2.96352130  0.034053675  0.310894658  1.14520306 -1.4652474
#> 62  -2.1275685  3.11086699 -0.062179860  0.463342052  0.64791079 -1.4547162
#> 63  -2.7731970  2.42039104  0.062368920  0.404002600  0.22157656 -2.2734122
#> 64  -2.8917303  1.89222297 -0.630643247  0.964478528  0.43956386 -1.1304358
#> 65  -0.9976832  3.03057805 -0.155979644 -0.346448012  0.40853955 -2.4714173
#> 66  -2.0495829  3.27128434  0.126460538  0.578084537  0.39127074 -1.0079091
#> 67  -2.2358876  1.14091577  0.202507342  0.568854450  0.68497612 -2.1746984
#> 68  -1.3604876  1.89972348 -0.296190174  0.147845219  0.45656989 -1.1928689
#> 69  -0.9137200  1.41017892 -0.185029861  0.434336214  0.65501632 -1.4978166
#> 70  -2.2975720  1.87491747  0.128433499  0.560915022  0.61775748 -1.3221432
#> 71  -2.9238779  2.42188947  0.197418039  0.186181216  0.50051336 -1.3383313
#> 72  -0.6036303  4.10365564  0.049509478  0.428123316  1.27205589 -2.0531130
#> 73  -2.6087465  0.98892975  0.280832673 -0.211562897  0.62930271 -2.1534559
#> 74  -3.4440660  2.13943766 -0.270868244  0.205283310  0.66546270 -1.5577168
#> 75  -0.8497911  1.83740580 -0.299600085  0.425198992  0.89333708 -0.7799189
#> 76  -2.1229975  1.69396811  0.119793050  0.393635067  0.69544827 -1.7905492
#> 77  -3.2307013  2.19236476  0.025527769  1.135516482  0.46072987 -1.3560364
#> 78  -2.9570203  1.26453773 -0.735149749  0.462310035  0.38799883 -2.3854624
#> 79  -2.6601594  2.55724258  0.138506999  0.242804482  0.55400278 -1.1917675
#> 80  -3.0089308  2.76504419  0.112059931  0.467911894  0.99632707 -1.9348311
#> 81  -2.7159605  2.36936740  0.240752741  0.336877937  1.09516737 -2.0295780
#> 82  -1.4058792  1.03320925  0.299838879  0.013954012  0.55713353 -1.9357347
#> 83  -2.2383399  1.26635498  0.158923551  0.275755154  0.62597797 -1.5956183
#> 84  -0.3064664  1.98689407 -0.077825318  0.509940159  0.71215042 -1.2853103
#> 85  -0.6512080  2.41225384  0.391987474  0.568729083  0.62942690 -0.9966019
#> 86  -2.0648452  1.80406455 -0.862614901  0.465386124  0.93885390 -1.9365981
#> 87  -0.6536069  2.19413491  0.282078014  0.465455587  0.42295457 -1.3423604
#> 88  -1.6520052  1.69636565  0.458576370  0.235497633  0.95784602 -1.4556369
#> 89  -1.4380542  3.15932218 -0.235585734  0.581893660  0.74147983 -1.4577552
#> 90  -2.9591387  1.30673934  0.121277947  0.218372224  0.30934273 -2.2653904
#> 91  -1.8173469  2.93676019 -0.189835945  0.647782355  0.92481188 -1.2696791
#> 92  -1.4146607  1.09067210  0.182371517  0.875950497  0.55486075 -0.8035008
#> 93  -1.9015772  2.81349238  0.496856169  1.018039744  0.97713340 -1.8411132
#> 94  -1.5130709  1.81369038  0.323287819  0.174914081  1.01524065 -1.8608134
#> 95  -1.8882245  2.67137782  0.019281062  0.166367250  0.78694165 -1.7143383
#> 96  -2.5782993  0.78087850  0.209203966  0.868715163  0.98125302 -1.3751932
#> 97  -1.6226691  1.78297806  0.673392883  0.221775355  0.77208475 -1.2684505
#> 98  -1.6118581  2.15239710 -0.389763428  0.250443886  0.37742008 -2.1410058
#> 99  -0.7360749  2.23790422  0.153631679  0.718014689  0.81516961 -1.6261244
#> 100 -1.0757333  1.75595825 -0.110107549  0.091823335  0.97619701 -2.3209984
#>                 Z          sv          SZ
#> 1   -0.0804993460 -1.03538032 -0.05451912
#> 2   -0.5482913119 -0.30252157  0.84291319
#> 3   -1.3826604898 -1.44536223 -1.78701061
#> 4    0.9299058413 -0.78826406  0.22282682
#> 5   -1.8535000262 -0.31918428 -1.87864131
#> 6   -0.2683895966 -1.10825777 -0.35076158
#> 7    0.4057566563 -0.06529362  0.26249064
#> 8   -1.1176769024 -0.77269165  1.27508330
#> 9   -0.1131751620 -1.22619182  0.74343626
#> 10   1.6836258638 -0.74745110 -0.82148662
#> 11   0.3699492865 -1.64327623 -0.87199436
#> 12   0.2518238794 -0.53817342 -0.48149279
#> 13   0.0557433466 -0.81591175 -0.94009092
#> 14  -0.5324284429 -0.70640905 -0.22033396
#> 15   1.4679465305 -0.95626371 -0.68662832
#> 16  -0.4600223081  0.63530152  1.34291063
#> 17   0.5843721477 -0.41400049  0.68057604
#> 18   0.5302629033 -0.61618763 -0.74734315
#> 19   0.2635561787 -0.53189309 -2.74913795
#> 20  -0.0803321113 -1.46174159  0.91142735
#> 21   0.0128097312 -1.32883701  0.73997179
#> 22  -0.8293658093 -0.67254095  0.19499112
#> 23  -2.9140854763 -1.13265495 -0.41937111
#> 24   1.3661549197  0.62486687  2.32764027
#> 25   1.4299102213 -0.93586338 -0.54273214
#> 26  -0.2614761123  0.36482519  0.34336902
#> 27  -0.5710017100 -0.50737177  0.02675517
#> 28  -1.4287898816 -1.24798550  0.18373212
#> 29  -0.7660674767 -0.58416128  0.54475339
#> 30  -0.5490693854 -0.51565592  0.41704485
#> 31  -1.9292175690  0.04290608 -0.87119566
#> 32   0.7392422717 -0.72249192 -0.22121274
#> 33  -1.5751050496 -0.47423347 -1.21204082
#> 34  -2.6317464897 -0.72755736 -0.05264594
#> 35  -0.2231439818 -0.56561133  0.80345423
#> 36   1.0930881173 -0.57281026  0.85663796
#> 37   0.7486580011 -0.85374634  0.96181572
#> 38  -1.1255719863 -1.24380112 -0.47229584
#> 39   0.4587493456 -0.20355143  2.88492848
#> 40   2.0315363666 -0.84654646 -0.23085054
#> 41   2.3910704193 -1.37776789 -1.60155030
#> 42  -0.4303562776 -1.07690498  0.87016710
#> 43  -0.6294121064 -0.85292507  1.66909934
#> 44   0.8240493994 -1.08501234 -1.28752407
#> 45   0.7595178333 -0.28286173 -0.05630831
#> 46   0.1020570671 -0.83504720 -0.71868226
#> 47   1.0327587010 -0.96770238 -1.07300354
#> 48  -0.5375215825 -0.78749526 -1.01185382
#> 49   0.2510000521 -1.30450959  0.48447429
#> 50  -1.0898658812 -1.10607986 -1.02444145
#> 51   0.4658704880 -0.71704180  0.58007778
#> 52   0.4893927515 -0.68260730 -1.66234698
#> 53  -1.6004295088 -0.95620331 -0.61342126
#> 54   0.5069126444 -0.69887697  0.60853879
#> 55  -0.3382699040 -0.51817666 -0.61290317
#> 56   0.6859934184 -0.29715771 -1.74073503
#> 57  -0.4533359843 -0.82362084  0.38699449
#> 58   0.8601328290 -1.04166241 -0.75203311
#> 59   0.2905533224 -0.61751877  1.06457992
#> 60   0.3559074550 -0.43735739 -1.59464155
#> 61   0.3098663878 -0.51155517  0.21543965
#> 62  -0.8358876828 -0.85396622  0.55343847
#> 63   2.1677794644 -0.87085468 -0.51039388
#> 64   0.1776892421 -0.59490355 -1.86703554
#> 65   1.3795462616 -0.12482091  0.23749947
#> 66   1.4082601802 -0.51384087  1.33773033
#> 67  -1.8857348597  0.12764313 -0.84498816
#> 68   0.3480893017 -0.86968233 -0.88537362
#> 69  -0.5067539160 -0.15591017  0.80116765
#> 70  -1.1522256689 -0.53339499 -1.32001970
#> 71  -0.5568429791 -0.78790183 -1.93760634
#> 72  -1.3556314528 -1.18097166 -0.86757181
#> 73  -1.7057472997 -0.19934657  1.14435022
#> 74   0.0006240421 -0.46974380  1.18727552
#> 75   0.4136559234 -0.68036733  1.53756455
#> 76  -1.4574743441 -0.46681804 -0.27604227
#> 77   0.4010619056 -0.27531429 -0.25249738
#> 78  -0.6391905427 -0.75923319  1.25150293
#> 79  -2.0953036105 -1.17093596 -1.73077408
#> 80   0.0016168513 -0.42387728 -0.75605973
#> 81   2.0803851872 -0.04910849  0.74901046
#> 82  -0.4476661032 -0.36930515 -1.34586966
#> 83   0.2469351470 -1.83938092 -1.55451024
#> 84  -2.1503035297 -0.71256000  1.36239378
#> 85   0.7583929736 -1.28531318  0.28461719
#> 86   0.3061721011 -1.14305847  0.54827690
#> 87  -0.1772581858 -1.13089471 -1.00240242
#> 88  -0.6300968317 -0.48925811 -0.41883392
#> 89   0.8721398255 -0.63036480  2.24180889
#> 90  -0.1182072299 -1.40931946  0.65388053
#> 91   0.3983806368 -0.51190741 -1.12666479
#> 92  -0.5160950276 -0.81829142 -0.21412914
#> 93   0.3045723242 -1.75899917  0.93739901
#> 94   1.6006054429 -0.89260586  0.22871894
#> 95  -0.5119700723 -0.64328114 -0.07735694
#> 96  -1.5282524249 -1.18756833  0.65843400
#> 97  -1.7743851314 -0.92398539  0.93738707
#> 98  -0.3863601607 -0.40760817 -0.62006534
#> 99  -0.2911600351 -0.12924675  0.67325885
#> 100  0.2569885545 -1.75547944 -0.85814368
# For posterior inference:
# Get 100 samples of the group-level mean (the default)
parameters(samples_LNR, N = 100)
#>              m       m_lMd            s         t0
#> 1   -0.9811196 -0.78261637 -0.550915549 -1.5701987
#> 2   -0.9661313 -0.94391805 -0.374097927 -1.6407631
#> 3   -0.9826687 -0.45909686 -0.453095494 -1.5494807
#> 4   -0.9091190 -0.58215610 -0.583051882 -1.6288851
#> 5   -0.2347931 -0.40856077 -0.318005546 -1.8275333
#> 6   -0.9630477 -0.48595524 -0.851498474 -1.8425821
#> 7   -0.9289952 -0.56134356 -0.564343650 -1.7183571
#> 8   -0.9624463 -0.44454834 -0.393587605 -1.5837815
#> 9   -0.9526650 -0.41553245 -0.582071394 -1.7019723
#> 10  -0.9554050 -0.71563385 -0.362949261 -1.4185379
#> 11  -0.8140972 -0.26612308 -0.768833193 -1.6967899
#> 12  -0.9700553 -0.65848436 -0.452489583 -1.6002041
#> 13  -0.9133665 -0.68743892 -1.093544394 -1.7526855
#> 14  -0.9676942 -0.50167285 -0.745485585 -1.7214540
#> 15  -0.9304879 -0.55000731 -0.422394639 -1.2028143
#> 16  -0.9978517 -0.34314664 -0.416217621 -1.6497508
#> 17  -0.8820441 -0.31286621 -1.054195718 -1.8928151
#> 18  -1.0833885 -0.56306190 -0.324133381 -1.4020840
#> 19  -0.9051631 -0.72807168 -0.607005848 -1.6111211
#> 20  -0.9304912 -0.45985783 -0.336760158 -1.5513197
#> 21  -0.7404763 -0.51271531 -0.523377481 -1.6586290
#> 22  -1.0187790 -0.50733200 -0.398569682 -1.5282437
#> 23  -0.9667462 -0.50268188  0.219079443 -1.7472997
#> 24  -0.9582441 -0.52672186 -0.497889445 -1.6796394
#> 25  -0.9562441 -0.49078944 -0.436195364 -1.5368417
#> 26  -0.9663854 -0.50560942 -0.717922815 -1.5744469
#> 27  -1.0639107 -0.67385177 -0.314030962 -1.4792762
#> 28  -1.0224155 -0.56975173 -0.603173569 -1.3855251
#> 29  -0.9982484 -0.54367301 -0.491158919 -1.7003176
#> 30  -0.9903443 -0.56946833 -0.708312367 -1.6394673
#> 31  -1.0251776 -0.68233892 -0.554351915 -1.6956811
#> 32  -1.1291780 -0.75686292 -0.248538853 -1.5232408
#> 33  -0.9796944 -0.61680796 -0.620540699 -1.6505963
#> 34  -0.8113945 -0.51235466 -0.693188889 -1.7945636
#> 35  -1.2200772 -0.97315003 -0.297626169 -1.2112230
#> 36  -0.9201198 -0.57223551 -0.259050612 -1.6016039
#> 37  -0.9142194 -0.49761480 -0.302216539 -1.5773389
#> 38  -0.9616561 -0.40710868 -0.534711732 -1.7406819
#> 39  -0.9492449 -0.55375334 -0.413224737 -1.5423728
#> 40  -0.9894384 -0.39938793 -0.682662659 -1.9928827
#> 41  -0.8806886 -0.51984748 -0.573467032 -1.4341995
#> 42  -0.9679518 -0.41888690 -0.787556953 -1.9248131
#> 43  -1.0198531 -0.41597227 -0.589020660 -1.6237633
#> 44  -1.0808396 -0.43678203 -0.469510191 -1.6236085
#> 45  -0.9283611 -0.42075473 -0.634612323 -1.7550151
#> 46  -1.1426424 -0.52704506 -0.499286173 -1.5166501
#> 47  -1.3017300 -0.66544623 -0.335280194 -1.2440264
#> 48  -0.9751348 -0.47633672 -0.386779406 -1.6351779
#> 49  -0.9342233 -0.50224789 -0.620946883 -1.6489467
#> 50  -0.9985964 -0.45330474 -0.630421878 -1.5630470
#> 51  -0.9315878 -0.51966481 -0.465783536 -1.7370025
#> 52  -0.9870514 -0.89181965 -0.111344331 -1.4699460
#> 53  -0.9169195 -0.55404444 -0.353231656 -1.5539168
#> 54  -0.8786327 -0.54303675 -0.346648089 -1.5058297
#> 55  -0.9477902 -0.61595771 -0.556930477 -1.1847693
#> 56  -0.9776619 -0.40318795 -0.531056056 -1.2977443
#> 57  -0.8434763 -0.31870875 -0.789633438 -2.0120383
#> 58  -0.9119672 -0.49924716 -0.293699621 -1.5967863
#> 59  -1.0127207 -0.50425518 -0.449933515 -1.5165119
#> 60  -0.7353512 -0.32838606 -0.819073724 -2.0949358
#> 61  -0.9699761 -0.68924491 -0.484697162 -1.5171890
#> 62  -0.9522664 -0.07751097 -0.483834810 -1.6472381
#> 63  -1.0367669 -0.60408727 -0.527622950 -1.4247506
#> 64  -1.3134263 -0.93463755 -0.007397619 -1.3883584
#> 65  -1.0032964 -0.64708915 -0.614916236 -1.7384065
#> 66  -0.9177661 -0.43309888 -0.461699589 -1.6036037
#> 67  -1.0272529 -0.72496932 -0.028530760 -1.8904442
#> 68  -0.7828386 -0.36599802 -0.887909725 -1.9098846
#> 69  -0.9893040 -0.50524756 -0.622618992 -1.7356624
#> 70  -0.9539401 -0.43292314 -0.708360568 -1.9231819
#> 71  -0.7878306 -0.39772668 -0.543149953 -1.6056142
#> 72  -0.9994864 -0.69351118 -0.405460899 -1.0117419
#> 73  -1.0497741 -0.50217850 -0.474831935 -2.0087113
#> 74  -1.0139782 -0.45067914 -0.662114619 -2.1055082
#> 75  -1.1013596 -0.54601993 -0.410663264 -1.5231039
#> 76  -0.9018078 -0.36172083 -0.813482345 -1.8808036
#> 77  -0.8985171 -0.53372756 -0.516235249 -1.6973150
#> 78  -0.9947448 -0.58090808 -0.574309776 -1.6330763
#> 79  -0.9571102 -0.28786304 -0.549653244 -1.6012989
#> 80  -0.9639456 -0.46655578 -0.537054621 -1.7091922
#> 81  -0.9073231 -0.13683258 -0.458140184 -1.2824003
#> 82  -0.8228464 -0.41598224 -0.901945703 -1.8475220
#> 83  -0.9821921 -0.51472643 -0.600772843 -1.6982912
#> 84  -0.9044736 -0.70538364 -0.811337429 -1.4543041
#> 85  -0.7528521 -0.58467753 -0.738577861 -0.6768273
#> 86  -1.0532265 -0.57457383 -0.346887444 -1.3267671
#> 87  -1.0273203 -0.43054779 -0.626083765 -1.5889587
#> 88  -0.9992775 -0.73110696 -0.520787631 -1.7935877
#> 89  -1.0463057 -0.74508312 -0.113591348 -1.2245920
#> 90  -0.9786128 -0.04573991 -0.535787685 -1.7672893
#> 91  -0.9862878 -0.55763510 -0.528719731 -1.5286193
#> 92  -0.8887304 -0.31649320 -0.638359478 -1.6257021
#> 93  -0.9084803 -0.55503472 -0.687918729 -1.8579733
#> 94  -0.9313091 -0.47280438 -0.731479665 -1.8076099
#> 95  -1.0004912 -0.48937830 -0.603968635 -1.5828861
#> 96  -0.9248655 -0.54723492 -0.301369406 -1.4633857
#> 97  -0.8763110 -0.46918520 -0.650341487 -1.4310851
#> 98  -0.8957991 -0.40858245 -0.605396604 -1.7572215
#> 99  -0.9637624 -0.58609448 -0.594637143 -1.2356965
#> 100 -0.9967384 -0.72002983 -0.302294464 -1.3738543
# or from the individual-level parameters and mapped
parameters(samples_LNR, selection = "alpha", map = TRUE)
#>     subjects   m_lMTRUE  m_lMFALSE         s         t0
#> 1      alpha -1.1570009 -0.7872438 0.4293839 0.16801658
#> 2      alpha -1.0670096 -0.7817840 0.3717159 0.14567906
#> 3      alpha -1.0295195 -0.7175014 0.3760064 0.13708996
#> 4      alpha -1.1770163 -0.7872028 0.4502336 0.17448345
#> 5      alpha -1.1393071 -0.8156149 0.4427031 0.16686474
#> 6      alpha -1.0305718 -0.7005892 0.3546520 0.12159670
#> 7      alpha -1.0596583 -0.7276130 0.4041022 0.14341162
#> 8      alpha -1.1365971 -0.6935565 0.4350651 0.15456096
#> 9      alpha -1.1358369 -0.7700018 0.4360685 0.16459899
#> 10     alpha -0.9864615 -0.6749056 0.3510998 0.11346567
#> 11     alpha -1.0364228 -0.7195639 0.3797011 0.13824446
#> 12     alpha -1.0233665 -0.6886324 0.3852578 0.12726889
#> 13     alpha -1.1124558 -0.7752946 0.4240553 0.16281209
#> 14     alpha -1.1012901 -0.7462984 0.4000909 0.14921527
#> 15     alpha -0.9069335 -0.6106710 0.3531630 0.09617918
#> 16     alpha -1.0258826 -0.7053379 0.3770549 0.13104244
#> 17     alpha -1.0791789 -0.7111617 0.4077908 0.14479556
#> 18     alpha -0.9917101 -0.6936841 0.3761668 0.12990626
#> 19     alpha -0.9866716 -0.6627641 0.3776492 0.12364275
#> 20     alpha -1.0138732 -0.7133038 0.3855606 0.13554326
#> 21     alpha -1.0104788 -0.7217930 0.3803812 0.12913389
#> 22     alpha -1.0842174 -0.7242903 0.4153931 0.15084367
#> 23     alpha -1.0589072 -0.7430621 0.3734719 0.13889447
#> 24     alpha -1.0165478 -0.7216585 0.3891099 0.12822946
#> 25     alpha -0.9176045 -0.6406681 0.3260429 0.08373377
#> 26     alpha -1.0351993 -0.7306344 0.4037468 0.14109172
#> 27     alpha -1.0112245 -0.7127998 0.3723586 0.12499177
#> 28     alpha -1.0676239 -0.7520092 0.3880921 0.13833993
#> 29     alpha -1.0776054 -0.7197328 0.4354251 0.14710947
#> 30     alpha -1.0773372 -0.7508607 0.4036126 0.14451288
#> 31     alpha -1.1358475 -0.8226264 0.4509304 0.17333695
#> 32     alpha -1.0697305 -0.7310913 0.4300572 0.15202498
#> 33     alpha -1.0969674 -0.7679438 0.4221040 0.15768440
#> 34     alpha -1.0786683 -0.7487090 0.4103467 0.15054198
#> 35     alpha -1.1214514 -0.7744996 0.4273802 0.16064738
#> 36     alpha -1.1579801 -0.7964629 0.4383740 0.17228062
#> 37     alpha -1.0641653 -0.7588610 0.4163428 0.14988124
#> 38     alpha -0.9791781 -0.7009921 0.3613880 0.12306163
#> 39     alpha -1.0893286 -0.7238464 0.4301851 0.15178784
#> 40     alpha -1.1000332 -0.7874175 0.4229469 0.16014793
#> 41     alpha -1.0853779 -0.7362570 0.4086294 0.14536702
#> 42     alpha -1.1307905 -0.7485656 0.4460703 0.16851289
#> 43     alpha -1.0286610 -0.7016076 0.3867210 0.12909833
#> 44     alpha -1.1115560 -0.7480559 0.4081617 0.15523067
#> 45     alpha -0.9043945 -0.6405006 0.3300171 0.08445788
#> 46     alpha -0.9626391 -0.6532296 0.3666730 0.10895569
#> 47     alpha -1.0324869 -0.7040148 0.3857555 0.12999493
#> 48     alpha -1.2122631 -0.8352284 0.4708365 0.18429295
#> 49     alpha -1.0798260 -0.7168437 0.4053278 0.14665948
#> 50     alpha -0.9358713 -0.6070935 0.3426830 0.09980805
#> 51     alpha -1.0873505 -0.7368567 0.4186030 0.14979388
#> 52     alpha -0.9806645 -0.6763175 0.3642130 0.11359139
#> 53     alpha -0.9637161 -0.6563292 0.3392152 0.10255932
#> 54     alpha -0.9678162 -0.6533387 0.3786926 0.11947505
#> 55     alpha -1.0141181 -0.6614038 0.3805165 0.12434530
#> 56     alpha -1.1200091 -0.7628625 0.4249748 0.15856903
#> 57     alpha -1.0545759 -0.7220144 0.3871085 0.13975675
#> 58     alpha -1.0279435 -0.7109095 0.3778541 0.13365437
#> 59     alpha -1.0651132 -0.7501589 0.4012592 0.14492181
#> 60     alpha -1.0563478 -0.7381576 0.4046139 0.14626901
#> 61     alpha -1.0399877 -0.7546098 0.3718251 0.13782127
#> 62     alpha -1.0632520 -0.7144881 0.4133738 0.14429825
#> 63     alpha -1.0881057 -0.7515780 0.4403547 0.16440378
#> 64     alpha -1.1556270 -0.7798450 0.4233238 0.16633895
#> 65     alpha -1.0001065 -0.6620950 0.3842340 0.12108479
#> 66     alpha -1.0198099 -0.6893782 0.3836754 0.13544359
#> 67     alpha -0.9439693 -0.6438970 0.3623025 0.11101413
#> 68     alpha -1.0808029 -0.7697418 0.3930781 0.15268079
#> 69     alpha -0.9473575 -0.6505703 0.3612524 0.10499307
#> 70     alpha -1.0355221 -0.7703158 0.3595023 0.13163747
#> 71     alpha -1.0192035 -0.7131389 0.3764897 0.12869942
#> 72     alpha -1.1597054 -0.7582343 0.4607373 0.17565856
#> 73     alpha -1.0493961 -0.7258235 0.3710301 0.13354864
#> 74     alpha -1.0734022 -0.7536389 0.4099012 0.14678890
#> 75     alpha -0.9055053 -0.6551253 0.3138597 0.08324387
#> 76     alpha -1.0160926 -0.6970351 0.3937478 0.12578117
#> 77     alpha -1.0431399 -0.7232982 0.3814829 0.13246155
#> 78     alpha -1.1271240 -0.8060048 0.4442186 0.16721360
#> 79     alpha -1.0670625 -0.7345962 0.4141387 0.14790988
#> 80     alpha -1.1114749 -0.7428704 0.4221870 0.15689487
#> 81     alpha -0.9443908 -0.6804092 0.3542292 0.10661996
#> 82     alpha -0.9729386 -0.6949539 0.3710433 0.11339297
#> 83     alpha -0.9456366 -0.6530277 0.3498557 0.09910587
#> 84     alpha -0.9683943 -0.6868948 0.3505727 0.10914970
#> 85     alpha -1.1186404 -0.7890304 0.4038738 0.15641637
#> 86     alpha -1.0697365 -0.7200119 0.4116401 0.14756638
#> 87     alpha -1.1054970 -0.7872433 0.4214496 0.15929478
#> 88     alpha -0.9471675 -0.6662281 0.3635858 0.09885393
#> 89     alpha -1.0787862 -0.7117497 0.3971265 0.14352790
#> 90     alpha -1.0730467 -0.7362955 0.4080333 0.14282382
#> 91     alpha -0.9914230 -0.7085223 0.3706027 0.12439652
#> 92     alpha -0.9993865 -0.6873189 0.3658255 0.12085522
#> 93     alpha -1.0646996 -0.7544896 0.3937386 0.14296763
#> 94     alpha -1.0580397 -0.7358224 0.4243844 0.14750745
#> 95     alpha -1.1130749 -0.7346204 0.4230006 0.15585541
#> 96     alpha -1.0558257 -0.7138059 0.3967373 0.13531140
#> 97     alpha -1.0695557 -0.7394513 0.4269228 0.15792581
#> 98     alpha -1.1564659 -0.7780094 0.4371126 0.16932010
#> 99     alpha -1.0237728 -0.7107497 0.3842627 0.13000473
#> 100    alpha -0.9618811 -0.6560182 0.3676605 0.11373738
#> 101    alpha -1.0235771 -0.6897257 0.3662859 0.12513523
#> 102    alpha -0.9393513 -0.6448309 0.3552599 0.10750156
#> 103    alpha -1.0443786 -0.7055384 0.4001313 0.12742383
#> 104    alpha -1.0847885 -0.7928594 0.4082327 0.15822779
#> 105    alpha -1.0292111 -0.7365530 0.3648745 0.12942239
#> 106    alpha -1.0995296 -0.7177378 0.4279220 0.15486906
#> 107    alpha -1.0667479 -0.7282988 0.3957993 0.14382537
#> 108    alpha -0.9850518 -0.6581731 0.3958601 0.12613933
#> 109    alpha -0.9367062 -0.6849032 0.3523640 0.10495299
#> 110    alpha -1.1277728 -0.7819916 0.4476395 0.16881901
#> 111    alpha -1.0571453 -0.7464033 0.3996287 0.14370013
#> 112    alpha -1.0237487 -0.7394211 0.3614820 0.12631052
#> 113    alpha -1.0025288 -0.6748217 0.3877831 0.12048503
#> 114    alpha -1.1086645 -0.8301767 0.4256795 0.16050426
#> 115    alpha -1.0403600 -0.7341616 0.3852665 0.13002827
#> 116    alpha -1.0900099 -0.7806913 0.4230767 0.16333041
#> 117    alpha -1.0471909 -0.7226619 0.4004733 0.14200933
#> 118    alpha -1.0127200 -0.6909032 0.3779048 0.12319567
#> 119    alpha -0.9335804 -0.6556374 0.3592914 0.10205537
#> 120    alpha -0.9541202 -0.6773315 0.3514517 0.10763646
#> 121    alpha -1.0693791 -0.7308815 0.4089236 0.14606731
#> 122    alpha -1.1640536 -0.8084877 0.4288452 0.17343929
#> 123    alpha -1.1375975 -0.7754573 0.4076061 0.15629489
#> 124    alpha -1.0507805 -0.6994638 0.3841095 0.13423306
#> 125    alpha -1.0846929 -0.7351153 0.4029959 0.14618218
#> 126    alpha -0.9881773 -0.7053132 0.3759156 0.12005471
#> 127    alpha -1.0521620 -0.7068017 0.4113915 0.13817453
#> 128    alpha -1.0667811 -0.7201335 0.4146586 0.14560177
#> 129    alpha -0.9819987 -0.6750864 0.3853135 0.12153688
#> 130    alpha -1.1377099 -0.7736763 0.4028053 0.15886852
#> 131    alpha -1.0467212 -0.7447093 0.3831709 0.13651159
#> 132    alpha -1.0786933 -0.7338470 0.4182813 0.15261339
#> 133    alpha -1.0312194 -0.7421994 0.3958864 0.13284598
#> 134    alpha -1.1543569 -0.8002066 0.4381499 0.17128686
#> 135    alpha -1.0832061 -0.7767270 0.4038499 0.15294995
#> 136    alpha -0.9798464 -0.6775429 0.3551291 0.11139055
#> 137    alpha -0.9587767 -0.6990903 0.3635988 0.11530345
#> 138    alpha -0.9534068 -0.6852145 0.3682995 0.11259044
#> 139    alpha -1.0243337 -0.7074813 0.3674736 0.12586707
#> 140    alpha -1.1842117 -0.8186786 0.4459840 0.18868285
#> 141    alpha -1.0433496 -0.7409972 0.3704495 0.13948958
#> 142    alpha -1.1117800 -0.8193618 0.4157302 0.16859284
#> 143    alpha -0.9432500 -0.6300303 0.3442393 0.09394759
#> 144    alpha -1.0718319 -0.7176486 0.3998227 0.13953618
#> 145    alpha -0.9898864 -0.6202097 0.3695391 0.10954006
#> 146    alpha -1.0706455 -0.7162288 0.3884072 0.14315521
#> 147    alpha -0.9637710 -0.6499030 0.3661698 0.10857226
#> 148    alpha -1.1166088 -0.7311733 0.4181847 0.15157460
#> 149    alpha -1.1899999 -0.8229935 0.4918894 0.18452999
#> 150    alpha -0.9450893 -0.6669010 0.3621561 0.11003797