Return Data Frame of Parameters

Return Data Frame of Parameters

Usage

# S3 method for class 'emc.prior'
parameters(x, selection = "mu", N = 1000, covariates = NULL, ...)

# S3 method for class 'emc'
parameters(x, selection = "mu", N = NULL, resample = FALSE, ...)

parameters(x, ...)

Arguments

x

An emc or emc.prior object

selection

String designating parameter type (e.g. mu, sigma2, correlation, alpha)

N

Integer. How many samples to take from the posterior/prior. If NULL will return the full posterior

covariates

For priors, possible covariates in the design

...

Optional arguments that can be passed to get_pars

resample

Boolean. If TRUE will sample N samples from the posterior with replacement

Value

A data frame with one row for each sample (with a subjects column if selection = "alpha" and using draws from the posterior)

Examples

# For prior inference:
# First set up a prior
design_DDMaE <- design(data = forstmann,model=DDM,
                       formula =list(v~0+S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
                       constants=c(s=log(1)))
#> Parameter(s) st0 not specified in formula and assumed constant.
#> 
#>  Sampled Parameters: 
#> [1] "v_Sleft"     "v_Sright"    "a"           "a_Eneutral"  "a_Eaccuracy"
#> [6] "t0"          "Z"           "sv"          "SZ"         
#> 
#>  Design Matrices: 
#> $v
#>      S v_Sleft v_Sright
#>   left       1        0
#>  right       0        1
#> 
#> $a
#>         E a a_Eneutral a_Eaccuracy
#>     speed 1          0           0
#>   neutral 1          1           0
#>  accuracy 1          0           1
#> 
#> $t0
#>  t0
#>   1
#> 
#> $s
#>  s
#>  1
#> 
#> $Z
#>  Z
#>  1
#> 
#> $sv
#>  sv
#>   1
#> 
#> $SZ
#>  SZ
#>   1
#> 
#> $st0
#>  st0
#>    1
#> 
# Then set up a prior using make_prior
p_vector=c(v_Sleft=-2,v_Sright=2,a=log(1),a_Eneutral=log(1.5),a_Eaccuracy=log(2),
           t0=log(.2),Z=qnorm(.5),sv=log(.5),SZ=qnorm(.5))
psd <- c(v_Sleft=1,v_Sright=1,a=.3,a_Eneutral=.3,a_Eaccuracy=.3,
         t0=.4,Z=1,sv=.4,SZ=1)
# Here we left the variance prior at default
prior_DDMaE <- prior(design_DDMaE,mu_mean=p_vector,mu_sd=psd)
# Get our prior samples
parameters(prior_DDMaE, N = 100)
#>          v_Sleft    v_Sright             a    a_Eneutral a_Eaccuracy         t0
#> 1   -2.625039268  0.31330669  0.2513361133  0.4514770435  0.35170610 -1.1079119
#> 2   -1.121866512  2.82158108  0.2065920762  0.5716404042  0.67457367 -1.7318230
#> 3   -3.265396352  4.16895597  0.3623885995  0.0685325331  0.57228173 -1.7961001
#> 4   -2.028546755  1.95712954  0.4105806852  0.3377338124  1.14808836 -2.2289390
#> 5   -1.620360517  1.49767655 -0.0999622151  0.0998924932  0.37160981 -1.4880265
#> 6    0.050084686  1.50896883 -0.6927506627  0.7071866654  0.48038695 -1.8846414
#> 7   -1.818696520  1.86110864  0.0017292558  0.5210492284  0.58194917 -1.3516873
#> 8   -1.564818509  1.67406841  0.3446422855  0.7035162649  0.85766627 -1.5139452
#> 9    0.187332993  3.53261063 -0.0707101077  0.0975388380  0.48002521 -1.5066844
#> 10  -2.045027725  1.21509553 -0.5003825810  0.2913971520  0.96884616 -1.8395767
#> 11  -1.480592796  2.30115336  0.0317028582  0.2132533056  0.43823588 -2.0190894
#> 12  -2.256092192  3.84386201 -0.1955849705  0.4760810798  0.71653544 -1.9941806
#> 13  -1.958767078  1.57750317 -0.6159741665  0.7448662721  0.25495516 -1.3134589
#> 14  -2.262197489  0.42785584 -0.4544002961 -0.0749957440  0.53387522 -2.1941401
#> 15  -1.212261153  2.76904224  0.0996607737  0.1029521256  0.65731140 -1.7215960
#> 16  -2.374580858  3.05271147 -0.3147531020  0.0274185347  1.66545916 -1.7761809
#> 17  -1.483137956  2.36896453 -0.0646141523  0.4250530182  0.68292700 -0.7580572
#> 18  -1.689519251  2.43652348 -0.1375095998  0.0864672679  1.07210273 -1.7492981
#> 19  -0.890079710  2.08473729  0.2262161356  0.2556775030  0.75748077 -1.7393123
#> 20  -0.002786615  2.60070882 -0.3753814085  0.2221153331  0.33750316 -0.7299138
#> 21  -2.414339948  1.52375311 -0.2365808514  0.2270799279  1.18841942 -1.6310492
#> 22  -2.516063831  1.00749285  0.5027090797  0.2731161430  0.47622739 -2.1039472
#> 23  -0.890151861  2.70758835 -0.1090971891  0.4233900893  0.48176824 -1.8963252
#> 24  -2.090319594  2.21453883 -0.2215583114  0.2331485012  0.29804234 -1.6826081
#> 25  -2.788621971  1.49780128  0.4488182010  0.0642740219  0.63943170 -0.8484932
#> 26  -1.514540021  1.62439713 -0.1685629091  0.3022899379  0.72029617 -0.9700344
#> 27  -2.113639896  0.46709800 -0.1563351953  0.2585039722  0.70729351 -1.0893584
#> 28  -3.756527396  1.61122014  0.0267621669  0.6589690093  0.98190557 -1.3357141
#> 29  -1.825197300  2.07455118  0.1284500295  0.4128676030  0.19290465 -1.3148395
#> 30  -1.865961355  2.22101947  0.4922538498  0.3397499944  0.74356680 -1.1420844
#> 31   0.002482730  2.06670087  0.5600555534  0.0001943023  0.69944226 -1.1094721
#> 32  -3.052513279  1.56284047  0.0993537519 -0.1987980413  0.75674131 -1.1147679
#> 33  -3.726730399  1.39849329 -0.1056139370  0.6165222789  0.66144578 -2.1128974
#> 34  -0.781891390  0.66122571  0.1982460893  0.2485913952  0.89827084 -1.6337667
#> 35  -0.982441363  0.81156596 -0.2164813321  0.8612304215  0.80636357 -2.4303270
#> 36  -2.101883256  2.62418747  0.2877016133  0.9067815568  0.70995220 -1.6302307
#> 37  -2.974009583  1.82009377  0.3044829518 -0.1923594385  0.56496339 -1.5627830
#> 38  -2.033036159 -0.46589819  0.7714374438  0.3438753309  0.88850517 -1.4999313
#> 39  -1.621832228  1.05459117  0.2570769033  0.2671536064  1.41817919 -2.2698575
#> 40  -2.589481039  1.00321926  0.0433427114  0.4011728842  0.15606281 -1.5956175
#> 41  -1.523866722  3.37857014  0.1368709210  0.0647885670  0.56245354 -1.4709965
#> 42  -2.829477612  1.42643973  0.4511701827  0.1732216292  0.94686664 -2.1137111
#> 43  -2.634748265  1.97115845  0.2012087906 -0.0896988549  0.58822091 -1.3068753
#> 44  -1.732164985  2.65325768 -0.0368125983  0.2813621540 -0.09979751 -1.6466143
#> 45  -0.220497090  2.28642442  0.0378947575  0.7871451419  0.47760731 -1.7895734
#> 46  -3.438506645  1.80948320  0.1135271711  0.4954766717  0.39145630 -1.6017342
#> 47  -4.224987696  3.23569346 -0.3723133490  0.5418958888  0.89111797 -1.6893938
#> 48  -1.116697180 -0.05233698 -0.4909137804  0.8345858105  1.00713583 -1.4353223
#> 49  -0.889722903  1.51501240  0.0691850492  0.3169177678  0.95473667 -1.7488269
#> 50  -0.789989490  2.74090001  0.5172786717  0.4250112879  1.03064800 -0.8192703
#> 51  -2.214041240  2.15168050  0.5136914932  0.3076219404  0.80504858 -1.7005115
#> 52  -1.878681623  2.71284232  0.2336580089  0.6798970894  0.52082881 -0.9586854
#> 53  -0.705916094  0.91000813 -0.2619213001 -0.0019586097  0.74770134 -1.5435016
#> 54  -2.993698591  3.02678506  0.2253183908 -0.0472848531  0.66460295 -1.9678170
#> 55  -2.097369268  2.21615254  0.2647395493  0.4671443595  0.50821643 -1.9033576
#> 56  -2.184308869  2.96726726 -0.0324840274  0.1959389078  0.61036363 -1.1635785
#> 57  -1.589724904  1.44154309  0.1816112007  0.2535650455  0.26697753 -1.5582407
#> 58  -1.641144277  1.39144282 -0.0606722565  0.3234906761  0.55253725 -1.3277710
#> 59  -3.198622359  2.63949200  0.7290679957  0.2383004635  0.94661845 -1.9223187
#> 60  -3.458970729  1.94870211 -0.1580775528  0.3462856474  0.50427356 -1.9429753
#> 61  -3.186206585  2.10107915  0.1598967860  0.5814857098  0.60262318 -1.5776371
#> 62  -1.679597686  1.55521802  0.4110011982  0.6074412671  0.71479721 -2.2125408
#> 63  -3.181559227  2.49865804 -0.3116869321  0.3375985137  0.80757493 -1.9228442
#> 64  -1.535032009  2.84053983 -0.0857536271  0.5567029845  0.34637222 -1.6602974
#> 65  -0.925987740  1.97265303 -0.0099991027 -0.0493551788  0.93026278 -1.6937316
#> 66  -2.849061135  1.60296948 -0.3652799963  0.9117419530  0.68834642 -1.1794599
#> 67  -3.222926179  3.54660915 -0.4245845759  0.5009821862  0.94707807 -1.5381618
#> 68  -3.063497906  0.61195095  0.6260152303  0.2019141645  0.13647568 -1.3961342
#> 69  -2.116302517  3.13939629  0.1908372106  0.2575838809  0.44289071 -1.5010112
#> 70  -1.100645947  1.16918847 -0.0991634100  0.6277094652  0.99013866 -2.3848398
#> 71  -1.519374400  1.17182573  0.3060759036  0.5670097177  0.92386287 -1.5611502
#> 72  -2.028395055 -0.24905109  0.0094578014  0.4671334706  0.64654358 -1.3821225
#> 73  -1.602157791  1.44977626  0.0273802150 -0.1830471716  0.35717726 -2.1405401
#> 74  -1.017886997  1.27261647 -0.2990516947  0.0929584504  0.56877056 -1.7050495
#> 75  -2.091434276  3.18718681  0.3574803806  0.1687761427  0.22881422 -0.6262137
#> 76  -3.614039457  1.27178089 -0.4621327216  0.1975367239  0.72880201 -2.1553217
#> 77  -1.773675058  2.74808116  0.3183285760  0.3416106245  0.66505614 -1.6441236
#> 78  -2.287340800  2.37324143  0.1209870994  0.0929631198  0.17465583 -1.3527059
#> 79  -1.436132610  2.18942624 -0.2198561419  0.7013748661  1.21473731 -1.2569664
#> 80  -1.475085721  2.62203324 -0.0290058220  0.3828861488  0.99889430 -1.3247971
#> 81  -1.792618843 -0.21063311  0.8075142010  0.2606620616  1.40556760 -1.4595805
#> 82  -1.786042020  1.81387930 -0.0361181474  0.7093154089  0.63270974 -2.4245109
#> 83  -1.383432183  0.30789848  0.1106226174  0.6958228711  1.07612079 -1.6994224
#> 84  -2.436829977  2.45746208 -0.4853321296  0.4893534668  1.25650639 -1.6110622
#> 85  -1.186599626  2.90443546  0.0008074984  0.0524574608  0.29768096 -1.8466369
#> 86  -2.653779825  2.39439485 -0.2740698144  0.6714898192  0.79315817 -1.6776938
#> 87  -1.768885066  2.64751336  0.1068850035  0.2080620457  0.94970784 -1.1482634
#> 88   0.161415846  2.27631553 -0.0474882086 -0.3469102326  0.22356265 -1.6405072
#> 89  -2.194152409  3.21458879 -0.2764548129  0.0429322920  0.32445133 -1.3125191
#> 90  -2.591892102  1.63164742 -0.5557850473  0.0545805300  0.26053679 -1.1877090
#> 91  -2.191774809  2.28387891 -0.5253202558  0.1598641749  0.71001164 -1.4898032
#> 92  -1.935756443  2.64979187 -0.0078055904  0.2123948920  1.00673888 -0.9632198
#> 93  -0.983544782  0.84383261  0.6962580672  0.2244057338  0.25549236 -1.7498050
#> 94  -1.857541568  0.61051648 -0.2598113234  0.3564796284  1.45905501 -2.3535289
#> 95  -3.139200640  2.14362323 -0.3298652825  0.6765200360  1.13828103 -0.8291495
#> 96  -2.131874912  2.47703724 -0.2915982634  0.3499044573  1.05943629 -1.3929243
#> 97  -2.764606014  2.39529587 -0.2971522891  0.5740775251  0.35822226 -0.8780257
#> 98  -2.352453199  2.37114796  0.0730598156  0.1012308563  0.45575302 -1.4896004
#> 99  -1.174077098  1.95143165  0.0903940955  0.4835735553  1.46578211 -2.0835534
#> 100 -0.903391528  3.44566224 -0.5775435756  0.5292959573  1.17115817 -1.7750443
#>               Z          sv          SZ
#> 1    0.42646422 -0.81117577  0.89512566
#> 2   -0.38047100 -0.97102997 -0.20791728
#> 3    0.77996512 -0.72649481  0.25331851
#> 4    0.58461375 -0.64360548  0.21594157
#> 5    0.44820978 -0.67194549  0.92226747
#> 6    1.02557137 -0.80705638 -1.22071771
#> 7   -0.22048656 -0.56043439  1.09683901
#> 8   -0.62790608 -0.14888620 -0.60025959
#> 9   -0.24669188 -0.83216422 -0.95161857
#> 10   0.60796432 -1.34030026 -0.05556197
#> 11   0.11764660 -1.07213703 -0.49055744
#> 12  -0.07130809 -0.11532684  0.45150405
#> 13   1.90910357 -1.27070444  0.70178434
#> 14   0.68791677  0.14689640 -1.28703048
#> 15   0.56298953 -0.84212268  0.97697339
#> 16   0.29822759 -0.43851931 -0.48378063
#> 17  -0.74133610 -1.13154569  0.03778840
#> 18  -0.86551286 -0.78765901 -0.19717589
#> 19   0.09458353 -1.05129252 -1.31080153
#> 20   1.31241298 -0.79920520  0.54319406
#> 21   0.11924524 -0.59567221  1.23247588
#> 22  -1.28471572 -0.92273657  0.61798582
#> 23   0.88465050 -1.09938421  1.95529397
#> 24   0.41898240 -0.56342544 -0.78153649
#> 25  -0.10097489 -1.23708346 -0.66476944
#> 26  -0.08856511 -0.26082738  0.63075412
#> 27   2.29307897 -0.07411476 -0.13315096
#> 28  -1.39527435 -0.35328996 -0.44655722
#> 29   0.38602657 -0.79940783  0.11814451
#> 30   1.05418102 -0.23504194 -0.57746800
#> 31  -0.71524219 -0.99422277 -0.93853870
#> 32   2.03757402 -0.17267678  0.75677476
#> 33   1.68443571 -0.32859066  0.23743027
#> 34   0.63296071 -0.15894013  0.00729009
#> 35  -1.36403745 -0.77345959  0.86577940
#> 36  -1.75323736 -0.65341614 -0.57185006
#> 37  -0.89320757 -0.55958600  0.41142992
#> 38   1.02467323 -0.36608340 -0.20979317
#> 39  -0.46398724 -0.36299524  0.51013255
#> 40   0.19023032 -0.62325662 -1.05501704
#> 41  -0.64704563 -1.55620571  0.88425082
#> 42  -0.35454240 -0.72256959 -1.16865142
#> 43  -0.53880916 -0.60223041  0.49222857
#> 44   0.43028470 -0.47898764 -0.55527835
#> 45   2.39745248 -0.68869551  1.63356842
#> 46  -1.07742065 -0.40806585  1.08477509
#> 47  -0.64511396 -0.62701877  0.43881870
#> 48   0.71517841 -0.32627721 -2.66092280
#> 49   0.51850377 -0.84942117 -1.09278721
#> 50  -0.28148212 -1.22232763 -0.23935157
#> 51   0.02045071 -0.56752412  1.32821470
#> 52  -0.38095674 -0.73546085  1.40405027
#> 53   0.36411469 -0.47228409 -0.60189285
#> 54  -2.07075107 -0.63309913 -0.07921171
#> 55  -0.13180279 -0.56914039 -1.03968035
#> 56   0.55004396 -0.19847686  0.13909786
#> 57   1.94585122 -0.37278144  1.16525339
#> 58  -1.19736350 -0.34660073  0.86415249
#> 59   1.11071142 -0.59321729  1.65191539
#> 60   0.57872237 -1.12817947  1.48403093
#> 61   0.96126415 -1.27573355 -0.78173971
#> 62   0.02610023 -0.81971353 -0.10234651
#> 63   0.58299141 -1.21975134 -2.80977468
#> 64  -1.94151838 -0.22067482  1.85991086
#> 65  -0.65674293 -1.25795750 -0.29976250
#> 66  -2.60169967 -0.87442631 -0.67548229
#> 67  -0.87525548 -0.31668086  0.17058808
#> 68   0.31023026 -1.23468091 -1.94295641
#> 69   0.15735335 -0.44126248 -0.39579795
#> 70   0.10719041 -0.44963558 -1.45082431
#> 71   0.86364843 -0.14094137  1.96624802
#> 72   1.01067796 -0.90034015 -0.29409533
#> 73  -0.85362370 -0.97046899  0.38230514
#> 74   0.48361753 -0.82167712 -2.07848927
#> 75  -0.16242194 -0.73212768  0.42057419
#> 76   0.58998268 -0.57740957 -0.90421503
#> 77   1.44146176 -0.24311842  0.83440157
#> 78  -1.52931053 -0.69247371  0.25024782
#> 79  -1.94365090 -0.13331671 -0.05605595
#> 80   0.99026225  0.26002350  0.66441586
#> 81   1.53843020 -0.73703131  0.51147075
#> 82  -0.19588925 -0.47723094  0.61645572
#> 83  -0.32189259 -0.09801205 -1.66792805
#> 84  -0.27845402 -0.50318249 -0.27907217
#> 85   0.79738050 -1.47642925 -1.88632516
#> 86   0.81882814 -0.53780112 -0.44593503
#> 87   0.27627456 -0.63550532 -0.07562508
#> 88   0.20629404 -0.58239820  0.82150678
#> 89  -0.08291994 -0.37722001 -0.26770642
#> 90  -0.59733009 -0.37736324  1.51649060
#> 91  -0.75939812  0.38079642 -0.45839014
#> 92  -0.02969397 -0.46824024 -0.09741250
#> 93   0.14670848 -0.04369870  0.91120968
#> 94   1.13105465 -0.90404088  1.66599090
#> 95   0.79760066  0.04415932  1.24642391
#> 96   0.45735733 -1.10839960 -0.60451323
#> 97   0.46059135 -0.97354863  0.24104593
#> 98   1.63905191 -0.25930038 -0.62456747
#> 99   0.10091986 -1.40513810  0.58983592
#> 100 -0.21215053 -0.70776207  0.36501875
# For posterior inference:
# Get 100 samples of the group-level mean (the default)
parameters(samples_LNR, N = 100)
#>              m       m_lMd            s         t0
#> 1   -0.9811196 -0.78261637 -0.550915549 -1.5701987
#> 2   -0.9661313 -0.94391805 -0.374097927 -1.6407631
#> 3   -0.9826687 -0.45909686 -0.453095494 -1.5494807
#> 4   -0.9091190 -0.58215610 -0.583051882 -1.6288851
#> 5   -0.2347931 -0.40856077 -0.318005546 -1.8275333
#> 6   -0.9630477 -0.48595524 -0.851498474 -1.8425821
#> 7   -0.9289952 -0.56134356 -0.564343650 -1.7183571
#> 8   -0.9624463 -0.44454834 -0.393587605 -1.5837815
#> 9   -0.9526650 -0.41553245 -0.582071394 -1.7019723
#> 10  -0.9554050 -0.71563385 -0.362949261 -1.4185379
#> 11  -0.8140972 -0.26612308 -0.768833193 -1.6967899
#> 12  -0.9700553 -0.65848436 -0.452489583 -1.6002041
#> 13  -0.9133665 -0.68743892 -1.093544394 -1.7526855
#> 14  -0.9676942 -0.50167285 -0.745485585 -1.7214540
#> 15  -0.9304879 -0.55000731 -0.422394639 -1.2028143
#> 16  -0.9978517 -0.34314664 -0.416217621 -1.6497508
#> 17  -0.8820441 -0.31286621 -1.054195718 -1.8928151
#> 18  -1.0833885 -0.56306190 -0.324133381 -1.4020840
#> 19  -0.9051631 -0.72807168 -0.607005848 -1.6111211
#> 20  -0.9304912 -0.45985783 -0.336760158 -1.5513197
#> 21  -0.7404763 -0.51271531 -0.523377481 -1.6586290
#> 22  -1.0187790 -0.50733200 -0.398569682 -1.5282437
#> 23  -0.9667462 -0.50268188  0.219079443 -1.7472997
#> 24  -0.9582441 -0.52672186 -0.497889445 -1.6796394
#> 25  -0.9562441 -0.49078944 -0.436195364 -1.5368417
#> 26  -0.9663854 -0.50560942 -0.717922815 -1.5744469
#> 27  -1.0639107 -0.67385177 -0.314030962 -1.4792762
#> 28  -1.0224155 -0.56975173 -0.603173569 -1.3855251
#> 29  -0.9982484 -0.54367301 -0.491158919 -1.7003176
#> 30  -0.9903443 -0.56946833 -0.708312367 -1.6394673
#> 31  -1.0251776 -0.68233892 -0.554351915 -1.6956811
#> 32  -1.1291780 -0.75686292 -0.248538853 -1.5232408
#> 33  -0.9796944 -0.61680796 -0.620540699 -1.6505963
#> 34  -0.8113945 -0.51235466 -0.693188889 -1.7945636
#> 35  -1.2200772 -0.97315003 -0.297626169 -1.2112230
#> 36  -0.9201198 -0.57223551 -0.259050612 -1.6016039
#> 37  -0.9142194 -0.49761480 -0.302216539 -1.5773389
#> 38  -0.9616561 -0.40710868 -0.534711732 -1.7406819
#> 39  -0.9492449 -0.55375334 -0.413224737 -1.5423728
#> 40  -0.9894384 -0.39938793 -0.682662659 -1.9928827
#> 41  -0.8806886 -0.51984748 -0.573467032 -1.4341995
#> 42  -0.9679518 -0.41888690 -0.787556953 -1.9248131
#> 43  -1.0198531 -0.41597227 -0.589020660 -1.6237633
#> 44  -1.0808396 -0.43678203 -0.469510191 -1.6236085
#> 45  -0.9283611 -0.42075473 -0.634612323 -1.7550151
#> 46  -1.1426424 -0.52704506 -0.499286173 -1.5166501
#> 47  -1.3017300 -0.66544623 -0.335280194 -1.2440264
#> 48  -0.9751348 -0.47633672 -0.386779406 -1.6351779
#> 49  -0.9342233 -0.50224789 -0.620946883 -1.6489467
#> 50  -0.9985964 -0.45330474 -0.630421878 -1.5630470
#> 51  -0.9315878 -0.51966481 -0.465783536 -1.7370025
#> 52  -0.9870514 -0.89181965 -0.111344331 -1.4699460
#> 53  -0.9169195 -0.55404444 -0.353231656 -1.5539168
#> 54  -0.8786327 -0.54303675 -0.346648089 -1.5058297
#> 55  -0.9477902 -0.61595771 -0.556930477 -1.1847693
#> 56  -0.9776619 -0.40318795 -0.531056056 -1.2977443
#> 57  -0.8434763 -0.31870875 -0.789633438 -2.0120383
#> 58  -0.9119672 -0.49924716 -0.293699621 -1.5967863
#> 59  -1.0127207 -0.50425518 -0.449933515 -1.5165119
#> 60  -0.7353512 -0.32838606 -0.819073724 -2.0949358
#> 61  -0.9699761 -0.68924491 -0.484697162 -1.5171890
#> 62  -0.9522664 -0.07751097 -0.483834810 -1.6472381
#> 63  -1.0367669 -0.60408727 -0.527622950 -1.4247506
#> 64  -1.3134263 -0.93463755 -0.007397619 -1.3883584
#> 65  -1.0032964 -0.64708915 -0.614916236 -1.7384065
#> 66  -0.9177661 -0.43309888 -0.461699589 -1.6036037
#> 67  -1.0272529 -0.72496932 -0.028530760 -1.8904442
#> 68  -0.7828386 -0.36599802 -0.887909725 -1.9098846
#> 69  -0.9893040 -0.50524756 -0.622618992 -1.7356624
#> 70  -0.9539401 -0.43292314 -0.708360568 -1.9231819
#> 71  -0.7878306 -0.39772668 -0.543149953 -1.6056142
#> 72  -0.9994864 -0.69351118 -0.405460899 -1.0117419
#> 73  -1.0497741 -0.50217850 -0.474831935 -2.0087113
#> 74  -1.0139782 -0.45067914 -0.662114619 -2.1055082
#> 75  -1.1013596 -0.54601993 -0.410663264 -1.5231039
#> 76  -0.9018078 -0.36172083 -0.813482345 -1.8808036
#> 77  -0.8985171 -0.53372756 -0.516235249 -1.6973150
#> 78  -0.9947448 -0.58090808 -0.574309776 -1.6330763
#> 79  -0.9571102 -0.28786304 -0.549653244 -1.6012989
#> 80  -0.9639456 -0.46655578 -0.537054621 -1.7091922
#> 81  -0.9073231 -0.13683258 -0.458140184 -1.2824003
#> 82  -0.8228464 -0.41598224 -0.901945703 -1.8475220
#> 83  -0.9821921 -0.51472643 -0.600772843 -1.6982912
#> 84  -0.9044736 -0.70538364 -0.811337429 -1.4543041
#> 85  -0.7528521 -0.58467753 -0.738577861 -0.6768273
#> 86  -1.0532265 -0.57457383 -0.346887444 -1.3267671
#> 87  -1.0273203 -0.43054779 -0.626083765 -1.5889587
#> 88  -0.9992775 -0.73110696 -0.520787631 -1.7935877
#> 89  -1.0463057 -0.74508312 -0.113591348 -1.2245920
#> 90  -0.9786128 -0.04573991 -0.535787685 -1.7672893
#> 91  -0.9862878 -0.55763510 -0.528719731 -1.5286193
#> 92  -0.8887304 -0.31649320 -0.638359478 -1.6257021
#> 93  -0.9084803 -0.55503472 -0.687918729 -1.8579733
#> 94  -0.9313091 -0.47280438 -0.731479665 -1.8076099
#> 95  -1.0004912 -0.48937830 -0.603968635 -1.5828861
#> 96  -0.9248655 -0.54723492 -0.301369406 -1.4633857
#> 97  -0.8763110 -0.46918520 -0.650341487 -1.4310851
#> 98  -0.8957991 -0.40858245 -0.605396604 -1.7572215
#> 99  -0.9637624 -0.58609448 -0.594637143 -1.2356965
#> 100 -0.9967384 -0.72002983 -0.302294464 -1.3738543
# or from the individual-level parameters and mapped
parameters(samples_LNR, selection = "alpha", map = TRUE)
#>     subjects  m_lMFALSE   m_lMTRUE         s         t0
#> 1       as1t -0.7872438 -1.1570009 0.4293839 0.16801658
#> 2       as1t -0.7817840 -1.0670096 0.3717159 0.14567906
#> 3       as1t -0.7175014 -1.0295195 0.3760064 0.13708996
#> 4       as1t -0.7872028 -1.1770163 0.4502336 0.17448345
#> 5       as1t -0.8156149 -1.1393071 0.4427031 0.16686474
#> 6       as1t -0.7005892 -1.0305718 0.3546520 0.12159670
#> 7       as1t -0.7276130 -1.0596583 0.4041022 0.14341162
#> 8       as1t -0.6935565 -1.1365971 0.4350651 0.15456096
#> 9       as1t -0.7700018 -1.1358369 0.4360685 0.16459899
#> 10      as1t -0.6749056 -0.9864615 0.3510998 0.11346567
#> 11      as1t -0.7195639 -1.0364228 0.3797011 0.13824446
#> 12      as1t -0.6886324 -1.0233665 0.3852579 0.12726889
#> 13      as1t -0.7752946 -1.1124558 0.4240553 0.16281209
#> 14      as1t -0.7462983 -1.1012901 0.4000910 0.14921527
#> 15      as1t -0.6106710 -0.9069335 0.3531630 0.09617918
#> 16      as1t -0.7053379 -1.0258826 0.3770549 0.13104244
#> 17      as1t -0.7111617 -1.0791789 0.4077908 0.14479556
#> 18      as1t -0.6936841 -0.9917101 0.3761668 0.12990626
#> 19      as1t -0.6627641 -0.9866716 0.3776492 0.12364275
#> 20      as1t -0.7133038 -1.0138732 0.3855606 0.13554326
#> 21      as1t -0.7217930 -1.0104788 0.3803812 0.12913389
#> 22      as1t -0.7242903 -1.0842174 0.4153931 0.15084367
#> 23      as1t -0.7430621 -1.0589072 0.3734719 0.13889447
#> 24      as1t -0.7216585 -1.0165478 0.3891099 0.12822946
#> 25      as1t -0.6406681 -0.9176045 0.3260429 0.08373377
#> 26      as1t -0.7306344 -1.0351993 0.4037467 0.14109172
#> 27      as1t -0.7127998 -1.0112245 0.3723586 0.12499177
#> 28      as1t -0.7520092 -1.0676239 0.3880921 0.13833993
#> 29      as1t -0.7197328 -1.0776054 0.4354251 0.14710947
#> 30      as1t -0.7508607 -1.0773372 0.4036126 0.14451288
#> 31      as1t -0.8226263 -1.1358475 0.4509304 0.17333695
#> 32      as1t -0.7310913 -1.0697305 0.4300573 0.15202498
#> 33      as1t -0.7679438 -1.0969674 0.4221040 0.15768440
#> 34      as1t -0.7487090 -1.0786683 0.4103467 0.15054198
#> 35      as1t -0.7744996 -1.1214514 0.4273803 0.16064738
#> 36      as1t -0.7964629 -1.1579801 0.4383740 0.17228062
#> 37      as1t -0.7588610 -1.0641653 0.4163428 0.14988124
#> 38      as1t -0.7009921 -0.9791781 0.3613880 0.12306163
#> 39      as1t -0.7238464 -1.0893286 0.4301851 0.15178784
#> 40      as1t -0.7874175 -1.1000332 0.4229469 0.16014793
#> 41      as1t -0.7362570 -1.0853779 0.4086294 0.14536702
#> 42      as1t -0.7485656 -1.1307904 0.4460703 0.16851289
#> 43      as1t -0.7016076 -1.0286610 0.3867210 0.12909833
#> 44      as1t -0.7480559 -1.1115560 0.4081617 0.15523067
#> 45      as1t -0.6405006 -0.9043945 0.3300171 0.08445788
#> 46      as1t -0.6532296 -0.9626391 0.3666730 0.10895569
#> 47      as1t -0.7040148 -1.0324869 0.3857556 0.12999493
#> 48      as1t -0.8352284 -1.2122631 0.4708365 0.18429295
#> 49      as1t -0.7168437 -1.0798260 0.4053278 0.14665948
#> 50      as1t -0.6070935 -0.9358713 0.3426830 0.09980805
#> 51      as1t -0.7368567 -1.0873505 0.4186030 0.14979388
#> 52      as1t -0.6763175 -0.9806645 0.3642130 0.11359139
#> 53      as1t -0.6563292 -0.9637161 0.3392152 0.10255932
#> 54      as1t -0.6533387 -0.9678162 0.3786926 0.11947505
#> 55      as1t -0.6614038 -1.0141181 0.3805165 0.12434530
#> 56      as1t -0.7628625 -1.1200091 0.4249748 0.15856903
#> 57      as1t -0.7220144 -1.0545759 0.3871085 0.13975675
#> 58      as1t -0.7109095 -1.0279436 0.3778541 0.13365437
#> 59      as1t -0.7501589 -1.0651132 0.4012592 0.14492181
#> 60      as1t -0.7381576 -1.0563478 0.4046139 0.14626901
#> 61      as1t -0.7546098 -1.0399877 0.3718251 0.13782127
#> 62      as1t -0.7144881 -1.0632520 0.4133738 0.14429825
#> 63      as1t -0.7515780 -1.0881057 0.4403547 0.16440378
#> 64      as1t -0.7798450 -1.1556270 0.4233238 0.16633895
#> 65      as1t -0.6620950 -1.0001065 0.3842340 0.12108479
#> 66      as1t -0.6893782 -1.0198099 0.3836754 0.13544359
#> 67      as1t -0.6438970 -0.9439693 0.3623025 0.11101413
#> 68      as1t -0.7697418 -1.0808029 0.3930781 0.15268079
#> 69      as1t -0.6505703 -0.9473575 0.3612524 0.10499307
#> 70      as1t -0.7703158 -1.0355221 0.3595024 0.13163747
#> 71      as1t -0.7131389 -1.0192035 0.3764897 0.12869942
#> 72      as1t -0.7582343 -1.1597054 0.4607373 0.17565856
#> 73      as1t -0.7258235 -1.0493961 0.3710301 0.13354864
#> 74      as1t -0.7536389 -1.0734022 0.4099012 0.14678890
#> 75      as1t -0.6551253 -0.9055053 0.3138597 0.08324387
#> 76      as1t -0.6970351 -1.0160926 0.3937478 0.12578117
#> 77      as1t -0.7232982 -1.0431399 0.3814829 0.13246155
#> 78      as1t -0.8060048 -1.1271240 0.4442186 0.16721360
#> 79      as1t -0.7345962 -1.0670625 0.4141387 0.14790988
#> 80      as1t -0.7428704 -1.1114749 0.4221870 0.15689487
#> 81      as1t -0.6804092 -0.9443908 0.3542292 0.10661996
#> 82      as1t -0.6949539 -0.9729386 0.3710433 0.11339297
#> 83      as1t -0.6530276 -0.9456366 0.3498557 0.09910587
#> 84      as1t -0.6868948 -0.9683943 0.3505727 0.10914970
#> 85      as1t -0.7890304 -1.1186404 0.4038738 0.15641637
#> 86      as1t -0.7200119 -1.0697364 0.4116401 0.14756638
#> 87      as1t -0.7872433 -1.1054970 0.4214496 0.15929478
#> 88      as1t -0.6662281 -0.9471675 0.3635858 0.09885393
#> 89      as1t -0.7117497 -1.0787862 0.3971265 0.14352790
#> 90      as1t -0.7362955 -1.0730466 0.4080333 0.14282382
#> 91      as1t -0.7085223 -0.9914230 0.3706027 0.12439652
#> 92      as1t -0.6873189 -0.9993865 0.3658255 0.12085522
#> 93      as1t -0.7544896 -1.0646996 0.3937386 0.14296763
#> 94      as1t -0.7358224 -1.0580397 0.4243844 0.14750745
#> 95      as1t -0.7346204 -1.1130749 0.4230006 0.15585541
#> 96      as1t -0.7138059 -1.0558257 0.3967372 0.13531140
#> 97      as1t -0.7394513 -1.0695557 0.4269228 0.15792581
#> 98      as1t -0.7780094 -1.1564659 0.4371126 0.16932010
#> 99      as1t -0.7107497 -1.0237728 0.3842626 0.13000473
#> 100     as1t -0.6560182 -0.9618811 0.3676605 0.11373738
#> 101     as1t -0.6897257 -1.0235771 0.3662859 0.12513523
#> 102     as1t -0.6448309 -0.9393513 0.3552599 0.10750156
#> 103     as1t -0.7055384 -1.0443786 0.4001313 0.12742383
#> 104     as1t -0.7928594 -1.0847885 0.4082327 0.15822779
#> 105     as1t -0.7365530 -1.0292112 0.3648745 0.12942239
#> 106     as1t -0.7177378 -1.0995296 0.4279220 0.15486906
#> 107     as1t -0.7282988 -1.0667479 0.3957993 0.14382537
#> 108     as1t -0.6581731 -0.9850518 0.3958601 0.12613933
#> 109     as1t -0.6849032 -0.9367062 0.3523640 0.10495299
#> 110     as1t -0.7819915 -1.1277728 0.4476395 0.16881901
#> 111     as1t -0.7464033 -1.0571454 0.3996287 0.14370013
#> 112     as1t -0.7394211 -1.0237487 0.3614820 0.12631052
#> 113     as1t -0.6748217 -1.0025288 0.3877831 0.12048503
#> 114     as1t -0.8301767 -1.1086645 0.4256795 0.16050426
#> 115     as1t -0.7341616 -1.0403601 0.3852665 0.13002827
#> 116     as1t -0.7806913 -1.0900099 0.4230767 0.16333041
#> 117     as1t -0.7226619 -1.0471909 0.4004733 0.14200933
#> 118     as1t -0.6909032 -1.0127200 0.3779048 0.12319567
#> 119     as1t -0.6556374 -0.9335804 0.3592914 0.10205537
#> 120     as1t -0.6773314 -0.9541202 0.3514517 0.10763646
#> 121     as1t -0.7308815 -1.0693791 0.4089236 0.14606731
#> 122     as1t -0.8084877 -1.1640536 0.4288452 0.17343929
#> 123     as1t -0.7754573 -1.1375975 0.4076061 0.15629489
#> 124     as1t -0.6994638 -1.0507805 0.3841095 0.13423306
#> 125     as1t -0.7351153 -1.0846929 0.4029959 0.14618218
#> 126     as1t -0.7053132 -0.9881773 0.3759156 0.12005471
#> 127     as1t -0.7068017 -1.0521620 0.4113915 0.13817453
#> 128     as1t -0.7201335 -1.0667811 0.4146586 0.14560177
#> 129     as1t -0.6750864 -0.9819987 0.3853135 0.12153688
#> 130     as1t -0.7736763 -1.1377099 0.4028054 0.15886852
#> 131     as1t -0.7447093 -1.0467212 0.3831709 0.13651159
#> 132     as1t -0.7338470 -1.0786933 0.4182813 0.15261339
#> 133     as1t -0.7421994 -1.0312194 0.3958864 0.13284598
#> 134     as1t -0.8002066 -1.1543569 0.4381499 0.17128686
#> 135     as1t -0.7767270 -1.0832061 0.4038499 0.15294995
#> 136     as1t -0.6775429 -0.9798464 0.3551291 0.11139055
#> 137     as1t -0.6990903 -0.9587767 0.3635988 0.11530345
#> 138     as1t -0.6852145 -0.9534068 0.3682995 0.11259044
#> 139     as1t -0.7074813 -1.0243337 0.3674736 0.12586707
#> 140     as1t -0.8186786 -1.1842117 0.4459840 0.18868285
#> 141     as1t -0.7409972 -1.0433496 0.3704495 0.13948958
#> 142     as1t -0.8193618 -1.1117800 0.4157302 0.16859284
#> 143     as1t -0.6300303 -0.9432500 0.3442393 0.09394759
#> 144     as1t -0.7176486 -1.0718319 0.3998227 0.13953618
#> 145     as1t -0.6202097 -0.9898864 0.3695391 0.10954006
#> 146     as1t -0.7162287 -1.0706455 0.3884072 0.14315521
#> 147     as1t -0.6499030 -0.9637710 0.3661698 0.10857226
#> 148     as1t -0.7311733 -1.1166088 0.4181847 0.15157460
#> 149     as1t -0.8229935 -1.1899999 0.4918894 0.18452999
#> 150     as1t -0.6669010 -0.9450893 0.3621561 0.11003797
#> 151     bd6t -0.6513525 -1.3135823 0.6245262 0.21992607
#> 152     bd6t -0.6440488 -1.2777198 0.5912560 0.20369259
#> 153     bd6t -0.6898805 -1.3221766 0.6241938 0.21207612
#> 154     bd6t -0.6849864 -1.3068814 0.6335140 0.21345184
#> 155     bd6t -0.6426868 -1.3510279 0.6368839 0.22033857
#> 156     bd6t -0.6271602 -1.2815950 0.6165215 0.20862328
#> 157     bd6t -0.6294156 -1.2741679 0.5904357 0.20226255
#> 158     bd6t -0.7174625 -1.2856999 0.5957751 0.20494825
#> 159     bd6t -0.6537927 -1.3473100 0.6198666 0.21563030
#> 160     bd6t -0.6851824 -1.3242719 0.6232013 0.21078805
#> 161     bd6t -0.6701575 -1.3716087 0.7002915 0.23431982
#> 162     bd6t -0.6998626 -1.4068322 0.6547853 0.22883210
#> 163     bd6t -0.6705283 -1.2608182 0.5683774 0.19922667
#> 164     bd6t -0.6852029 -1.2846741 0.6191840 0.20682661
#> 165     bd6t -0.7100346 -1.3235248 0.6095770 0.21577093
#> 166     bd6t -0.6710051 -1.1786403 0.5882826 0.19457873
#> 167     bd6t -0.6831252 -1.2803101 0.6464260 0.21401069
#> 168     bd6t -0.6946855 -1.3773605 0.6488808 0.22091591
#> 169     bd6t -0.6667655 -1.4163196 0.6594769 0.23118701
#> 170     bd6t -0.6974807 -1.3313847 0.6098617 0.21468964
#> 171     bd6t -0.7147610 -1.2703825 0.6453647 0.21689730
#> 172     bd6t -0.6730285 -1.2854458 0.5724651 0.20110736
#> 173     bd6t -0.6563006 -1.2742327 0.6069726 0.20020162
#> 174     bd6t -0.6791531 -1.2394442 0.5865296 0.20220978
#> 175     bd6t -0.7393746 -1.3087292 0.6072111 0.20906910
#> 176     bd6t -0.7129752 -1.3608107 0.6737275 0.22855023
#> 177     bd6t -0.7721832 -1.3436021 0.6094075 0.22252458
#> 178     bd6t -0.6572253 -1.3429096 0.6422432 0.21677493
#> 179     bd6t -0.6783492 -1.2400829 0.5822197 0.20308948
#> 180     bd6t -0.6437099 -1.2789448 0.6620428 0.21360788
#> 181     bd6t -0.7442113 -1.3054751 0.5802184 0.20687400
#> 182     bd6t -0.7455560 -1.2963851 0.5922387 0.21197528
#> 183     bd6t -0.6882040 -1.4459176 0.6792136 0.22948214
#> 184     bd6t -0.6431430 -1.2483815 0.6147576 0.20211201
#> 185     bd6t -0.6291594 -1.3136204 0.6285886 0.21626432
#> 186     bd6t -0.6193622 -1.3206329 0.6216105 0.20725513
#> 187     bd6t -0.6461730 -1.3262467 0.6410877 0.21208191
#> 188     bd6t -0.6757738 -1.3662900 0.6366853 0.21637044
#> 189     bd6t -0.6868235 -1.3239032 0.6476556 0.21831540
#> 190     bd6t -0.7159189 -1.3036833 0.6190394 0.21553141
#> 191     bd6t -0.7033656 -1.3578430 0.6269355 0.21907479
#> 192     bd6t -0.6976549 -1.3280984 0.6186824 0.21560577
#> 193     bd6t -0.6550737 -1.2996069 0.6186268 0.21244290
#> 194     bd6t -0.7487778 -1.3398290 0.6336352 0.22403247
#> 195     bd6t -0.7698773 -1.3678082 0.6527377 0.22894367
#> 196     bd6t -0.7099829 -1.3626568 0.6427447 0.22511511
#> 197     bd6t -0.7150821 -1.3105394 0.6151128 0.22087087
#> 198     bd6t -0.7210995 -1.3178158 0.6032981 0.21341860
#> 199     bd6t -0.6927786 -1.3649830 0.6737934 0.22474371
#> 200     bd6t -0.7042026 -1.3515560 0.5989583 0.21483830
#> 201     bd6t -0.7063591 -1.3626855 0.6651333 0.22734926
#> 202     bd6t -0.8258738 -1.3434636 0.6598113 0.22411554
#> 203     bd6t -0.7149521 -1.3273738 0.6318721 0.22635221
#> 204     bd6t -0.6591988 -1.1994517 0.5626243 0.19160973
#> 205     bd6t -0.7818350 -1.3406022 0.5969514 0.21995111
#> 206     bd6t -0.6814883 -1.2771950 0.6183380 0.20929941
#> 207     bd6t -0.6861687 -1.3194858 0.6349531 0.21580174
#> 208     bd6t -0.6466297 -1.2736365 0.6278820 0.21033520
#> 209     bd6t -0.6306682 -1.2622646 0.6011097 0.20118942
#> 210     bd6t -0.7217633 -1.2424506 0.5781905 0.20713212
#> 211     bd6t -0.6758677 -1.2665301 0.5518050 0.19887163
#> 212     bd6t -0.7475346 -1.3425651 0.6442243 0.22472406
#> 213     bd6t -0.7147900 -1.3660713 0.6520303 0.22431005
#> 214     bd6t -0.6250579 -1.3059337 0.6198061 0.20529211
#> 215     bd6t -0.6689934 -1.2147724 0.5537144 0.18906186
#> 216     bd6t -0.6661217 -1.3263685 0.6259941 0.21973081
#> 217     bd6t -0.6970342 -1.2502642 0.6143346 0.21041247
#> 218     bd6t -0.6578478 -1.2753348 0.6159787 0.20845757
#> 219     bd6t -0.7066331 -1.3821611 0.6504432 0.22298865
#> 220     bd6t -0.6629161 -1.2887062 0.6178499 0.20680869
#> 221     bd6t -0.6809803 -1.3603889 0.7094094 0.23429223
#> 222     bd6t -0.6676181 -1.2704152 0.5977281 0.20701192
#> 223     bd6t -0.6666943 -1.2731741 0.5769706 0.19953605
#> 224     bd6t -0.6693308 -1.3356488 0.6405377 0.21645913
#> 225     bd6t -0.6970132 -1.4172956 0.6927711 0.23421873
#> 226     bd6t -0.7162136 -1.4181931 0.6760360 0.23174386
#> 227     bd6t -0.6744792 -1.2452509 0.6080220 0.20538680
#> 228     bd6t -0.6482426 -1.3369358 0.6490138 0.22073318
#> 229     bd6t -0.6499383 -1.2757764 0.6033925 0.20452393
#> 230     bd6t -0.6929636 -1.3031066 0.6005276 0.21080828
#> 231     bd6t -0.6430733 -1.2638529 0.6191193 0.20526767
#> 232     bd6t -0.7078777 -1.3553834 0.6422639 0.21783240
#> 233     bd6t -0.7326696 -1.3327081 0.6439006 0.22110455
#> 234     bd6t -0.6921695 -1.2799153 0.5779720 0.20114157
#> 235     bd6t -0.6193310 -1.3228109 0.6493343 0.22319146
#> 236     bd6t -0.6624924 -1.2686299 0.6211480 0.21374307
#> 237     bd6t -0.6875806 -1.3230154 0.6495446 0.22010250
#> 238     bd6t -0.6753231 -1.2472992 0.5696760 0.20046454
#> 239     bd6t -0.6858468 -1.2993867 0.6236215 0.21534728
#> 240     bd6t -0.6576676 -1.3435743 0.6371050 0.22679132
#> 241     bd6t -0.6371954 -1.1861169 0.5762295 0.18932944
#> 242     bd6t -0.6349382 -1.3082052 0.6098963 0.21460627
#> 243     bd6t -0.7165922 -1.3373150 0.6602195 0.22937064
#> 244     bd6t -0.6259452 -1.2979280 0.5874952 0.20600045
#> 245     bd6t -0.7496883 -1.3689600 0.6972915 0.22461493
#> 246     bd6t -0.5981908 -1.2761600 0.6024936 0.20059397
#> 247     bd6t -0.7407439 -1.3523539 0.6272850 0.22180162
#> 248     bd6t -0.6912643 -1.2917587 0.6255867 0.21329626
#> 249     bd6t -0.6818949 -1.3468699 0.6574501 0.22033002
#> 250     bd6t -0.6499379 -1.2602769 0.5935858 0.19884548
#> 251     bd6t -0.7386345 -1.3090391 0.6051388 0.21434828
#> 252     bd6t -0.6677728 -1.3106525 0.6453151 0.21359980
#> 253     bd6t -0.7323352 -1.2575568 0.5703215 0.20671072
#> 254     bd6t -0.6998982 -1.2901009 0.5731684 0.20287694
#> 255     bd6t -0.7100538 -1.3132648 0.6584313 0.22604815
#> 256     bd6t -0.6265683 -1.3077185 0.6424235 0.21672062
#> 257     bd6t -0.6885798 -1.2724156 0.6029552 0.20875770
#> 258     bd6t -0.6821826 -1.3070054 0.6333662 0.21510314
#> 259     bd6t -0.6808222 -1.3436394 0.6527520 0.21701643
#> 260     bd6t -0.7183147 -1.2851893 0.6206698 0.21313935
#> 261     bd6t -0.6533935 -1.2589865 0.5559380 0.19860465
#> 262     bd6t -0.6738609 -1.2250497 0.5563053 0.19079286
#> 263     bd6t -0.7004082 -1.3325583 0.6163258 0.21770827
#> 264     bd6t -0.6823598 -1.3125692 0.6119460 0.21450920
#> 265     bd6t -0.7163883 -1.3633873 0.6342104 0.22479841
#> 266     bd6t -0.5937707 -1.3537364 0.6834280 0.22222801
#> 267     bd6t -0.6197231 -1.2853540 0.6048389 0.20605635
#> 268     bd6t -0.6744634 -1.3568515 0.6296059 0.21734391
#> 269     bd6t -0.7116277 -1.3148357 0.5872098 0.21348009
#> 270     bd6t -0.7187063 -1.2500753 0.5917608 0.20255116
#> 271     bd6t -0.6284374 -1.2470759 0.5945706 0.19575401
#> 272     bd6t -0.6544477 -1.3416845 0.6114220 0.21676606
#> 273     bd6t -0.6480485 -1.3652910 0.6471427 0.22303292
#> 274     bd6t -0.6866722 -1.2732820 0.6302779 0.20976229
#> 275     bd6t -0.6706104 -1.4130307 0.6406655 0.22435742
#> 276     bd6t -0.7326783 -1.4206500 0.6345549 0.22798450
#> 277     bd6t -0.6953945 -1.2992604 0.5841897 0.21147700
#> 278     bd6t -0.6737497 -1.3855504 0.6446499 0.22667556
#> 279     bd6t -0.6578104 -1.2352484 0.5977712 0.19121206
#> 280     bd6t -0.6821221 -1.3463002 0.6394038 0.21899726
#> 281     bd6t -0.6808072 -1.3449408 0.6326422 0.22197110
#> 282     bd6t -0.6398785 -1.2523714 0.5754070 0.19280316
#> 283     bd6t -0.7100033 -1.3321439 0.6601261 0.22693283
#> 284     bd6t -0.6947523 -1.2844014 0.6083885 0.20662422
#> 285     bd6t -0.6301285 -1.2461180 0.5784337 0.19665033
#> 286     bd6t -0.6402351 -1.1935122 0.5538978 0.18704626
#> 287     bd6t -0.7105553 -1.3627025 0.6326102 0.21754027
#> 288     bd6t -0.6996594 -1.3733172 0.6756695 0.23139392
#> 289     bd6t -0.7280160 -1.2966949 0.6365696 0.21849810
#> 290     bd6t -0.6970143 -1.3200086 0.6021421 0.21230308
#> 291     bd6t -0.6293012 -1.2325210 0.5927175 0.19820251
#> 292     bd6t -0.6973401 -1.3284179 0.6281564 0.20839674
#> 293     bd6t -0.6493191 -1.1795296 0.5629378 0.18621668
#> 294     bd6t -0.6805691 -1.3042673 0.6081048 0.21033989
#> 295     bd6t -0.6629757 -1.2222250 0.5869393 0.19585204
#> 296     bd6t -0.5897831 -1.2922985 0.6593161 0.21792809
#> 297     bd6t -0.5979289 -1.2287556 0.6041833 0.20544221
#> 298     bd6t -0.6604203 -1.3247229 0.6308341 0.21000323
#> 299     bd6t -0.6391888 -1.2565431 0.6252097 0.20679997
#> 300     bd6t -0.6549316 -1.2101131 0.5612868 0.19202108
#> 301     bl1t -0.6425259 -1.3150365 0.7363968 0.22279882
#> 302     bl1t -0.6255176 -1.2794691 0.7788502 0.22945684
#> 303     bl1t -0.7024789 -1.3804388 0.7665932 0.23720112
#> 304     bl1t -0.6747313 -1.3203565 0.7004331 0.21923165
#> 305     bl1t -0.6801801 -1.3191207 0.7436223 0.23163801
#> 306     bl1t -0.6981021 -1.2958516 0.7449468 0.22630224
#> 307     bl1t -0.6404600 -1.3596984 0.7845784 0.23256799
#> 308     bl1t -0.7094233 -1.2811712 0.7519419 0.21980108
#> 309     bl1t -0.7031465 -1.2652594 0.6971242 0.22044790
#> 310     bl1t -0.6002651 -1.2551225 0.7420754 0.22218404
#> 311     bl1t -0.6918574 -1.3060069 0.7166001 0.22096943
#> 312     bl1t -0.7189801 -1.3316324 0.7061819 0.22476580
#> 313     bl1t -0.7081653 -1.2919567 0.7454678 0.22608158
#> 314     bl1t -0.6589224 -1.2926582 0.7047279 0.21746556
#> 315     bl1t -0.7139431 -1.3286014 0.7931653 0.23479876
#> 316     bl1t -0.6880031 -1.3350333 0.7266994 0.22065584
#> 317     bl1t -0.6666357 -1.3108046 0.7765503 0.23264006
#> 318     bl1t -0.7165069 -1.2766482 0.7197293 0.21556985
#> 319     bl1t -0.7209080 -1.3099882 0.7452070 0.22983333
#> 320     bl1t -0.7525637 -1.3436897 0.7726297 0.23439923
#> 321     bl1t -0.7696770 -1.3567591 0.7938226 0.23954913
#> 322     bl1t -0.7129788 -1.3429108 0.7569677 0.23023590
#> 323     bl1t -0.7685169 -1.3587155 0.7336059 0.22574655
#> 324     bl1t -0.7270599 -1.2668246 0.6994483 0.21904876
#> 325     bl1t -0.6929205 -1.3132227 0.7519789 0.23032217
#> 326     bl1t -0.7508666 -1.3516455 0.7227470 0.22627236
#> 327     bl1t -0.7265757 -1.2684610 0.7371101 0.22239297
#> 328     bl1t -0.6461273 -1.3656607 0.7716244 0.23197263
#> 329     bl1t -0.7027315 -1.3106121 0.6918173 0.21506402
#> 330     bl1t -0.7530403 -1.3608641 0.7665630 0.23598347
#> 331     bl1t -0.6789778 -1.2961807 0.7158178 0.22109019
#> 332     bl1t -0.7582933 -1.2746832 0.6971084 0.21920971
#> 333     bl1t -0.7297764 -1.3781070 0.7138871 0.22316201
#> 334     bl1t -0.6413004 -1.3277531 0.7496622 0.23377654
#> 335     bl1t -0.6350292 -1.2781609 0.7272252 0.21286221
#> 336     bl1t -0.6726633 -1.3047938 0.7349912 0.22225665
#> 337     bl1t -0.7480874 -1.3383307 0.7488354 0.23454298
#> 338     bl1t -0.7535620 -1.3212156 0.7154615 0.22660095
#> 339     bl1t -0.6323505 -1.3235338 0.7502076 0.22639478
#> 340     bl1t -0.7511055 -1.3140514 0.7471710 0.22767624
#> 341     bl1t -0.8018702 -1.3205480 0.7252066 0.22342755
#> 342     bl1t -0.7245678 -1.3426947 0.6846748 0.21686119
#> 343     bl1t -0.7070918 -1.3714854 0.7998689 0.23732296
#> 344     bl1t -0.7385217 -1.2656082 0.7149736 0.22274236
#> 345     bl1t -0.7563972 -1.2668008 0.6934949 0.21616367
#> 346     bl1t -0.7298247 -1.3892367 0.7528917 0.23622227
#> 347     bl1t -0.6946585 -1.3183701 0.6952566 0.21802587
#> 348     bl1t -0.7389478 -1.2602555 0.7087188 0.21832872
#> 349     bl1t -0.6479551 -1.3055832 0.7992516 0.23211247
#> 350     bl1t -0.7594781 -1.3691365 0.7001022 0.22644805
#> 351     bl1t -0.6892592 -1.3681923 0.7359670 0.22484477
#> 352     bl1t -0.6895020 -1.3141755 0.7052336 0.22286415
#> 353     bl1t -0.6872580 -1.3099381 0.6952813 0.21791406
#> 354     bl1t -0.6563578 -1.2783442 0.7378265 0.22308686
#> 355     bl1t -0.7294925 -1.3863723 0.7947785 0.23910812
#> 356     bl1t -0.6511401 -1.3239657 0.7632875 0.22538128
#> 357     bl1t -0.6871650 -1.3320128 0.7362987 0.22595829
#> 358     bl1t -0.6832203 -1.3619220 0.7471103 0.22924068
#> 359     bl1t -0.7050625 -1.2558508 0.7056678 0.21995902
#> 360     bl1t -0.6698606 -1.3118140 0.6812678 0.21353042
#> 361     bl1t -0.7113797 -1.3093046 0.6984311 0.22295457
#> 362     bl1t -0.7972408 -1.3553240 0.7742162 0.23641167
#> 363     bl1t -0.6767465 -1.3253185 0.7574205 0.22470270
#> 364     bl1t -0.7276962 -1.3464585 0.8135278 0.23682517
#> 365     bl1t -0.6900629 -1.4217722 0.7952317 0.23565912
#> 366     bl1t -0.6622068 -1.2972868 0.6733936 0.21077955
#> 367     bl1t -0.6845068 -1.3499912 0.7400225 0.22865034
#> 368     bl1t -0.6525561 -1.2818679 0.6672929 0.20579903
#> 369     bl1t -0.7322513 -1.3036308 0.6466137 0.21244189
#> 370     bl1t -0.7418096 -1.3124488 0.7648888 0.23560544
#> 371     bl1t -0.7472292 -1.3169718 0.7438015 0.22610892
#> 372     bl1t -0.7121042 -1.2707056 0.7360128 0.22208134
#> 373     bl1t -0.7594214 -1.3218921 0.7600333 0.23260717
#> 374     bl1t -0.6699538 -1.3308481 0.7090700 0.21544508
#> 375     bl1t -0.7046195 -1.3142365 0.6925003 0.22021711
#> 376     bl1t -0.6412613 -1.3564876 0.7788940 0.23067137
#> 377     bl1t -0.7022379 -1.3151763 0.7105215 0.21898813
#> 378     bl1t -0.6601695 -1.2799191 0.6650373 0.20638399
#> 379     bl1t -0.6360542 -1.3150655 0.7176246 0.22308899
#> 380     bl1t -0.6767123 -1.2596359 0.6647197 0.21007997
#> 381     bl1t -0.7404283 -1.3213675 0.8034864 0.23698461
#> 382     bl1t -0.7018473 -1.3496325 0.7935591 0.23369937
#> 383     bl1t -0.7482937 -1.2799502 0.7117000 0.22389904
#> 384     bl1t -0.7002652 -1.2670464 0.7378487 0.22382038
#> 385     bl1t -0.6270800 -1.3365645 0.7924893 0.23070852
#> 386     bl1t -0.7475376 -1.3359938 0.7774005 0.23401093
#> 387     bl1t -0.6883527 -1.3544764 0.7476107 0.22482603
#> 388     bl1t -0.7231836 -1.3833655 0.7698436 0.23518384
#> 389     bl1t -0.7374950 -1.3559736 0.7295434 0.22857156
#> 390     bl1t -0.7117947 -1.2803006 0.7296215 0.22204952
#> 391     bl1t -0.7305692 -1.2636509 0.7186275 0.22377276
#> 392     bl1t -0.6427365 -1.2981797 0.7287554 0.21803839
#> 393     bl1t -0.7330085 -1.3358368 0.8085800 0.22887968
#> 394     bl1t -0.7255767 -1.3657404 0.7374130 0.23160980
#> 395     bl1t -0.6682339 -1.3244342 0.7407305 0.22533037
#> 396     bl1t -0.7117651 -1.3330417 0.7412962 0.22284786
#> 397     bl1t -0.6459263 -1.3028929 0.7491241 0.21998467
#> 398     bl1t -0.7624016 -1.2967695 0.6943597 0.21749172
#> 399     bl1t -0.6871591 -1.2659610 0.6710133 0.21071884
#> 400     bl1t -0.7500396 -1.2572505 0.7041646 0.22563021
#> 401     bl1t -0.6656479 -1.2492469 0.7021696 0.21492997
#> 402     bl1t -0.7589363 -1.3589792 0.7222205 0.23107347
#> 403     bl1t -0.7150904 -1.2953883 0.7060700 0.22182030
#> 404     bl1t -0.6468855 -1.2603937 0.6822754 0.21055147
#> 405     bl1t -0.7365705 -1.2780602 0.7304555 0.22532237
#> 406     bl1t -0.6744379 -1.2934210 0.7520711 0.22730409
#> 407     bl1t -0.7235483 -1.3785759 0.7035508 0.22353761
#> 408     bl1t -0.7325335 -1.4144151 0.7978611 0.23756639
#> 409     bl1t -0.7263850 -1.3012013 0.7302913 0.22285525
#> 410     bl1t -0.7240447 -1.3855253 0.7812781 0.23737288
#> 411     bl1t -0.6633044 -1.3041093 0.7205184 0.21690335
#> 412     bl1t -0.6728115 -1.3310603 0.7909516 0.22870552
#> 413     bl1t -0.6838013 -1.2607632 0.6808377 0.20983291
#> 414     bl1t -0.7159751 -1.3232140 0.6611157 0.21503266
#> 415     bl1t -0.7273238 -1.3756692 0.7723927 0.23712819
#> 416     bl1t -0.6412638 -1.3036523 0.7311275 0.21942872
#> 417     bl1t -0.6922304 -1.2641163 0.7049883 0.21940047
#> 418     bl1t -0.6496863 -1.3196942 0.7273203 0.22295811
#> 419     bl1t -0.7513628 -1.2598428 0.7312818 0.22762959
#> 420     bl1t -0.6941665 -1.3111917 0.7283716 0.22238208
#> 421     bl1t -0.7125374 -1.3108630 0.7393155 0.22419590
#> 422     bl1t -0.7194083 -1.3674738 0.7838579 0.23278268
#> 423     bl1t -0.6959866 -1.2404131 0.6519118 0.21138845
#> 424     bl1t -0.6892855 -1.3869124 0.7795217 0.23277260
#> 425     bl1t -0.6947145 -1.2469215 0.7092772 0.22080720
#> 426     bl1t -0.7224618 -1.3518398 0.7547425 0.23259646
#> 427     bl1t -0.7549161 -1.2468322 0.6894893 0.21758283
#> 428     bl1t -0.7561273 -1.2999979 0.7256994 0.22824315
#> 429     bl1t -0.7380485 -1.4338722 0.7718012 0.22660038
#> 430     bl1t -0.7559814 -1.2948797 0.6919549 0.22525424
#> 431     bl1t -0.7332289 -1.3852987 0.8059830 0.24434084
#> 432     bl1t -0.7459842 -1.2674477 0.6938678 0.21745504
#> 433     bl1t -0.6905369 -1.2845522 0.6830678 0.21739782
#> 434     bl1t -0.7507927 -1.3854020 0.7992228 0.24100618
#> 435     bl1t -0.6463572 -1.2574277 0.7078887 0.21868149
#> 436     bl1t -0.7259746 -1.4354514 0.7895601 0.23879250
#> 437     bl1t -0.6704558 -1.3720466 0.8053171 0.23329701
#> 438     bl1t -0.6845892 -1.3636331 0.7141532 0.22740105
#> 439     bl1t -0.6529809 -1.3113653 0.7512166 0.23193179
#> 440     bl1t -0.7420397 -1.2871301 0.7011978 0.22188850
#> 441     bl1t -0.6620421 -1.3204437 0.7298145 0.22020755
#> 442     bl1t -0.7735172 -1.2896468 0.7547983 0.22328284
#> 443     bl1t -0.7125990 -1.2788768 0.6937024 0.21897353
#> 444     bl1t -0.7061854 -1.2318731 0.6931918 0.21429552
#> 445     bl1t -0.7466843 -1.3090349 0.7262220 0.22296616
#> 446     bl1t -0.6533619 -1.3234566 0.7699138 0.22786518
#> 447     bl1t -0.6962337 -1.3756648 0.7510258 0.22583584
#> 448     bl1t -0.6631555 -1.3210536 0.7035021 0.21629529
#> 449     bl1t -0.7606564 -1.3373018 0.7431862 0.22996980
#> 450     bl1t -0.6347594 -1.3800775 0.8153273 0.23606915