Generate Posterior/Prior Predictives

Simulate n_post data sets using the posterior/prior parameter estimates

Usage

# S3 method for class 'emc.prior'
predict(object, data = NULL, n_post = 50, n_cores = 1, n_trials = NULL, ...)

# S3 method for class 'emc'
predict(
  object,
  hyper = FALSE,
  n_post = 50,
  n_cores = 1,
  stat = c("random", "mean", "median")[1],
  ...
)

Arguments

object

An emc or emc.prior object from which to generate predictives

data

A data frame needed to exactly match the original design

n_post

Integer. Number of generated datasets

n_cores

Integer. Number of cores across which there should be parallellized

n_trials

An integer. If data isn't provided (although preferred), can generate data based on n_trials per cell of design

...

Optional additional arguments passed to get_pars or make_data

hyper

Boolean. Defaults to FALSE. If TRUE, simulates from the group-level (hyper) parameters instead of the subject-level parameters.

stat

Character. Can be mean, median or random (i.e., the default). Will take either random samples from the chain(s) or use the mean or median of the parameter estimates.

Value

A list of simulated data sets of length n_post

Examples

# \donttest{
# based on an emc object ran by fit() we can generate posterior predictives
predict(samples_LNR, n_cores = 1, n_post = 10)
#>     postn subjects        E     S     R        rt trials
#> 1       1     as1t accuracy right  left 0.6254139      1
#> 3       1     as1t    speed right  left 0.4117822      2
#> 5       1     as1t    speed  left  left 0.3569358      3
#> 7       1     as1t accuracy right right 0.4658908      4
#> 9       1     as1t accuracy  left right 0.4670008      5
#> 11      1     as1t accuracy  left  left 0.4242591      6
#> 13      1     as1t  neutral  left  left 0.2657214      7
#> 15      1     as1t  neutral right right 0.4082121      8
#> 17      1     as1t  neutral  left right 0.4003372      9
#> 19      1     as1t accuracy right  left 0.4549237     10
#> 21      1     as1t accuracy right  left 0.4530192     11
#> 23      1     as1t  neutral  left  left 0.3948881     12
#> 25      1     as1t accuracy  left  left 0.5030849     13
#> 27      1     as1t accuracy  left  left 0.5952443     14
#> 29      1     as1t  neutral right  left 0.4313421     15
#> 31      1     as1t    speed right right 0.4414972     16
#> 33      1     as1t accuracy right right 0.6575966     17
#> 35      1     as1t    speed right right 0.8082370     18
#> 37      1     as1t accuracy right right 0.5082482     19
#> 39      1     as1t    speed  left  left 0.6374160     20
#> 41      1     as1t accuracy right right 0.4203186     21
#> 43      1     as1t  neutral  left  left 0.4267183     22
#> 45      1     as1t  neutral  left  left 0.4144285     23
#> 47      1     as1t    speed  left  left 0.3971042     24
#> 49      1     as1t accuracy right right 0.5475085     25
#> 51      1     as1t  neutral right right 0.3843506     26
#> 53      1     as1t accuracy  left  left 0.3719949     27
#> 55      1     as1t  neutral  left  left 0.4273270     28
#> 57      1     as1t  neutral right  left 0.5023278     29
#> 59      1     as1t    speed  left  left 0.3084145     30
#> 61      1     as1t    speed  left  left 0.3244708     31
#> 63      1     as1t  neutral right right 0.5790864     32
#> 65      1     as1t    speed  left right 0.4883445     33
#> 67      1     as1t accuracy  left  left 0.4941199     34
#> 69      1     as1t accuracy  left  left 0.4290373     35
#> 71      1     as1t    speed right right 0.4446584     36
#> 73      1     as1t  neutral right right 0.4367800     37
#> 75      1     as1t accuracy right right 0.4655631     38
#> 77      1     as1t    speed  left  left 0.3613322     39
#> 79      1     as1t  neutral right  left 0.3306966     40
#> 81      1     as1t    speed right  left 0.4340761     41
#> 83      1     as1t    speed  left right 0.4680188     42
#> 85      1     as1t  neutral right  left 0.5885984     43
#> 87      1     as1t  neutral right right 0.5932288     44
#> 89      1     as1t  neutral right right 0.7987355     45
#> 91      1     as1t accuracy  left  left 0.2967151     46
#> 93      1     as1t    speed  left  left 0.6149597     47
#> 95      1     as1t accuracy right  left 0.3713583     48
#> 97      1     as1t  neutral right  left 0.3955865     49
#> 99      1     as1t    speed right right 0.5150043     50
#> 101     1     as1t    speed  left  left 0.4849034     51
#> 103     1     as1t  neutral right right 0.4411967     52
#> 105     1     as1t accuracy right  left 0.4039128     53
#> 107     1     as1t  neutral  left  left 0.3661824     54
#> 109     1     as1t accuracy right  left 0.8139798     55
#> 111     1     as1t    speed right right 0.4558547     56
#> 113     1     as1t    speed  left  left 0.6914065     57
#> 115     1     as1t    speed right right 0.4755739     58
#> 117     1     as1t accuracy right right 0.3753376     59
#> 119     1     as1t accuracy right right 0.6869336     60
#> 121     1     as1t    speed right right 0.5263887     61
#> 123     1     as1t accuracy right right 0.6989235     62
#> 125     1     as1t    speed  left  left 0.5788471     63
#> 127     1     as1t  neutral  left  left 0.3695878     64
#> 129     1     as1t  neutral  left  left 0.6352199     65
#> 131     1     as1t accuracy  left  left 0.4674349     66
#> 133     1     as1t    speed  left right 0.5439651     67
#> 135     1     as1t  neutral right right 0.2961434     68
#> 137     1     as1t  neutral  left  left 0.7410274     69
#> 139     1     as1t  neutral  left  left 0.3790847     70
#> 141     1     as1t    speed right  left 0.5192558     71
#> 143     1     as1t    speed  left  left 0.3374245     72
#> 145     1     as1t  neutral right  left 0.4486350     73
#> 147     1     as1t  neutral right right 0.2953331     74
#> 149     1     as1t    speed  left  left 0.4348216     75
#> 151     1     as1t    speed  left  left 0.4647976     76
#> 153     1     as1t    speed right right 0.5407136     77
#> 155     1     as1t  neutral  left  left 0.3462920     78
#> 157     1     as1t accuracy right right 0.3781027     79
#> 159     1     as1t    speed right  left 0.2859385     80
#> 161     1     as1t    speed  left  left 0.3487933     81
#> 163     1     as1t accuracy right  left 0.5425322     82
#> 165     1     as1t accuracy  left right 0.4299138     83
#> 167     1     as1t  neutral right right 0.4502385     84
#> 169     1     as1t    speed right right 0.3574119     85
#> 171     1     as1t    speed right  left 0.4593865     86
#> 173     1     as1t accuracy  left  left 0.5774532     87
#> 175     1     as1t    speed  left  left 0.4669459     88
#> 177     1     as1t accuracy right  left 0.4536489     89
#> 179     1     as1t accuracy  left  left 0.4436052     90
#> 181     1     as1t accuracy right right 0.4933910     91
#> 183     1     as1t    speed  left  left 0.2963580     92
#> 185     1     as1t    speed  left  left 0.6212734     93
#> 187     1     as1t accuracy right right 0.6399800     94
#> 189     1     as1t    speed  left  left 0.6447397     95
#> 191     1     as1t  neutral right  left 0.4713033     96
#> 193     1     as1t  neutral  left  left 0.4176305     97
#> 195     1     as1t    speed  left right 0.3971596     98
#> 197     1     as1t accuracy  left  left 0.4488975     99
#> 199     1     as1t  neutral right right 0.7569792    100
#> 201     1     as1t accuracy right  left 0.5636566    101
#> 203     1     as1t accuracy right right 0.4887500    102
#> 205     1     as1t  neutral right right 0.2828674    103
#> 207     1     as1t accuracy right right 0.3406199    104
#> 209     1     as1t    speed right right 0.4729912    105
#> 211     1     as1t accuracy right right 0.3675573    106
#> 213     1     as1t  neutral  left  left 0.7320758    107
#> 215     1     as1t  neutral right  left 0.5241622    108
#> 217     1     as1t  neutral right right 0.5753958    109
#> 219     1     as1t accuracy right  left 0.6250898    110
#> 221     1     as1t  neutral right right 0.5600237    111
#> 223     1     as1t    speed  left  left 0.2935740    112
#> 225     1     as1t  neutral  left  left 0.5118903    113
#> 227     1     as1t  neutral right  left 0.4202074    114
#> 229     1     as1t accuracy  left  left 0.4193253    115
#> 231     1     as1t  neutral right right 0.6851206    116
#> 233     1     as1t    speed right right 0.4115369    117
#> 235     1     as1t accuracy  left  left 0.3975062    118
#> 237     1     as1t  neutral right  left 0.4845970    119
#> 239     1     as1t  neutral  left  left 0.4305321    120
#> 241     1     as1t accuracy right  left 0.4798481    121
#> 243     1     as1t  neutral  left  left 0.5451094    122
#> 245     1     as1t accuracy  left  left 0.4063619    123
#> 247     1     as1t accuracy right  left 0.3143698    124
#> 249     1     as1t accuracy right right 0.3176393    125
#> 251     1     as1t  neutral  left  left 0.4089291    126
#> 253     1     as1t accuracy right right 0.4115995    127
#> 255     1     as1t  neutral  left  left 0.3907702    128
#> 257     1     as1t    speed  left  left 0.4578902    129
#> 259     1     as1t accuracy right  left 0.7416022    130
#> 261     1     as1t  neutral right right 0.5184400    131
#> 263     1     as1t  neutral right right 0.5724417    132
#> 265     1     as1t  neutral  left  left 0.5117118    133
#> 267     1     as1t accuracy right right 0.3906642    134
#> 269     1     as1t    speed right right 0.3899849    135
#> 271     1     as1t accuracy right right 0.5092735    136
#> 273     1     as1t  neutral right right 0.5281823    137
#> 275     1     as1t accuracy right right 0.5646688    138
#> 277     1     as1t  neutral  left right 0.5197626    139
#> 279     1     as1t  neutral  left  left 0.2667961    140
#> 281     1     as1t accuracy right  left 0.3733536    141
#> 283     1     as1t accuracy right  left 0.3378385    142
#>  [ reached 'max' / getOption("max.print") -- omitted 24878 rows ]
# }